Module Core.Arg

INRIA's original command-line parsing library.

The Command module is generally recommended over direct use of this library.

type spec = Stdlib__arg.spec =

| Unit of unit -> unit

| Bool of bool -> unit

| Set of bool Stdlib.ref

| Clear of bool Stdlib.ref

| String of string -> unit

| Set_string of string Stdlib.ref

| Int of int -> unit

| Set_int of int Stdlib.ref

| Float of float -> unit

| Set_float of float Stdlib.ref

| Tuple of spec list

| Symbol of string list * string -> unit

| Rest of string -> unit

| Rest_all of string list -> unit

| Expand of string -> string array

type key = string

type doc = string

type usage_msg = string

type anon_fun = string -> unit

val parse : (key * spec * doc) list -> anon_fun -> usage_msg -> unit

val parse_dynamic : (key * spec * doc) list Stdlib.ref -> anon_fun -> usage_msg -> unit

val parse_argv : ?current:int Stdlib.ref -> string array -> (key * spec * doc) list -> anon_fun -> usage_msg -> unit

val parse_argv_dynamic : ?current:int Stdlib.ref -> string array -> (key * spec * doc) list Stdlib.ref -> anon_fun -> string -> unit

val parse_and_expand_argv_dynamic : int Stdlib.ref -> string array Stdlib.ref -> (key * spec * doc) list Stdlib.ref -> anon_fun -> string -> unit

val parse_expand : (key * spec * doc) list -> anon_fun -> usage_msg -> unit

exception Help of string

exception Bad of string

val usage : (key * spec * doc) list -> usage_msg -> unit

val usage_string : (key * spec * doc) list -> usage_msg -> string

val align : ?limit:int -> (key * spec * doc) list -> (key * spec * doc) list

val current : int Stdlib.ref

val read_arg : string -> string array

val read_arg0 : string -> string array

val write_arg : string -> string array -> unit

val write_arg0 : string -> string array -> unit

type t = key * spec * doc

val sort_and_align : (key * spec * doc) Base.List.t -> (key * spec * doc) Base.List.t

Like align, except that the specification list is also sorted by key