Module Random.State

type t

The type of PRNG states.

val make : int array -> t

Create a new state and initialize it with the given seed.

val make_self_init : unit -> t

Create a new state and initialize it with a system-dependent low-entropy seed.

val copy : t -> t

Return a copy of the given state.

val bits : t -> int

val int : t -> int -> int

val int32 : t -> Stdlib.Int32.t -> Stdlib.Int32.t

val nativeint : t -> Stdlib.Nativeint.t -> Stdlib.Nativeint.t

val int64 : t -> Stdlib.Int64.t -> Stdlib.Int64.t

val float : t -> float -> float

val bool : t -> bool

These functions are the same as the basic functions, except that they use (and update) the given PRNG state instead of the default one.