Module Spacetime.Snapshot

val take : ?⁠time:float -> Series.t -> unit

take series takes a snapshot of the profiling annotations on the values in the minor and major heaps, together with GC stats, and write the result to the series file. This function triggers a minor GC but does not allocate any memory itself. If the optional time is specified, it will be used instead of the result of Sys.time as the timestamp of the snapshot. Such times should start from zero and be monotonically increasing. This parameter is intended to be used so that snapshots can be correlated against wall clock time (which is not supported in the standard library) rather than elapsed CPU time.