Module Core.Char

This module extends Base.Char, adding Identifiable for making char identifiers and Quickcheckable to facilitate automated testing with pseudorandom data.

type t = char

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

The signature included from Base.Char

val all : t list

include Sexplib0.Sexpable.S with type t := t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

include Base.Identifiable.S with type t := t

include Sexplib0.Sexpable.S with type t := t

include Base.Stringable.S with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

include Base.Comparator.S with type t := t

type comparator_witness

include Base.Pretty_printer.S with type t := t

include Base.Invariant.S with type t := t

val invariant : t -> unit

module O : Base.Comparisons.Infix with type t := t

val to_int : t -> int

Returns the ASCII code of the argument.

val of_int : int -> t option

Returns the character with the given ASCII code or None is the argument is outside the range 0 to 255.

val of_int_exn : int -> t

Returns the character with the given ASCII code. Raises Failure if the argument is outside the range 0 to 255.

val unsafe_of_int : int -> t

val escaped : t -> string

Returns a string representing the given character, with special characters escaped following the lexical conventions of OCaml.

val lowercase : t -> t

Converts the given character to its equivalent lowercase character.

val uppercase : t -> t

Converts the given character to its equivalent uppercase character.

val is_digit : t -> bool

'0' - '9'

val is_lowercase : t -> bool

'a' - 'z'

val is_uppercase : t -> bool

'A' - 'Z'

val is_alpha : t -> bool

'a' - 'z' or 'A' - 'Z'

val is_alphanum : t -> bool

'a' - 'z' or 'A' - 'Z' or '0' - '9'

val is_print : t -> bool

' ' - '~'

val is_whitespace : t -> bool

' ' or '\t' or '\r' or '\n'

val get_digit : t -> int option

Returns Some i if is_digit c and None otherwise.

val get_digit_exn : t -> int

Returns i if is_digit c and raises Failure otherwise.

val is_hex_digit : t -> bool

'0' - '9' or 'a' - 'f' or 'A' - 'F'

val is_hex_digit_lower : t -> bool

'0' - '9' or 'a' - 'f'

val is_hex_digit_upper : t -> bool

'0' - '9' or 'A' - 'F'

val get_hex_digit : t -> int option

Returns Some i where 0 <= i && i < 16 if is_hex_digit c and None otherwise.

val get_hex_digit_exn : t -> int

Same as get_hex_digit but raises instead of returning None.

val min_value : t

val max_value : t

Extensions

module Caseless : sig ... end

Caseless compares and hashes characters ignoring case, so that for example Caseless.equal 'A' 'a' and Caseless.('a' < 'B') are true, and Caseless.Map, Caseless.Table lookup and Caseless.Set membership is case-insensitive.

include Identifiable.S with type t := t and type comparator_witness := comparator_witness

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Ppx_compare_lib.Comparable.S with type t := t

include Ppx_hash_lib.Hashable.S with type t := t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Base.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Base.Pretty_printer.S with type t := t

val pp : Base.Formatter.t -> t -> unit

include Comparable.S_binable with type t := t with type comparator_witness := comparator_witness

include Base.Comparable.S with type t := t with type comparator_witness := comparator_witness

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t with type comparator_witness := comparator_witness

val validate_lbound : min:t Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Maybe_bound.t -> t Validate.check

val validate_bound : min:t Maybe_bound.t -> max:t Maybe_bound.t -> t Validate.check

module Replace_polymorphic_compare : Base.Comparable.Polymorphic_compare with type t := t

include Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : ( t, comparator_witness ) Comparator.comparator

module Map : Map.S_binable with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Set.S_binable with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Hashable.S_binable with type t := t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val hashable : t Hashtbl.Hashable.t

module Table : Hashtbl.S_binable with type key = t

module Hash_set : Hash_set.S_binable with type elt = t

module Hash_queue : Hash_queue.S with type key = t

Quickcheck Support

include Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val gen_digit : t Quickcheck.Generator.t

val gen_lowercase : t Quickcheck.Generator.t

val gen_uppercase : t Quickcheck.Generator.t

val gen_alpha : t Quickcheck.Generator.t

val gen_alphanum : t Quickcheck.Generator.t

val gen_print : t Quickcheck.Generator.t

val gen_whitespace : t Quickcheck.Generator.t

val gen_uniform_inclusive : t -> t -> t Quickcheck.Generator.t

Generates characters between the given inclusive bounds in ASCII order. Raises if bounds are in decreasing order.