Module Core.Command

module type Enumerable_sexpable = sig ... end

module type Enumerable_stringable = sig ... end

module Auto_complete : sig ... end

Specifications for command-line auto-completion.

module Arg_type : sig ... end

Argument types.

module Flag : sig ... end

Command-line flag specifications.

module Anons : sig ... end

Anonymous command-line argument specification.

module Param : sig ... end

Command-line parameter specification.

module Let_syntax : sig ... end

module Spec : sig ... end

The old interface for command-line specifications -- Do Not Use.

type t

Commands which can be combined into a hierarchy of subcommands.

type ('main, 'result) basic_spec_command = summary:Base.String.t -> ?readme:( Base.Unit.t -> Base.String.t ) -> ( 'main, Base.Unit.t -> 'result ) Spec.t -> 'main -> t

val basic_spec : ( 'main, Base.Unit.t ) basic_spec_command

basic_spec ~summary ?readme spec main is a basic command that executes a function main which is passed parameters parsed from the command line according to spec. summary is to contain a short one-line description of its behavior. readme is to contain any longer description of its behavior that will go on that command's help screen.

type 'result basic_command = summary:Base.String.t -> ?readme:( Base.Unit.t -> Base.String.t ) -> ( Base.Unit.t -> 'result ) Param.t -> t

val basic : Base.Unit.t basic_command

Same general behavior as basic_spec, but takes a command line specification built up using Params instead of Spec.

val group : summary:Base.String.t -> ?readme:( Base.Unit.t -> Base.String.t ) -> ?preserve_subcommand_order:Base.Unit.t -> ?body:( path:Base.String.t Base.List.t -> Base.Unit.t ) -> (Base.String.t * t) Base.List.t -> t

group ~summary subcommand_alist is a compound command with named subcommands, as found in subcommand_alist. summary is to contain a short one-line description of the command group. readme is to contain any longer description of its behavior that will go on that command's help screen.

NOTE: subcommand names containing underscores will be rejected; use dashes instead.

body is called when no additional arguments are passed -- in particular, when no subcommand is passed. Its path argument is the subcommand path by which the group command was reached.

val lazy_group : summary:Base.String.t -> ?readme:( Base.Unit.t -> Base.String.t ) -> ?preserve_subcommand_order:Base.Unit.t -> ?body:( path:Base.String.t Base.List.t -> Base.Unit.t ) -> (Base.String.t * t) Base.List.t Lazy.t -> t

lazy_group is the same as group, except that the list of subcommands may be generated lazily.

val exec : summary:Base.String.t -> ?readme:( Base.Unit.t -> Base.String.t ) -> ?child_subcommand:Base.String.t Base.List.t -> ?env: [ `Replace of (Base.String.t * Base.String.t) Base.List.t | `Extend of (Base.String.t * Base.String.t) Base.List.t | `Override of (Base.String.t * Base.String.t Base.Option.t) Base.List.t | `Replace_raw of Base.String.t Base.List.t ] -> path_to_exe: [ `Absolute of Base.String.t | `Relative_to_argv0 of Base.String.t | `Relative_to_me of Base.String.t ] -> Base.Unit.t -> t

exec ~summary ~path_to_exe runs exec on the executable at path_to_exe. If path_to_exe is `Absolute path then path is executed without any further qualification. If it is `Relative_to_me path then Filename.dirname Sys.executable_name ^ "/" ^ path is executed instead. All of the usual caveats about Sys.executable_name apply: specifically, it may only return an absolute path in Linux. On other operating systems it will return Sys.argv.(0). If it is `Relative_to_argv0 path then Sys.argv.(0) ^ "/" ^ path is executed.

The child_subcommand argument allows referencing a subcommand one or more levels below the top-level of the child executable. It should not be used to pass flags or anonymous arguments to the child.

Care has been taken to support nesting multiple executables built with Command. In particular, recursive help and autocompletion should work as expected.

NOTE: Non-Command executables can be used with this function but will still be executed when help -recursive is called or autocompletion is attempted (despite the fact that neither will be particularly helpful in this case). This means that if you have a shell script called "reboot-everything.sh" that takes no arguments and reboots everything no matter how it is called, you shouldn't use it with exec.

Additionally, no loop detection is attempted, so if you nest an executable within itself, help -recursive and autocompletion will hang forever (although actually running the subcommand will work).

val of_lazy : t Lazy.t -> t

of_lazy thunk constructs a lazy command that is forced only when necessary to run it or extract its shape.

val summary : t -> Base.String.t

Extracts the summary string for a command.

module Shape : sig ... end

val exit : Base.Int.t -> _

call this instead of Core.exit if in command-related code that you want to run in tests. For example, in the body of Command.Param.no_arg_abort

module Deprecated : sig ... end

Deprecated should be used only by Deprecated_command. At some point it will go away.

Deprecated values that moved from Core.Command to Command_unix.

val run : [ `Use_Command_unix ]

module Path : sig ... end

val shape : [ `Use_Command_unix ]