Module Command.Path

type t

Path.t is a top-level executable name and sequence of subcommand names that can be used to identify a command.

val create : path_to_exe:string -> t

create creates a path from a toplevel executable given by path_to_exe.

val append : t -> subcommand:string -> t

append appends a subcommand to t.

val parts : t -> string list

parts returns a list containing the path's executable name followed by its subcommands.