Module Shape.Base_info

type t = {

summary : Base.String.t;

readme : Base.String.t Base.Option.t;

anons : Anons.t;

flags : Flag_info.t Base.List.t;

}

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

val flags : t -> Flag_info.t Base.List.t

val anons : t -> Anons.t

val readme : t -> Base.String.t Base.Option.t

val summary : t -> Base.String.t

module Fields : sig ... end

val sexp_of_t : t -> Sexplib0.Sexp.t

val find_flag : t -> Base.String.t -> Flag_info.t Or_error.t

find_flag t prefix looks up the flag, if any, to which prefix refers.

It raises if prefix does not begin with - as all flags should.

find_flag does not consider aliases_excluded_from_help, and it assumes that all flags can be passed by prefix. These are limitations in the underlying shape representation.

val get_usage : t -> Base.String.t

include Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val t_of_sexp : Sexp.t -> t