Module Shape.Flag_info

type t = {

name : Base.String.t;	(*
See flag_name below.
*)

doc : Base.String.t;

aliases : Base.String.t Base.List.t;

}

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

val aliases : t -> Base.String.t Base.List.t

val doc : t -> Base.String.t

val name : t -> Base.String.t

module Fields : sig ... end

val sexp_of_t : t -> Sexplib0.Sexp.t

val flag_name : t -> Base.String.t Or_error.t

flag_name infers the string which one would pass on the command line. It is not the same as the raw name field, which additionally encodes num_occurrences and requires_arg (sort of).

val num_occurrences : t -> Num_occurrences.t Or_error.t

val requires_arg : t -> Base.Bool.t Or_error.t

requires_arg gives undefined behavior on escape flags. This is a limitation of the underlying shape representation.

include Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val t_of_sexp : Sexp.t -> t