Module Core_private.Stable_internal

include module type of struct include Bin_prot.Std end

include module type of struct include Bin_prot.Size end

type 'a sizer = 'a -> int

type ('a, 'b) sizer1 = 'a sizer -> 'b sizer

type ('a, 'b, 'c) sizer2 = 'a sizer -> ('b, 'c) sizer1

type ('a, 'b, 'c, 'd) sizer3 = 'a sizer -> ('b, 'c, 'd) sizer2

type 'a sizer_local = 'a -> int

type ('a, 'b) sizer_local1 = 'a sizer_local -> 'b sizer_local

type ('a, 'b, 'c) sizer_local2 = 'a sizer_local -> ('b, 'c) sizer_local1

type ('a, 'b, 'c, 'd) sizer_local3 = 'a sizer_local -> ('b, 'c, 'd) sizer_local2

val bin_size_unit : unit sizer

val bin_size_bool : bool sizer

val bin_size_string : string sizer

val bin_size_bytes : bytes sizer

val bin_size_char : char sizer

val bin_size_int : int sizer

val bin_size_float : float sizer

val bin_size_int32 : int32 sizer

val bin_size_int64 : int64 sizer

val bin_size_nativeint : nativeint sizer

val bin_size_nat0 : Bin_prot.Nat0.t sizer

val bin_size_ref : ('a, 'a ref) sizer1

val bin_size_lazy_t : ('a, 'a lazy_t) sizer1

val bin_size_lazy : ('a, 'a lazy_t) sizer1

val bin_size_option : ('a, 'a option) sizer1

val bin_size_pair : ('a, 'b, 'a * 'b) sizer2

val bin_size_triple : ('a, 'b, 'c, 'a * 'b * 'c) sizer3

val bin_size_list : ('a, 'a list) sizer1

val bin_size_array : ('a, 'a array) sizer1

val bin_size_hashtbl : ('a, 'b, ('a, 'b) Hashtbl.t) sizer2

val bin_size_float32_vec : Bin_prot.Common.vec32 sizer

val bin_size_float64_vec : Bin_prot.Common.vec64 sizer

val bin_size_vec : Bin_prot.Common.vec sizer

val bin_size_float32_mat : Bin_prot.Common.mat32 sizer

val bin_size_float64_mat : Bin_prot.Common.mat64 sizer

val bin_size_mat : Bin_prot.Common.mat sizer

val bin_size_bigstring : Bin_prot.Common.buf sizer

val bin_size_floatarray : floatarray sizer

val bin_size_variant_int : int sizer

val bin_size_int_8bit : int sizer

val bin_size_int_16bit : int sizer

val bin_size_int_32bit : int sizer

val bin_size_int_64bit : int sizer

val bin_size_int64_bits : int64 sizer

val bin_size_network16_int : int sizer

val bin_size_network32_int : int sizer

val bin_size_network32_int32 : int32 sizer

val bin_size_network64_int : int sizer

val bin_size_network64_int64 : int64 sizer

val bin_size_md5 : Md5_lib.t sizer

val bin_size_unit__local : unit sizer_local

val bin_size_bool__local : bool sizer_local

val bin_size_string__local : string sizer_local

val bin_size_bytes__local : bytes sizer_local

val bin_size_char__local : char sizer_local

val bin_size_int__local : int sizer_local

val bin_size_float__local : float sizer_local

val bin_size_int32__local : int32 sizer_local

val bin_size_int64__local : int64 sizer_local

val bin_size_nativeint__local : nativeint sizer_local

val bin_size_nat0__local : Bin_prot.Nat0.t sizer_local

val bin_size_ref__local : ('a, 'a ref) sizer_local1

val bin_size_lazy_t__local : ('a, 'a lazy_t) sizer_local1

val bin_size_lazy__local : ('a, 'a lazy_t) sizer_local1

val bin_size_option__local : ('a, 'a option) sizer_local1

val bin_size_pair__local : ('a, 'b, 'a * 'b) sizer_local2

val bin_size_triple__local : ('a, 'b, 'c, 'a * 'b * 'c) sizer_local3

val bin_size_list__local : ('a, 'a list) sizer_local1

val bin_size_array__local : ('a, 'a array) sizer_local1

val bin_size_float32_vec__local : Bin_prot.Common.vec32 sizer_local

val bin_size_float64_vec__local : Bin_prot.Common.vec64 sizer_local

val bin_size_vec__local : Bin_prot.Common.vec sizer_local

val bin_size_float32_mat__local : Bin_prot.Common.mat32 sizer_local

val bin_size_float64_mat__local : Bin_prot.Common.mat64 sizer_local

val bin_size_mat__local : Bin_prot.Common.mat sizer_local

val bin_size_bigstring__local : Bin_prot.Common.buf sizer_local

val bin_size_floatarray__local : floatarray sizer_local

val bin_size_variant_int__local : int sizer_local

val bin_size_int_8bit__local : int sizer_local

val bin_size_int_16bit__local : int sizer_local

val bin_size_int_32bit__local : int sizer_local

val bin_size_int_64bit__local : int sizer_local

val bin_size_int64_bits__local : int64 sizer_local

val bin_size_network16_int__local : int sizer_local

val bin_size_network32_int__local : int sizer_local

val bin_size_network32_int32__local : int32 sizer_local

val bin_size_network64_int__local : int sizer_local

val bin_size_network64_int64__local : int64 sizer_local

val bin_size_md5__local : Md5_lib.t sizer_local

module Maximum = Bin_prot.Std.Maximum

module Minimum = Bin_prot.Std.Minimum

val bin_unit : unit Bin_prot.Type_class.t

val bin_shape_unit : Bin_prot.Shape.t

val bin_writer_unit : unit Bin_prot.Type_class.writer

val bin_write_unit : unit Bin_prot.Write.writer

val bin_write_unit__local : unit Bin_prot.Write.writer_local

val bin_reader_unit : unit Bin_prot.Type_class.reader

val bin_read_unit : unit Bin_prot.Read.reader

val __bin_read_unit__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_bool : bool Bin_prot.Type_class.t

val bin_shape_bool : Bin_prot.Shape.t

val bin_writer_bool : bool Bin_prot.Type_class.writer

val bin_write_bool : bool Bin_prot.Write.writer

val bin_write_bool__local : bool Bin_prot.Write.writer_local

val bin_reader_bool : bool Bin_prot.Type_class.reader

val bin_read_bool : bool Bin_prot.Read.reader

val __bin_read_bool__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_string : string Bin_prot.Type_class.t

val bin_shape_string : Bin_prot.Shape.t

val bin_writer_string : string Bin_prot.Type_class.writer

val bin_write_string : string Bin_prot.Write.writer

val bin_write_string__local : string Bin_prot.Write.writer_local

val bin_reader_string : string Bin_prot.Type_class.reader

val bin_read_string : string Bin_prot.Read.reader

val __bin_read_string__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_bytes : bytes Bin_prot.Type_class.t

val bin_shape_bytes : Bin_prot.Shape.t

val bin_writer_bytes : bytes Bin_prot.Type_class.writer

val bin_write_bytes : bytes Bin_prot.Write.writer

val bin_write_bytes__local : bytes Bin_prot.Write.writer_local

val bin_reader_bytes : bytes Bin_prot.Type_class.reader

val bin_read_bytes : bytes Bin_prot.Read.reader

val __bin_read_bytes__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_char : char Bin_prot.Type_class.t

val bin_shape_char : Bin_prot.Shape.t

val bin_writer_char : char Bin_prot.Type_class.writer

val bin_write_char : char Bin_prot.Write.writer

val bin_write_char__local : char Bin_prot.Write.writer_local

val bin_reader_char : char Bin_prot.Type_class.reader

val bin_read_char : char Bin_prot.Read.reader

val __bin_read_char__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_int : int Bin_prot.Type_class.t

val bin_shape_int : Bin_prot.Shape.t

val bin_writer_int : int Bin_prot.Type_class.writer

val bin_write_int : int Bin_prot.Write.writer

val bin_write_int__local : int Bin_prot.Write.writer_local

val bin_reader_int : int Bin_prot.Type_class.reader

val bin_read_int : int Bin_prot.Read.reader

val __bin_read_int__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_float : float Bin_prot.Type_class.t

val bin_shape_float : Bin_prot.Shape.t

val bin_writer_float : float Bin_prot.Type_class.writer

val bin_write_float : float Bin_prot.Write.writer

val bin_write_float__local : float Bin_prot.Write.writer_local

val bin_reader_float : float Bin_prot.Type_class.reader

val bin_read_float : float Bin_prot.Read.reader

val __bin_read_float__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_floatarray : floatarray Bin_prot.Type_class.t

val bin_shape_floatarray : Bin_prot.Shape.t

val bin_writer_floatarray : floatarray Bin_prot.Type_class.writer

val bin_write_floatarray : floatarray Bin_prot.Write.writer

val bin_write_floatarray__local : floatarray Bin_prot.Write.writer_local

val bin_reader_floatarray : floatarray Bin_prot.Type_class.reader

val bin_read_floatarray : floatarray Bin_prot.Read.reader

val __bin_read_floatarray__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_int32 : int32 Bin_prot.Type_class.t

val bin_shape_int32 : Bin_prot.Shape.t

val bin_writer_int32 : int32 Bin_prot.Type_class.writer

val bin_write_int32 : int32 Bin_prot.Write.writer

val bin_write_int32__local : int32 Bin_prot.Write.writer_local

val bin_reader_int32 : int32 Bin_prot.Type_class.reader

val bin_read_int32 : int32 Bin_prot.Read.reader

val __bin_read_int32__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_int64 : int64 Bin_prot.Type_class.t

val bin_shape_int64 : Bin_prot.Shape.t

val bin_writer_int64 : int64 Bin_prot.Type_class.writer

val bin_write_int64 : int64 Bin_prot.Write.writer

val bin_write_int64__local : int64 Bin_prot.Write.writer_local

val bin_reader_int64 : int64 Bin_prot.Type_class.reader

val bin_read_int64 : int64 Bin_prot.Read.reader

val __bin_read_int64__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_nativeint : nativeint Bin_prot.Type_class.t

val bin_shape_nativeint : Bin_prot.Shape.t

val bin_writer_nativeint : nativeint Bin_prot.Type_class.writer

val bin_write_nativeint : nativeint Bin_prot.Write.writer

val bin_write_nativeint__local : nativeint Bin_prot.Write.writer_local

val bin_reader_nativeint : nativeint Bin_prot.Type_class.reader

val bin_read_nativeint : nativeint Bin_prot.Read.reader

val __bin_read_nativeint__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_ref : ('a, 'a ref) Bin_prot.Type_class.S1.t

val bin_shape_ref : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_ref : ('a, 'a ref) Bin_prot.Type_class.S1.writer

val bin_write_ref : ('a, 'a ref) Bin_prot.Write.writer1

val bin_write_ref__local : ('a, 'a ref) Bin_prot.Write.writer_local1

val bin_reader_ref : ('a, 'a ref) Bin_prot.Type_class.S1.reader

val bin_read_ref : ('a, 'a ref) Bin_prot.Read.reader1

val __bin_read_ref__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos ref -> 'c -> 'd

val bin_lazy_t : ('a, 'a lazy_t) Bin_prot.Type_class.S1.t

val bin_shape_lazy_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_lazy_t : ('a, 'a lazy_t) Bin_prot.Type_class.S1.writer

val bin_write_lazy_t : ('a, 'a lazy_t) Bin_prot.Write.writer1

val bin_write_lazy_t__local : ('a, 'a lazy_t) Bin_prot.Write.writer_local1

val bin_reader_lazy_t : ('a, 'a lazy_t) Bin_prot.Type_class.S1.reader

val bin_read_lazy_t : ('a, 'a lazy_t) Bin_prot.Read.reader1

val __bin_read_lazy_t__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos ref -> 'c -> 'd

val bin_lazy : ('a, 'a lazy_t) Bin_prot.Type_class.S1.t

val bin_shape_lazy : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_lazy : ('a, 'a lazy_t) Bin_prot.Type_class.S1.writer

val bin_write_lazy : ('a, 'a lazy_t) Bin_prot.Write.writer1

val bin_write_lazy__local : ('a, 'a lazy_t) Bin_prot.Write.writer_local1

val bin_reader_lazy : ('a, 'a lazy_t) Bin_prot.Type_class.S1.reader

val bin_read_lazy : ('a, 'a lazy_t) Bin_prot.Read.reader1

val __bin_read_lazy__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos ref -> 'c -> 'd

val bin_option : ('a, 'a option) Bin_prot.Type_class.S1.t

val bin_shape_option : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_option : ('a, 'a option) Bin_prot.Type_class.S1.writer

val bin_write_option : ('a, 'a option) Bin_prot.Write.writer1

val bin_write_option__local : ('a, 'a option) Bin_prot.Write.writer_local1

val bin_reader_option : ('a, 'a option) Bin_prot.Type_class.S1.reader

val bin_read_option : ('a, 'a option) Bin_prot.Read.reader1

val __bin_read_option__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos ref -> 'c -> 'd

val bin_list : ('a, 'a list) Bin_prot.Type_class.S1.t

val bin_shape_list : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_list : ('a, 'a list) Bin_prot.Type_class.S1.writer

val bin_write_list : ('a, 'a list) Bin_prot.Write.writer1

val bin_write_list__local : ('a, 'a list) Bin_prot.Write.writer_local1

val bin_reader_list : ('a, 'a list) Bin_prot.Type_class.S1.reader

val bin_read_list : ('a, 'a list) Bin_prot.Read.reader1

val __bin_read_list__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos ref -> 'c -> 'd

val bin_array : ('a, 'a array) Bin_prot.Type_class.S1.t

val bin_shape_array : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_array : ('a, 'a array) Bin_prot.Type_class.S1.writer

val bin_write_array : ('a, 'a array) Bin_prot.Write.writer1

val bin_write_array__local : ('a, 'a array) Bin_prot.Write.writer_local1

val bin_reader_array : ('a, 'a array) Bin_prot.Type_class.S1.reader

val bin_read_array : ('a, 'a array) Bin_prot.Read.reader1

val __bin_read_array__ : 'a -> 'b -> pos_ref:Bin_prot.Common.pos ref -> 'c -> 'd

val bin_hashtbl : ('a, 'b, ('a, 'b) Hashtbl.t) Bin_prot.Type_class.S2.t

val bin_shape_hashtbl : Bin_prot.Shape.t -> Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_writer_hashtbl : ('a, 'b, ('a, 'b) Hashtbl.t) Bin_prot.Type_class.S2.writer

val bin_write_hashtbl : ('a, 'b, ('a, 'b) Hashtbl.t) Bin_prot.Write.writer2

val bin_reader_hashtbl : ('a, 'b, ('a, 'b) Hashtbl.t) Bin_prot.Type_class.S2.reader

val bin_read_hashtbl : ('a, 'b, ('a, 'b) Hashtbl.t) Bin_prot.Read.reader2

val __bin_read_hashtbl__ : 'a -> 'b -> 'c -> pos_ref:Bin_prot.Common.pos ref -> 'd -> 'e

val bin_bigstring : Bin_prot.Common.buf Bin_prot.Type_class.t

val bin_shape_bigstring : Bin_prot.Shape.t

val bin_writer_bigstring : Bin_prot.Common.buf Bin_prot.Type_class.writer

val bin_write_bigstring : Bin_prot.Common.buf Bin_prot.Write.writer

val bin_write_bigstring__local : Bin_prot.Common.buf Bin_prot.Write.writer_local

val bin_reader_bigstring : Bin_prot.Common.buf Bin_prot.Type_class.reader

val bin_read_bigstring : Bin_prot.Common.buf Bin_prot.Read.reader

val __bin_read_bigstring__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_mat : Bin_prot.Common.mat Bin_prot.Type_class.t

val bin_shape_mat : Bin_prot.Shape.t

val bin_writer_mat : Bin_prot.Common.mat Bin_prot.Type_class.writer

val bin_write_mat : Bin_prot.Common.mat Bin_prot.Write.writer

val bin_write_mat__local : Bin_prot.Common.mat Bin_prot.Write.writer_local

val bin_reader_mat : Bin_prot.Common.mat Bin_prot.Type_class.reader

val bin_read_mat : Bin_prot.Common.mat Bin_prot.Read.reader

val __bin_read_mat__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_float32_mat : Bin_prot.Common.mat32 Bin_prot.Type_class.t

val bin_shape_float32_mat : Bin_prot.Shape.t

val bin_writer_float32_mat : Bin_prot.Common.mat32 Bin_prot.Type_class.writer

val bin_write_float32_mat : Bin_prot.Common.mat32 Bin_prot.Write.writer

val bin_write_float32_mat__local : Bin_prot.Common.mat32 Bin_prot.Write.writer_local

val bin_reader_float32_mat : Bin_prot.Common.mat32 Bin_prot.Type_class.reader

val bin_read_float32_mat : Bin_prot.Common.mat32 Bin_prot.Read.reader

val __bin_read_float32_mat__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_float64_mat : Bin_prot.Common.mat64 Bin_prot.Type_class.t

val bin_shape_float64_mat : Bin_prot.Shape.t

val bin_writer_float64_mat : Bin_prot.Common.mat64 Bin_prot.Type_class.writer

val bin_write_float64_mat : Bin_prot.Common.mat64 Bin_prot.Write.writer

val bin_write_float64_mat__local : Bin_prot.Common.mat64 Bin_prot.Write.writer_local

val bin_reader_float64_mat : Bin_prot.Common.mat64 Bin_prot.Type_class.reader

val bin_read_float64_mat : Bin_prot.Common.mat64 Bin_prot.Read.reader

val __bin_read_float64_mat__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_vec : Bin_prot.Common.vec Bin_prot.Type_class.t

val bin_shape_vec : Bin_prot.Shape.t

val bin_writer_vec : Bin_prot.Common.vec Bin_prot.Type_class.writer

val bin_write_vec : Bin_prot.Common.vec Bin_prot.Write.writer

val bin_write_vec__local : Bin_prot.Common.vec Bin_prot.Write.writer_local

val bin_reader_vec : Bin_prot.Common.vec Bin_prot.Type_class.reader

val bin_read_vec : Bin_prot.Common.vec Bin_prot.Read.reader

val __bin_read_vec__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_float32_vec : Bin_prot.Common.vec32 Bin_prot.Type_class.t

val bin_shape_float32_vec : Bin_prot.Shape.t

val bin_writer_float32_vec : Bin_prot.Common.vec32 Bin_prot.Type_class.writer

val bin_write_float32_vec : Bin_prot.Common.vec32 Bin_prot.Write.writer

val bin_write_float32_vec__local : Bin_prot.Common.vec32 Bin_prot.Write.writer_local

val bin_reader_float32_vec : Bin_prot.Common.vec32 Bin_prot.Type_class.reader

val bin_read_float32_vec : Bin_prot.Common.vec32 Bin_prot.Read.reader

val __bin_read_float32_vec__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

val bin_float64_vec : Bin_prot.Common.vec64 Bin_prot.Type_class.t

val bin_shape_float64_vec : Bin_prot.Shape.t

val bin_writer_float64_vec : Bin_prot.Common.vec64 Bin_prot.Type_class.writer

val bin_write_float64_vec : Bin_prot.Common.vec64 Bin_prot.Write.writer

val bin_write_float64_vec__local : Bin_prot.Common.vec64 Bin_prot.Write.writer_local

val bin_reader_float64_vec : Bin_prot.Common.vec64 Bin_prot.Type_class.reader

val bin_read_float64_vec : Bin_prot.Common.vec64 Bin_prot.Read.reader

val __bin_read_float64_vec__ : 'a -> pos_ref:Bin_prot.Common.pos ref -> 'b -> 'c

include module type of struct include Base.Hash.Builtin end

val hash_fold_nativeint : nativeint Base.Hash.folder

val hash_fold_int64 : int64 Base.Hash.folder

val hash_fold_int32 : int32 Base.Hash.folder

val hash_fold_char : char Base.Hash.folder

val hash_fold_int : int Base.Hash.folder

val hash_fold_bool : bool Base.Hash.folder

val hash_fold_string : string Base.Hash.folder

val hash_fold_float : float Base.Hash.folder

val hash_fold_unit : unit Base.Hash.folder

val hash_fold_option : 'a Base.Hash.folder -> 'a option Base.Hash.folder

val hash_fold_list : 'a Base.Hash.folder -> 'a list Base.Hash.folder

val hash_fold_lazy_t : 'a Base.Hash.folder -> 'a lazy_t Base.Hash.folder

val hash_fold_ref_frozen : 'a Base.Hash.folder -> 'a ref Base.Hash.folder

Hash support for array and ref is provided, but is potentially DANGEROUS, since it incorporates the current contents of the array/ref into the hash value. Because of this we add a _frozen suffix to the function name.

Hash support for string is also potentially DANGEROUS, but strings are mutated less often, so we don't append _frozen to it.

Also note that we don't support bytes.

val hash_fold_array_frozen : 'a Base.Hash.folder -> 'a array Base.Hash.folder

val hash_nativeint : nativeint -> Base.Hash.hash_value

val hash_int64 : int64 -> Base.Hash.hash_value

val hash_int32 : int32 -> Base.Hash.hash_value

val hash_char : char -> Base.Hash.hash_value

val hash_int : int -> Base.Hash.hash_value

val hash_bool : bool -> Base.Hash.hash_value

val hash_string : string -> Base.Hash.hash_value

val hash_float : float -> Base.Hash.hash_value

val hash_unit : unit -> Base.Hash.hash_value

include module type of struct include Stable_witness.Export end

val stable_witness_array : 'a Stable_witness.t -> 'a array Stable_witness.t

val stable_witness_bool : bool Stable_witness.t

val stable_witness_bytes : bytes Stable_witness.t

val stable_witness_char : char Stable_witness.t

val stable_witness_exn : exn Stable_witness.t

val stable_witness_float : float Stable_witness.t

val stable_witness_int : int Stable_witness.t

val stable_witness_int32 : int32 Stable_witness.t

val stable_witness_int64 : int64 Stable_witness.t

val stable_witness_lazy_t : 'a Stable_witness.t -> 'a lazy_t Stable_witness.t

val stable_witness_list : 'a Stable_witness.t -> 'a list Stable_witness.t

val stable_witness_nativeint : nativeint Stable_witness.t

val stable_witness_option : 'a Stable_witness.t -> 'a option Stable_witness.t

val stable_witness_ref : 'a Stable_witness.t -> 'a ref Stable_witness.t

val stable_witness_string : string Stable_witness.t

val stable_witness_unit : unit Stable_witness.t

include sig ... end

val sexp_of_array : ('a -> Sexplib0.Sexp.t) -> 'a Base.Array.t -> Sexplib0.Sexp.t

val array_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a Base.Array.t

val array_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'a Base.Array.t Sexplib0.Sexp_grammar.t

val sexp_of_bool : Base.Bool.t -> Sexplib0.Sexp.t

val bool_of_sexp : Sexplib0.Sexp.t -> Base.Bool.t

val bool_sexp_grammar : Base.Bool.t Sexplib0.Sexp_grammar.t

val sexp_of_char : Base.Char.t -> Sexplib0.Sexp.t

val char_of_sexp : Sexplib0.Sexp.t -> Base.Char.t

val char_sexp_grammar : Base.Char.t Sexplib0.Sexp_grammar.t

val sexp_of_exn : Base.Exn.t -> Sexplib0.Sexp.t

val sexp_of_float : Base.Float.t -> Sexplib0.Sexp.t

val float_of_sexp : Sexplib0.Sexp.t -> Base.Float.t

val float_sexp_grammar : Base.Float.t Sexplib0.Sexp_grammar.t

val sexp_of_int : Base.Int.t -> Sexplib0.Sexp.t

val int_of_sexp : Sexplib0.Sexp.t -> Base.Int.t

val int_sexp_grammar : Base.Int.t Sexplib0.Sexp_grammar.t

val sexp_of_int32 : Base.Int32.t -> Sexplib0.Sexp.t

val int32_of_sexp : Sexplib0.Sexp.t -> Base.Int32.t

val int32_sexp_grammar : Base.Int32.t Sexplib0.Sexp_grammar.t

val sexp_of_int64 : Base.Int64.t -> Sexplib0.Sexp.t

val int64_of_sexp : Sexplib0.Sexp.t -> Base.Int64.t

val int64_sexp_grammar : Base.Int64.t Sexplib0.Sexp_grammar.t

val sexp_of_list : ('a -> Sexplib0.Sexp.t) -> 'a Base.List.t -> Sexplib0.Sexp.t

val list_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a Base.List.t

val list_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'a Base.List.t Sexplib0.Sexp_grammar.t

val sexp_of_nativeint : Base.Nativeint.t -> Sexplib0.Sexp.t

val nativeint_of_sexp : Sexplib0.Sexp.t -> Base.Nativeint.t

val nativeint_sexp_grammar : Base.Nativeint.t Sexplib0.Sexp_grammar.t

val sexp_of_option : ('a -> Sexplib0.Sexp.t) -> 'a Base.Option.t -> Sexplib0.Sexp.t

val option_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a Base.Option.t

val option_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'a Base.Option.t Sexplib0.Sexp_grammar.t

val sexp_of_ref : ('a -> Sexplib0.Sexp.t) -> 'a Base.Ref.t -> Sexplib0.Sexp.t

val ref_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a Base.Ref.t

val ref_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'a Base.Ref.t Sexplib0.Sexp_grammar.t

val sexp_of_string : Base.String.t -> Sexplib0.Sexp.t

val string_of_sexp : Sexplib0.Sexp.t -> Base.String.t

val string_sexp_grammar : Base.String.t Sexplib0.Sexp_grammar.t

val sexp_of_bytes : Base.Bytes.t -> Sexplib0.Sexp.t

val bytes_of_sexp : Sexplib0.Sexp.t -> Base.Bytes.t

val bytes_sexp_grammar : Base.Bytes.t Sexplib0.Sexp_grammar.t

val sexp_of_unit : Base.Unit.t -> Sexplib0.Sexp.t

val unit_of_sexp : Sexplib0.Sexp.t -> Base.Unit.t

val unit_sexp_grammar : Base.Unit.t Sexplib0.Sexp_grammar.t