Module Core.Core_stable

module type Stable_int63able = sig ... end

module type Stable_int63able_without_comparator = sig ... end

module type Stable_int63able_with_witness = sig ... end

module type Stable_int63able__local = sig ... end

module type Stable_int63able_without_comparator__local = sig ... end

module type Stable_int63able_with_witness__local = sig ... end

module type Stable__portable = sig ... end

module type Stable = sig ... end

module type Stable_without_comparator = sig ... end

module type Stable1 = sig ... end

module type Stable2 = sig ... end

module type Stable3 = sig ... end

module type Stable4 = sig ... end

module type Stable_with_witness__portable = sig ... end

module type Stable_with_witness = sig ... end

module type Stable_without_comparator_with_witness = sig ... end

module type Stable1_with_witness = sig ... end

module type Stable2_with_witness = sig ... end

module type Stable3_with_witness = sig ... end

module type Stable4_with_witness = sig ... end

module type Stable__local__portable = sig ... end

module type Stable__local = sig ... end

module type Stable_without_comparator__local = sig ... end

module type Stable1__local = sig ... end

module type Stable2__local = sig ... end

module type Stable3__local = sig ... end

module type Stable4__local = sig ... end

module type Stable_with_witness__local__portable = sig ... end

module type Stable_with_witness__local = sig ... end

module type Stable_without_comparator_with_witness__local = sig ... end

module type Stable1_with_witness__local = sig ... end

module type Stable2_with_witness__local = sig ... end

module type Stable3_with_witness__local = sig ... end

module type Stable4_with_witness__local = sig ... end

module Unit_test = Stable_unit_test.Make

module Make_stable : sig ... end

module Bigstring = Bigstring.Stable

module Binable = Binable.Stable

module Blang = Blang.Stable

module Bool = Bool.Stable

module Byte_units = Byte_units.Stable

module Bytes = Bytes.Stable

module Char = Char.Stable

module Comparable = Comparable.Stable

module Comparator = Comparator.Stable

module Date = Date.Stable

module Day_of_week = Day_of_week.Stable

module Either = Either.Stable

module Error = Error.Stable

module Fdeque = Fdeque.Stable

module Filename = Filename.Stable

module Float = Float.Stable

module Float_with_finite_only_serialization = Float_with_finite_only_serialization.Stable

module Fqueue = Fqueue.Stable

module Gc = Gc.Stable

module Hashable = Hashable.Stable

module Host_and_port = Host_and_port.Stable

module Iarray = Iarray.Stable

module Info = Info.Stable

module Int = Int.Stable

module Int63 = Int63.Stable

module Lazy = Lazy.Stable

module List = List.Stable

module Map = Map.Stable

module Maybe_bound = Maybe_bound.Stable

module Md5 = Md5.Stable

module Modes = Modes.Stable

module Month = Month.Stable

module Nonempty_list = Nonempty_list.Stable

module Nothing = Nothing.Stable

module Option = Option.Stable

module Or_error = Or_error.Stable

module Percent = Percent.Stable

module Perms = Perms.Stable

module Pid = Pid.Stable

module Portable_lazy = Portable_lazy.Stable

module Queue = Queue.Stable

module Result = Result.Stable

module Set = Set.Stable

module Sexp = Sexp.Stable

module Sexpable = Sexpable.Stable

module Source_code_position = Source_code_position.Stable

module String = String.Stable

module String_id = String_id.Stable

module Time_float = Time_float.Stable

module Time_ns = Time_ns.Stable

module Timezone = Timezone.Stable

module Uchar = Uchar.Stable

module Unit = Unit.Stable

include module type of struct include Bin_prot.Std end

type ('a : any) sizer = 'a -> Base.int

type (('a : any), ('b : any)) sizer1 = 'a sizer -> 'b sizer

type (('a : any), ('b : any), ('c : any)) sizer2 = 'a sizer -> ('b, 'c) sizer1

type (('a : any), ('b : any), ('c : any), ('d : any)) sizer3 = 'a sizer -> ('b, 'c, 'd) sizer2

type ('a : any) sizer__local = 'a @ local -> Base.int

type (('a : any), ('b : any)) sizer1__local = 'a sizer__local -> 'b sizer__local

type (('a : any), ('b : any), ('c : any)) sizer2__local = 'a sizer__local -> ('b, 'c) sizer1__local

type (('a : any), ('b : any), ('c : any), ('d : any)) sizer3__local = 'a sizer__local -> ('b, 'c, 'd) sizer2__local

val bin_size_unit : Base.unit sizer

val bin_size_bool : Base.bool sizer

val bin_size_string : Base.string sizer

val bin_size_bytes : Base.bytes sizer

val bin_size_char : Base.char sizer

val bin_size_int : Base.int sizer

val bin_size_float : Base.float sizer

val bin_size_int32 : Base.int32 sizer

val bin_size_int64 : Base.int64 sizer

val bin_size_nativeint : Base.nativeint sizer

val bin_size_nat0 : Bin_prot.Nat0.t sizer

val bin_size_ref : 'a. ('a, 'a Base.ref) sizer1

val bin_size_lazy_t : ('a, 'a lazy_t) sizer1

val bin_size_lazy : ('a, 'a lazy_t) sizer1

val bin_size_option : 'a. ('a, 'a Base.option) sizer1

val bin_size_or_null : ('a, 'a Base.or_null) sizer1

val bin_size_pair : ('a, 'b, 'a * 'b) sizer2

val bin_size_triple : ('a, 'b, 'c, 'a * 'b * 'c) sizer3

val bin_size_list : 'a. ('a, 'a Base.list) sizer1

val bin_size_array : 'a. ('a, 'a Base.array) sizer1

val bin_size_iarray : 'a. ('a, 'a Base.iarray) sizer1

val bin_size_float32_vec : Bin_prot.Common.vec32 sizer

val bin_size_float64_vec : Bin_prot.Common.vec64 sizer

val bin_size_vec : Bin_prot.Common.vec sizer

val bin_size_float32_mat : Bin_prot.Common.mat32 sizer

val bin_size_float64_mat : Bin_prot.Common.mat64 sizer

val bin_size_mat : Bin_prot.Common.mat sizer

val bin_size_bigstring : Bin_prot.Common.buf sizer

val bin_size_floatarray : floatarray sizer

val bin_size_variant_int : Base.int sizer

val bin_size_int_8bit : Base.int sizer

val bin_size_int_16bit : Base.int sizer

val bin_size_int_32bit : Base.int sizer

val bin_size_int_64bit : Base.int sizer

val bin_size_int32_bits : Base.int32 sizer

val bin_size_int64_bits : Base.int64 sizer

val bin_size_network16_int : Base.int sizer

val bin_size_network32_int : Base.int sizer

val bin_size_network32_int32 : Base.int32 sizer

val bin_size_network64_int : Base.int sizer

val bin_size_network64_int64 : Base.int64 sizer

val bin_size_md5 : Md5_lib.t sizer

module Maximum = Bin_prot.Std.Maximum

module Minimum = Bin_prot.Std.Minimum

Monomorphic types

shapes

val bin_shape_bigstring : Bin_shape.t @@ portable

val bin_shape_bool : Bin_shape.t @@ portable

val bin_shape_bytes : Bin_shape.t @@ portable

val bin_shape_char : Bin_shape.t @@ portable

val bin_shape_float : Bin_shape.t @@ portable

val bin_shape_float32_mat : Bin_shape.t @@ portable

val bin_shape_float32_vec : Bin_shape.t @@ portable

val bin_shape_float64_mat : Bin_shape.t @@ portable

val bin_shape_float64_vec : Bin_shape.t @@ portable

val bin_shape_floatarray : Bin_shape.t @@ portable

val bin_shape_int : Bin_shape.t @@ portable

val bin_shape_int32 : Bin_shape.t @@ portable

val bin_shape_int64 : Bin_shape.t @@ portable

val bin_shape_mat : Bin_shape.t @@ portable

val bin_shape_nativeint : Bin_shape.t @@ portable

val bin_shape_string : Bin_shape.t @@ portable

val bin_shape_unit : Bin_shape.t @@ portable

val bin_shape_vec : Bin_shape.t @@ portable

readers

val bin_read_bigstring : Bin_prot.Common.buf Bin_prot.Read.reader @@ portable

val bin_read_bool : Base.bool Bin_prot.Read.reader @@ portable

val bin_read_bytes : Base.bytes Bin_prot.Read.reader @@ portable

val bin_read_char : Base.char Bin_prot.Read.reader @@ portable

val bin_read_float : Base.float Bin_prot.Read.reader @@ portable

val bin_read_float32_mat : Bin_prot.Common.mat32 Bin_prot.Read.reader @@ portable

val bin_read_float32_vec : Bin_prot.Common.vec32 Bin_prot.Read.reader @@ portable

val bin_read_float64_mat : Bin_prot.Common.mat64 Bin_prot.Read.reader @@ portable

val bin_read_float64_vec : Bin_prot.Common.vec64 Bin_prot.Read.reader @@ portable

val bin_read_floatarray : floatarray Bin_prot.Read.reader @@ portable

val bin_read_int : Base.int Bin_prot.Read.reader @@ portable

val bin_read_int32 : Base.int32 Bin_prot.Read.reader @@ portable

val bin_read_int64 : Base.int64 Bin_prot.Read.reader @@ portable

val bin_read_mat : Bin_prot.Common.mat Bin_prot.Read.reader @@ portable

val bin_read_nativeint : Base.nativeint Bin_prot.Read.reader @@ portable

val bin_read_string : Base.string Bin_prot.Read.reader @@ portable

val bin_read_unit : Base.unit Bin_prot.Read.reader @@ portable

val bin_read_vec : Bin_prot.Common.vec Bin_prot.Read.reader @@ portable

vtag readers

val __bin_read_bigstring__ : Bin_prot.Common.buf Bin_prot.Read.vtag_reader @@ portable

val __bin_read_bool__ : Base.bool Bin_prot.Read.vtag_reader @@ portable

val __bin_read_bytes__ : Base.bytes Bin_prot.Read.vtag_reader @@ portable

val __bin_read_char__ : Base.char Bin_prot.Read.vtag_reader @@ portable

val __bin_read_float32_mat__ : Bin_prot.Common.mat32 Bin_prot.Read.vtag_reader @@ portable

val __bin_read_float32_vec__ : Bin_prot.Common.vec32 Bin_prot.Read.vtag_reader @@ portable

val __bin_read_float64_mat__ : Bin_prot.Common.mat64 Bin_prot.Read.vtag_reader @@ portable

val __bin_read_float64_vec__ : Bin_prot.Common.vec64 Bin_prot.Read.vtag_reader @@ portable

val __bin_read_float__ : Base.float Bin_prot.Read.vtag_reader @@ portable

val __bin_read_floatarray__ : floatarray Bin_prot.Read.vtag_reader @@ portable

val __bin_read_int32__ : Base.int32 Bin_prot.Read.vtag_reader @@ portable

val __bin_read_int64__ : Base.int64 Bin_prot.Read.vtag_reader @@ portable

val __bin_read_int__ : Base.int Bin_prot.Read.vtag_reader @@ portable

val __bin_read_mat__ : Bin_prot.Common.mat Bin_prot.Read.vtag_reader @@ portable

val __bin_read_nativeint__ : Base.nativeint Bin_prot.Read.vtag_reader @@ portable

val __bin_read_string__ : Base.string Bin_prot.Read.vtag_reader @@ portable

val __bin_read_unit__ : Base.unit Bin_prot.Read.vtag_reader @@ portable

val __bin_read_vec__ : Bin_prot.Common.vec Bin_prot.Read.vtag_reader @@ portable

writers

val bin_write_bigstring : Bin_prot.Common.buf Bin_prot.Write.writer

val bin_write_bool : Base.bool Bin_prot.Write.writer

val bin_write_bytes : Base.bytes Bin_prot.Write.writer

val bin_write_char : Base.char Bin_prot.Write.writer

val bin_write_float : Base.float Bin_prot.Write.writer

val bin_write_float32_mat : Bin_prot.Common.mat32 Bin_prot.Write.writer

val bin_write_float32_vec : Bin_prot.Common.vec32 Bin_prot.Write.writer

val bin_write_float64_mat : Bin_prot.Common.mat64 Bin_prot.Write.writer

val bin_write_float64_vec : Bin_prot.Common.vec64 Bin_prot.Write.writer

val bin_write_floatarray : floatarray Bin_prot.Write.writer

val bin_write_int : Base.int Bin_prot.Write.writer

val bin_write_int32 : Base.int32 Bin_prot.Write.writer

val bin_write_int64 : Base.int64 Bin_prot.Write.writer

val bin_write_mat : Bin_prot.Common.mat Bin_prot.Write.writer

val bin_write_nativeint : Base.nativeint Bin_prot.Write.writer

val bin_write_string : Base.string Bin_prot.Write.writer

val bin_write_unit : Base.unit Bin_prot.Write.writer

val bin_write_vec : Bin_prot.Common.vec Bin_prot.Write.writer

type class readers

val bin_reader_bigstring : Bin_prot.Common.buf Bin_prot.Type_class.reader @@ portable

val bin_reader_bool : Base.bool Bin_prot.Type_class.reader @@ portable

val bin_reader_bytes : Base.bytes Bin_prot.Type_class.reader @@ portable

val bin_reader_char : Base.char Bin_prot.Type_class.reader @@ portable

val bin_reader_float : Base.float Bin_prot.Type_class.reader @@ portable

val bin_reader_float32_mat : Bin_prot.Common.mat32 Bin_prot.Type_class.reader @@ portable

val bin_reader_float32_vec : Bin_prot.Common.vec32 Bin_prot.Type_class.reader @@ portable

val bin_reader_float64_mat : Bin_prot.Common.mat64 Bin_prot.Type_class.reader @@ portable

val bin_reader_float64_vec : Bin_prot.Common.vec64 Bin_prot.Type_class.reader @@ portable

val bin_reader_floatarray : floatarray Bin_prot.Type_class.reader @@ portable

val bin_reader_int : Base.int Bin_prot.Type_class.reader @@ portable

val bin_reader_int32 : Base.int32 Bin_prot.Type_class.reader @@ portable

val bin_reader_int64 : Base.int64 Bin_prot.Type_class.reader @@ portable

val bin_reader_mat : Bin_prot.Common.mat Bin_prot.Type_class.reader @@ portable

val bin_reader_nativeint : Base.nativeint Bin_prot.Type_class.reader @@ portable

val bin_reader_string : Base.string Bin_prot.Type_class.reader @@ portable

val bin_reader_unit : Base.unit Bin_prot.Type_class.reader @@ portable

val bin_reader_vec : Bin_prot.Common.vec Bin_prot.Type_class.reader @@ portable

type class writers

val bin_writer_bigstring : Bin_prot.Common.buf Bin_prot.Type_class.writer @@ portable

val bin_writer_bool : Base.bool Bin_prot.Type_class.writer @@ portable

val bin_writer_bytes : Base.bytes Bin_prot.Type_class.writer @@ portable

val bin_writer_char : Base.char Bin_prot.Type_class.writer @@ portable

val bin_writer_float : Base.float Bin_prot.Type_class.writer @@ portable

val bin_writer_float32_mat : Bin_prot.Common.mat32 Bin_prot.Type_class.writer @@ portable

val bin_writer_float32_vec : Bin_prot.Common.vec32 Bin_prot.Type_class.writer @@ portable

val bin_writer_float64_mat : Bin_prot.Common.mat64 Bin_prot.Type_class.writer @@ portable

val bin_writer_float64_vec : Bin_prot.Common.vec64 Bin_prot.Type_class.writer @@ portable

val bin_writer_floatarray : floatarray Bin_prot.Type_class.writer @@ portable

val bin_writer_int : Base.int Bin_prot.Type_class.writer @@ portable

val bin_writer_int32 : Base.int32 Bin_prot.Type_class.writer @@ portable

val bin_writer_int64 : Base.int64 Bin_prot.Type_class.writer @@ portable

val bin_writer_mat : Bin_prot.Common.mat Bin_prot.Type_class.writer @@ portable

val bin_writer_nativeint : Base.nativeint Bin_prot.Type_class.writer @@ portable

val bin_writer_string : Base.string Bin_prot.Type_class.writer @@ portable

val bin_writer_unit : Base.unit Bin_prot.Type_class.writer @@ portable

val bin_writer_vec : Bin_prot.Common.vec Bin_prot.Type_class.writer @@ portable

type classes

val bin_bigstring : Bin_prot.Common.buf Bin_prot.Type_class.t @@ portable

val bin_bool : Base.bool Bin_prot.Type_class.t @@ portable

val bin_bytes : Base.bytes Bin_prot.Type_class.t @@ portable

val bin_char : Base.char Bin_prot.Type_class.t @@ portable

val bin_float : Base.float Bin_prot.Type_class.t @@ portable

val bin_float32_mat : Bin_prot.Common.mat32 Bin_prot.Type_class.t @@ portable

val bin_float32_vec : Bin_prot.Common.vec32 Bin_prot.Type_class.t @@ portable

val bin_float64_mat : Bin_prot.Common.mat64 Bin_prot.Type_class.t @@ portable

val bin_float64_vec : Bin_prot.Common.vec64 Bin_prot.Type_class.t @@ portable

val bin_floatarray : floatarray Bin_prot.Type_class.t @@ portable

val bin_int : Base.int Bin_prot.Type_class.t @@ portable

val bin_int32 : Base.int32 Bin_prot.Type_class.t @@ portable

val bin_int64 : Base.int64 Bin_prot.Type_class.t @@ portable

val bin_mat : Bin_prot.Common.mat Bin_prot.Type_class.t @@ portable

val bin_nativeint : Base.nativeint Bin_prot.Type_class.t @@ portable

val bin_string : Base.string Bin_prot.Type_class.t @@ portable

val bin_unit : Base.unit Bin_prot.Type_class.t @@ portable

val bin_vec : Bin_prot.Common.vec Bin_prot.Type_class.t @@ portable

polymorphic types

shapes

val bin_shape_array : Bin_shape.t -> Bin_shape.t @@ portable

val bin_shape_iarray : Bin_shape.t -> Bin_shape.t @@ portable

val bin_shape_lazy : Bin_shape.t -> Bin_shape.t @@ portable

val bin_shape_lazy_t : Bin_shape.t -> Bin_shape.t @@ portable

val bin_shape_list : Bin_shape.t -> Bin_shape.t @@ portable

val bin_shape_option : Bin_shape.t -> Bin_shape.t @@ portable

val bin_shape_or_null : Bin_shape.t -> Bin_shape.t @@ portable

val bin_shape_ref : Bin_shape.t -> Bin_shape.t @@ portable

readers

val bin_read_array : 'a. ('a, 'a Base.array) Bin_prot.Read.reader1 @@ portable

val bin_read_iarray : 'a. ('a, 'a Base.iarray) Bin_prot.Read.reader1 @@ portable

val bin_read_lazy : ('a, 'a lazy_t) Bin_prot.Read.reader1 @@ portable

val bin_read_lazy_t : ('a, 'a lazy_t) Bin_prot.Read.reader1 @@ portable

val bin_read_list : 'a. ('a, 'a Base.list) Bin_prot.Read.reader1 @@ portable

val bin_read_option : 'a. ('a, 'a Base.option) Bin_prot.Read.reader1 @@ portable

val bin_read_or_null : ('a, 'a Base.or_null) Bin_prot.Read.reader1 @@ portable

val bin_read_ref : 'a. ('a, 'a Base.ref) Bin_prot.Read.reader1 @@ portable

vtag readers

val __bin_read_array__ : 'a. ('a, 'a Base.array) Bin_prot.Read.vtag_reader1 @@ portable

val __bin_read_iarray__ : 'a. ('a, 'a Base.iarray) Bin_prot.Read.vtag_reader1 @@ portable

val __bin_read_lazy__ : ('a, 'a lazy_t) Bin_prot.Read.vtag_reader1 @@ portable

val __bin_read_lazy_t__ : ('a, 'a lazy_t) Bin_prot.Read.vtag_reader1 @@ portable

val __bin_read_list__ : 'a. ('a, 'a Base.list) Bin_prot.Read.vtag_reader1 @@ portable

val __bin_read_option__ : 'a. ('a, 'a Base.option) Bin_prot.Read.vtag_reader1 @@ portable

val __bin_read_or_null__ : ('a, 'a Base.or_null) Bin_prot.Read.vtag_reader1 @@ portable

val __bin_read_ref__ : 'a. ('a, 'a Base.ref) Bin_prot.Read.vtag_reader1 @@ portable

writers

val bin_write_array : 'a. ('a, 'a Base.array) Bin_prot.Write.writer1

val bin_write_iarray : 'a. ('a, 'a Base.iarray) Bin_prot.Write.writer1

val bin_write_lazy : ('a, 'a lazy_t) Bin_prot.Write.writer1

val bin_write_lazy_t : ('a, 'a lazy_t) Bin_prot.Write.writer1

val bin_write_list : 'a. ('a, 'a Base.list) Bin_prot.Write.writer1

val bin_write_option : 'a. ('a, 'a Base.option) Bin_prot.Write.writer1

val bin_write_or_null : ('a, 'a Base.or_null) Bin_prot.Write.writer1

val bin_write_ref : 'a. ('a, 'a Base.ref) Bin_prot.Write.writer1

type class readers

val bin_reader_array : 'a. ('a, 'a Base.array) Bin_prot.Type_class.S1.reader @@ portable

val bin_reader_iarray : 'a. ('a, 'a Base.iarray) Bin_prot.Type_class.S1.reader @@ portable

val bin_reader_lazy : ('a, 'a lazy_t) Bin_prot.Type_class.S1.reader @@ portable

val bin_reader_lazy_t : ('a, 'a lazy_t) Bin_prot.Type_class.S1.reader @@ portable

val bin_reader_list : 'a. ('a, 'a Base.list) Bin_prot.Type_class.S1.reader @@ portable

val bin_reader_option : 'a. ('a, 'a Base.option) Bin_prot.Type_class.S1.reader @@ portable

val bin_reader_or_null : ('a, 'a Base.or_null) Bin_prot.Type_class.S1.reader @@ portable

val bin_reader_ref : 'a. ('a, 'a Base.ref) Bin_prot.Type_class.S1.reader @@ portable

type class writers

val bin_writer_array : 'a. ('a, 'a Base.array) Bin_prot.Type_class.S1.writer @@ portable

val bin_writer_iarray : 'a. ('a, 'a Base.iarray) Bin_prot.Type_class.S1.writer @@ portable

val bin_writer_lazy : ('a, 'a lazy_t) Bin_prot.Type_class.S1.writer @@ portable

val bin_writer_lazy_t : ('a, 'a lazy_t) Bin_prot.Type_class.S1.writer @@ portable

val bin_writer_list : 'a. ('a, 'a Base.list) Bin_prot.Type_class.S1.writer @@ portable

val bin_writer_option : 'a. ('a, 'a Base.option) Bin_prot.Type_class.S1.writer @@ portable

val bin_writer_or_null : ('a, 'a Base.or_null) Bin_prot.Type_class.S1.writer @@ portable

val bin_writer_ref : 'a. ('a, 'a Base.ref) Bin_prot.Type_class.S1.writer @@ portable

type classes

val bin_array : 'a. ('a, 'a Base.array) Bin_prot.Type_class.S1.t @@ portable

val bin_iarray : 'a. ('a, 'a Base.iarray) Bin_prot.Type_class.S1.t @@ portable

val bin_lazy : ('a, 'a lazy_t) Bin_prot.Type_class.S1.t @@ portable

val bin_lazy_t : ('a, 'a lazy_t) Bin_prot.Type_class.S1.t @@ portable

val bin_list : 'a. ('a, 'a Base.list) Bin_prot.Type_class.S1.t @@ portable

val bin_option : 'a. ('a, 'a Base.option) Bin_prot.Type_class.S1.t @@ portable

val bin_or_null : ('a, 'a Base.or_null) Bin_prot.Type_class.S1.t @@ portable

val bin_ref : 'a. ('a, 'a Base.ref) Bin_prot.Type_class.S1.t @@ portable

include module type of struct include Base.Hash.Builtin end

val hash_fold_nativeint : Base_internalhash_types.state -> nativeint -> Base_internalhash_types.state @@ portable

val hash_fold_int64 : Base_internalhash_types.state -> int64 -> Base_internalhash_types.state @@ portable

val hash_fold_int32 : Base_internalhash_types.state -> int32 -> Base_internalhash_types.state @@ portable

val hash_fold_char : Base_internalhash_types.state -> char -> Base_internalhash_types.state @@ portable

val hash_fold_int : Base_internalhash_types.state -> int -> Base_internalhash_types.state @@ portable

val hash_fold_bool : Base_internalhash_types.state -> bool -> Base_internalhash_types.state @@ portable

val hash_fold_string : Base_internalhash_types.state -> string -> Base_internalhash_types.state @@ portable

val hash_fold_float : Base_internalhash_types.state -> float -> Base_internalhash_types.state @@ portable

val hash_fold_unit : Base_internalhash_types.state -> unit -> Base_internalhash_types.state @@ portable

val hash_fold_option : 'a. (Base_internalhash_types.state -> 'a -> Base_internalhash_types.state) -> Base_internalhash_types.state -> 'a option -> Base_internalhash_types.state @@ portable

val hash_fold_or_null : (Base_internalhash_types.state -> 'a -> Base_internalhash_types.state) -> Base_internalhash_types.state -> 'a Basement.Or_null_shim.t -> Base_internalhash_types.state @@ portable

val hash_fold_list : 'a. (Base_internalhash_types.state -> 'a -> Base_internalhash_types.state) -> Base_internalhash_types.state -> 'a list -> Base_internalhash_types.state @@ portable

val hash_fold_lazy_t : (Base_internalhash_types.state -> 'a -> Base_internalhash_types.state) -> Base_internalhash_types.state -> 'a lazy_t -> Base_internalhash_types.state @@ portable

val hash_fold_ref_frozen : (Base_internalhash_types.state -> 'a -> Base_internalhash_types.state) -> Base_internalhash_types.state -> 'a ref -> Base_internalhash_types.state @@ portable

Hash support for array and ref is provided, but is potentially DANGEROUS, since it incorporates the current contents of the array/ref into the hash value. Because of this we add a _frozen suffix to the function name.

Hash support for string is also potentially DANGEROUS, but strings are mutated less often, so we don't append _frozen to it.

Also note that we don't support bytes.

val hash_fold_array_frozen : (Base_internalhash_types.state -> 'a -> Base_internalhash_types.state) -> Base_internalhash_types.state -> 'a array -> Base_internalhash_types.state @@ portable

val hash_nativeint : nativeint -> Base_internalhash_types.hash_value @@ portable

val hash_int64 : int64 -> Base_internalhash_types.hash_value @@ portable

val hash_int32 : int32 -> Base_internalhash_types.hash_value @@ portable

val hash_char : char -> Base_internalhash_types.hash_value @@ portable

val hash_int : int -> Base_internalhash_types.hash_value @@ portable

val hash_bool : bool -> Base_internalhash_types.hash_value @@ portable

val hash_string : string -> Base_internalhash_types.hash_value @@ portable

val hash_float : float -> Base_internalhash_types.hash_value @@ portable

val hash_unit : unit -> Base_internalhash_types.hash_value @@ portable

include module type of struct include Stable_witness.Export end

val stable_witness_array : 'a Stable_witness.t -> 'a array Stable_witness.t

val stable_witness_bool : bool Stable_witness.t

val stable_witness_bytes : bytes Stable_witness.t

val stable_witness_char : char Stable_witness.t

val stable_witness_exn : exn Stable_witness.t

val stable_witness_float : float Stable_witness.t

val stable_witness_iarray : 'a Stable_witness.t -> 'a Basement.Stdlib_iarray_labels.t Stable_witness.t

val stable_witness_int : int Stable_witness.t

val stable_witness_int32 : int32 Stable_witness.t

val stable_witness_int64 : int64 Stable_witness.t

val stable_witness_lazy_t : 'a Stable_witness.t -> 'a lazy_t Stable_witness.t

val stable_witness_list : 'a Stable_witness.t -> 'a list Stable_witness.t

val stable_witness_nativeint : nativeint Stable_witness.t

val stable_witness_option : 'a Stable_witness.t -> 'a option Stable_witness.t

val stable_witness_or_null : 'a Stable_witness.t -> 'a Basement.Or_null_shim.t Stable_witness.t

val stable_witness_ref : 'a Stable_witness.t -> 'a ref Stable_witness.t

val stable_witness_string : string Stable_witness.t

val stable_witness_unit : unit Stable_witness.t

include sig ... end

val sexp_of_array : ('a -> Sexplib0.Sexp.t) -> 'a Base.Array.t -> Sexplib0.Sexp.t

val sexp_of_array__stack : ('a @ local -> Sexplib0.Sexp.t @ local) -> 'a Base.Array.t @ local -> Sexplib0.Sexp.t @ local

val array_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a Base.Array.t

val array_sexp_grammar : 'a. 'a Sexplib0.Sexp_grammar.t -> 'a Base.Array.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_bool : Base.Bool.t -> Sexplib0.Sexp.t

val sexp_of_bool__stack : Base.Bool.t @ local -> Sexplib0.Sexp.t @ local

val bool_of_sexp : Sexplib0.Sexp.t -> Base.Bool.t

val bool_sexp_grammar : Base.Bool.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_char : Base.Char.t -> Sexplib0.Sexp.t

val sexp_of_char__stack : Base.Char.t @ local -> Sexplib0.Sexp.t @ local

val char_of_sexp : Sexplib0.Sexp.t -> Base.Char.t

val char_sexp_grammar : Base.Char.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_exn : Base.Exn.t -> Sexplib0.Sexp.t

val sexp_of_float : Base.Float.t -> Sexplib0.Sexp.t

val sexp_of_float__stack : Base.Float.t @ local -> Sexplib0.Sexp.t @ local

val float_of_sexp : Sexplib0.Sexp.t -> Base.Float.t

val float_sexp_grammar : Base.Float.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_int : Base.Int.t -> Sexplib0.Sexp.t

val sexp_of_int__stack : Base.Int.t @ local -> Sexplib0.Sexp.t @ local

val int_of_sexp : Sexplib0.Sexp.t -> Base.Int.t

val int_sexp_grammar : Base.Int.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_int32 : Base.Int32.t -> Sexplib0.Sexp.t

val sexp_of_int32__stack : Base.Int32.t @ local -> Sexplib0.Sexp.t @ local

val int32_of_sexp : Sexplib0.Sexp.t -> Base.Int32.t

val int32_sexp_grammar : Base.Int32.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_int64 : Base.Int64.t -> Sexplib0.Sexp.t

val sexp_of_int64__stack : Base.Int64.t @ local -> Sexplib0.Sexp.t @ local

val int64_of_sexp : Sexplib0.Sexp.t -> Base.Int64.t

val int64_sexp_grammar : Base.Int64.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_list : ('a -> Sexplib0.Sexp.t) -> 'a Base.List.t -> Sexplib0.Sexp.t

val sexp_of_list__stack : ('a @ local -> Sexplib0.Sexp.t @ local) -> 'a Base.List.t @ local -> Sexplib0.Sexp.t @ local

val list_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a Base.List.t

val list_sexp_grammar : 'a. 'a Sexplib0.Sexp_grammar.t -> 'a Base.List.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_nativeint : Base.Nativeint.t -> Sexplib0.Sexp.t

val sexp_of_nativeint__stack : Base.Nativeint.t @ local -> Sexplib0.Sexp.t @ local

val nativeint_of_sexp : Sexplib0.Sexp.t -> Base.Nativeint.t

val nativeint_sexp_grammar : Base.Nativeint.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_option : ('a -> Sexplib0.Sexp.t) -> 'a Base.Option.t -> Sexplib0.Sexp.t

val sexp_of_option__stack : ('a @ local -> Sexplib0.Sexp.t @ local) -> 'a Base.Option.t @ local -> Sexplib0.Sexp.t @ local

val option_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a Base.Option.t

val option_sexp_grammar : 'a. 'a Sexplib0.Sexp_grammar.t -> 'a Base.Option.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_ref : ('a -> Sexplib0.Sexp.t) -> 'a Base.Ref.t -> Sexplib0.Sexp.t

val sexp_of_ref__stack : ('a @ local -> Sexplib0.Sexp.t @ local) -> 'a Base.Ref.t @ local -> Sexplib0.Sexp.t @ local

val ref_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a Base.Ref.t

val ref_sexp_grammar : 'a. 'a Sexplib0.Sexp_grammar.t -> 'a Base.Ref.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_string : Base.String.t -> Sexplib0.Sexp.t

val sexp_of_string__stack : Base.String.t @ local -> Sexplib0.Sexp.t @ local

val string_of_sexp : Sexplib0.Sexp.t -> Base.String.t

val string_sexp_grammar : Base.String.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_bytes : Base.Bytes.t -> Sexplib0.Sexp.t

val sexp_of_bytes__stack : Base.Bytes.t @ local -> Sexplib0.Sexp.t @ local

val bytes_of_sexp : Sexplib0.Sexp.t -> Base.Bytes.t

val bytes_sexp_grammar : Base.Bytes.t Sexplib0.Sexp_grammar.t @@ portable

val sexp_of_unit : Base.Unit.t -> Sexplib0.Sexp.t

val sexp_of_unit__stack : Base.Unit.t @ local -> Sexplib0.Sexp.t @ local

val unit_of_sexp : Sexplib0.Sexp.t -> Base.Unit.t

val unit_sexp_grammar : Base.Unit.t Sexplib0.Sexp_grammar.t @@ portable

include module type of struct include Perms.Export end

type read = Perms.Read.t

val bin_shape_read : Bin_prot.Shape.t

val bin_size_read : read Bin_prot.Size.sizer

val bin_size_read__local : read Bin_prot.Size.sizer__local

val bin_write_read : read Bin_prot.Write.writer

val bin_write_read__local : read Bin_prot.Write.writer__local

val bin_writer_read : read Bin_prot.Type_class.writer

val bin_read_read : read Bin_prot.Read.reader

val __bin_read_read__ : read Bin_prot.Read.vtag_reader

val bin_reader_read : read Bin_prot.Type_class.reader

val bin_read : read Bin_prot.Type_class.t

val compare_read : read -> read -> Base.Int.t

val compare_read__local : read @ local -> (read @ local -> Base.Int.t) @ local

val equal_read : read -> read -> Base.Bool.t

val equal_read__local : read @ local -> (read @ local -> Base.Bool.t) @ local

val globalize_read : read @ local -> read

val hash_fold_read : Ppx_hash_lib.Std.Hash.state -> read -> Ppx_hash_lib.Std.Hash.state

val hash_read : read -> Ppx_hash_lib.Std.Hash.hash_value

val sexp_of_read : read -> Sexplib0.Sexp.t

val read_of_sexp : Sexplib0.Sexp.t -> read

val read_sexp_grammar : read Sexplib0.Sexp_grammar.t @@ portable

val stable_witness_read : read Ppx_stable_witness_runtime.Stable_witness.t

type write = Perms.Write.t

We don't expose bin_io for write due to a naming conflict with the functions exported by bin_io for read_write. If you want bin_io for write, use Write.t.

val compare_write : write -> write -> Base.Int.t

val compare_write__local : write @ local -> (write @ local -> Base.Int.t) @ local

val equal_write : write -> write -> Base.Bool.t

val equal_write__local : write @ local -> (write @ local -> Base.Bool.t) @ local

val hash_fold_write : Ppx_hash_lib.Std.Hash.state -> write -> Ppx_hash_lib.Std.Hash.state

val hash_write : write -> Ppx_hash_lib.Std.Hash.hash_value

val globalize_write : write @ local -> write

val sexp_of_write : write -> Sexplib0.Sexp.t

val write_of_sexp : Sexplib0.Sexp.t -> write

val write_sexp_grammar : write Sexplib0.Sexp_grammar.t @@ portable

val stable_witness_write : write Ppx_stable_witness_runtime.Stable_witness.t

type immutable = Perms.Immutable.t

val bin_shape_immutable : Bin_prot.Shape.t

val bin_size_immutable : immutable Bin_prot.Size.sizer

val bin_size_immutable__local : immutable Bin_prot.Size.sizer__local

val bin_write_immutable : immutable Bin_prot.Write.writer

val bin_write_immutable__local : immutable Bin_prot.Write.writer__local

val bin_writer_immutable : immutable Bin_prot.Type_class.writer

val bin_read_immutable : immutable Bin_prot.Read.reader

val __bin_read_immutable__ : immutable Bin_prot.Read.vtag_reader

val bin_reader_immutable : immutable Bin_prot.Type_class.reader

val bin_immutable : immutable Bin_prot.Type_class.t

val compare_immutable : immutable -> immutable -> Base.Int.t

val compare_immutable__local : immutable @ local -> (immutable @ local -> Base.Int.t) @ local

val equal_immutable : immutable -> immutable -> Base.Bool.t

val equal_immutable__local : immutable @ local -> (immutable @ local -> Base.Bool.t) @ local

val globalize_immutable : immutable @ local -> immutable

val hash_fold_immutable : Ppx_hash_lib.Std.Hash.state -> immutable -> Ppx_hash_lib.Std.Hash.state

val hash_immutable : immutable -> Ppx_hash_lib.Std.Hash.hash_value

val sexp_of_immutable : immutable -> Sexplib0.Sexp.t

val immutable_of_sexp : Sexplib0.Sexp.t -> immutable

val immutable_sexp_grammar : immutable Sexplib0.Sexp_grammar.t @@ portable

val stable_witness_immutable : immutable Ppx_stable_witness_runtime.Stable_witness.t

type read_write = Perms.Read_write.t

val bin_shape_read_write : Bin_prot.Shape.t

val bin_size_read_write : read_write Bin_prot.Size.sizer

val bin_size_read_write__local : read_write Bin_prot.Size.sizer__local

val bin_write_read_write : read_write Bin_prot.Write.writer

val bin_write_read_write__local : read_write Bin_prot.Write.writer__local

val bin_writer_read_write : read_write Bin_prot.Type_class.writer

val bin_read_read_write : read_write Bin_prot.Read.reader

val __bin_read_read_write__ : read_write Bin_prot.Read.vtag_reader

val bin_reader_read_write : read_write Bin_prot.Type_class.reader

val bin_read_write : read_write Bin_prot.Type_class.t

val compare_read_write : read_write -> read_write -> Base.Int.t

val compare_read_write__local : read_write @ local -> (read_write @ local -> Base.Int.t) @ local

val equal_read_write : read_write -> read_write -> Base.Bool.t

val equal_read_write__local : read_write @ local -> (read_write @ local -> Base.Bool.t) @ local

val globalize_read_write : read_write @ local -> read_write

val hash_fold_read_write : Ppx_hash_lib.Std.Hash.state -> read_write -> Ppx_hash_lib.Std.Hash.state

val hash_read_write : read_write -> Ppx_hash_lib.Std.Hash.hash_value

val sexp_of_read_write : read_write -> Sexplib0.Sexp.t

val read_write_of_sexp : Sexplib0.Sexp.t -> read_write

val read_write_sexp_grammar : read_write Sexplib0.Sexp_grammar.t @@ portable

val stable_witness_read_write : read_write Ppx_stable_witness_runtime.Stable_witness.t

type 'a perms = 'a Perms.Upper_bound.t

val bin_shape_perms : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_perms : 'a. 'a Bin_prot.Size.sizer -> 'a perms Bin_prot.Size.sizer

val bin_size_perms__local : 'a. 'a Bin_prot.Size.sizer__local -> 'a perms Bin_prot.Size.sizer__local

val bin_write_perms : 'a. 'a Bin_prot.Write.writer -> 'a perms Bin_prot.Write.writer

val bin_write_perms__local : 'a. 'a Bin_prot.Write.writer__local -> 'a perms Bin_prot.Write.writer__local

val bin_writer_perms : 'a. 'a Bin_prot.Type_class.writer -> 'a perms Bin_prot.Type_class.writer

val bin_read_perms : 'a. 'a Bin_prot.Read.reader -> 'a perms Bin_prot.Read.reader

val __bin_read_perms__ : 'a. 'a Bin_prot.Read.reader -> 'a perms Bin_prot.Read.vtag_reader

val bin_reader_perms : 'a. 'a Bin_prot.Type_class.reader -> 'a perms Bin_prot.Type_class.reader

val bin_perms : 'a. 'a Bin_prot.Type_class.t -> 'a perms Bin_prot.Type_class.t

val compare_perms : 'a. ('a -> 'a -> Base.Int.t) -> 'a perms -> 'a perms -> Base.Int.t

val compare_perms__local : 'a. ('a @ local -> ('a @ local -> Base.Int.t) @ local) -> 'a perms @ local -> ('a perms @ local -> Base.Int.t) @ local

val equal_perms : 'a. ('a -> 'a -> Base.Bool.t) -> 'a perms -> 'a perms -> Base.Bool.t

val equal_perms__local : 'a. ('a @ local -> ('a @ local -> Base.Bool.t) @ local) -> 'a perms @ local -> ('a perms @ local -> Base.Bool.t) @ local

val globalize_perms : 'a. ('a @ local -> 'a) -> 'a perms @ local -> 'a perms

val hash_fold_perms : 'a. (Ppx_hash_lib.Std.Hash.state -> 'a -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 'a perms -> Ppx_hash_lib.Std.Hash.state

val sexp_of_perms : ('a -> Sexplib0.Sexp.t) -> 'a perms -> Sexplib0.Sexp.t

val perms_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a perms

val perms_sexp_grammar : 'a. 'a Sexplib0.Sexp_grammar.t -> 'a perms Sexplib0.Sexp_grammar.t @@ portable

val stable_witness_perms : 'a Ppx_stable_witness_runtime.Stable_witness.t -> 'a perms Ppx_stable_witness_runtime.Stable_witness.t

include module type of struct include Ppx_compare_lib.Builtin end

val compare_bool : bool Ppx_compare_lib.compare

val compare_char : char Ppx_compare_lib.compare

val compare_float : float Ppx_compare_lib.compare

val compare_int : int Ppx_compare_lib.compare

val compare_int32 : int32 Ppx_compare_lib.compare

val compare_int64 : int64 Ppx_compare_lib.compare

val compare_nativeint : nativeint Ppx_compare_lib.compare

val compare_string : string Ppx_compare_lib.compare

val compare_bytes : bytes Ppx_compare_lib.compare

val compare_unit : unit Ppx_compare_lib.compare

val compare_array : 'a. 'a Ppx_compare_lib.compare -> 'a array Ppx_compare_lib.compare

val compare_list : 'a. 'a Ppx_compare_lib.compare -> 'a list Ppx_compare_lib.compare

val compare_option : 'a. 'a Ppx_compare_lib.compare -> 'a option Ppx_compare_lib.compare

val compare_ref : 'a. 'a Ppx_compare_lib.compare -> 'a ref Ppx_compare_lib.compare

val equal_bool : bool Ppx_compare_lib.equal

val equal_char : char Ppx_compare_lib.equal

val equal_float : float Ppx_compare_lib.equal

val equal_int : int Ppx_compare_lib.equal

val equal_int32 : int32 Ppx_compare_lib.equal

val equal_int64 : int64 Ppx_compare_lib.equal

val equal_nativeint : nativeint Ppx_compare_lib.equal

val equal_string : string Ppx_compare_lib.equal

val equal_bytes : bytes Ppx_compare_lib.equal

val equal_unit : unit Ppx_compare_lib.equal

val equal_array : 'a. 'a Ppx_compare_lib.equal -> 'a array Ppx_compare_lib.equal

val equal_list : 'a. 'a Ppx_compare_lib.equal -> 'a list Ppx_compare_lib.equal

val equal_option : 'a. 'a Ppx_compare_lib.equal -> 'a option Ppx_compare_lib.equal

val equal_ref : 'a. 'a Ppx_compare_lib.equal -> 'a ref Ppx_compare_lib.equal

val globalize_bool : bool @ local -> bool @@ portable

val globalize_char : char @ local -> char @@ portable

val globalize_float : float @ local -> float @@ portable

val globalize_int : int @ local -> int @@ portable

val globalize_int32 : int32 @ local -> int32 @@ portable

val globalize_int64 : int64 @ local -> int64 @@ portable

val globalize_nativeint : nativeint @ local -> nativeint @@ portable

val globalize_bytes : bytes @ local -> bytes @@ portable

val globalize_string : string @ local -> string @@ portable

val globalize_unit : unit @ local -> unit @@ portable

val globalize_array : 'a 'b. ('a @ local -> 'b) -> 'a array @ local -> 'a array @@ portable

val globalize_floatarray : floatarray @ local -> floatarray @@ portable

val globalize_lazy_t : ('a @ local -> 'b) -> 'a lazy_t @ local -> 'a lazy_t @@ portable

val globalize_list : 'a 'b. ('a @ local -> 'b) -> 'a list @ local -> 'b list @@ portable

val globalize_option : 'a 'b. ('a @ local -> 'b) -> 'a option @ local -> 'b option @@ portable

val globalize_or_null : ('a @ local -> 'b) -> 'a Basement.Or_null_shim.t @ local -> 'b Basement.Or_null_shim.t @@ portable

val globalize_result : 'ok 'err. ('ok @ local -> 'ok) -> ('err @ local -> 'err) -> ('ok, 'err) result @ local -> ('ok, 'err) result @@ portable

val globalize_ref : 'a 'b. ('a @ local -> 'b) -> 'a ref @ local -> 'a ref @@ portable

exception Not_found

deprecated [since 2018-02] Instead of raising [Not_found], consider using [raise_s] with an informative error message. If code needs to distinguish [Not_found] from other exceptions, please change it to handle both [Not_found] and [Not_found_s]. Then, instead of raising [Not_found], raise [Not_found_s] with an informative error message.

exception Not_found_s of Sexplib0.Sexp.t