Module Date.Days

Days provides a linear representation of dates that is optimized for arithmetic on the number of days between dates, rather than for representing year/month/day components. This module is intended for use only in performance-sensitive contexts where dates are manipulated more often than they are constructed or deconstructed; most clients should use the ordinary t.

type t

val of_date : t -> t

val to_date : t -> t

val diff : t -> t -> Base.Int.t

val add_days : t -> Base.Int.t -> t

val unix_epoch : t

The starting date of the UNIX epoch: 1970-01-01