Module Core.Fdeque

A simple polymorphic functional double-ended queue. Use this if you need a queue-like data structure that provides enqueue and dequeue accessors on both ends. For strictly first-in, first-out access, see Fqueue.

Amortized running times assume that enqueue/dequeue are used sequentially, threading the changing deque through the calls.

type 'a t

include Bin_prot.Binable.S1 with type 'a t := 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ( 'a, 'a t ) Bin_prot.Size.sizer1

val bin_write_t : ( 'a, 'a t ) Bin_prot.Write.writer1

val bin_read_t : ( 'a, 'a t ) Bin_prot.Read.reader1

val __bin_read_t__ : ( 'a, int -> 'a t ) Bin_prot.Read.reader1

val bin_writer_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.writer

val bin_reader_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.reader

val bin_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.t

include Ppx_compare_lib.Comparable.S1 with type 'a t := 'a t

val compare : ( 'a -> 'a -> int ) -> 'a t -> 'a t -> int

include Ppx_compare_lib.Equal.S1 with type 'a t := 'a t

val equal : ( 'a -> 'a -> bool ) -> 'a t -> 'a t -> bool

include Ppx_hash_lib.Hashable.S1 with type 'a t := 'a t

val hash_fold_t : ( Base.Hash.state -> 'a -> Base.Hash.state ) -> Base.Hash.state -> 'a t -> Base.Hash.state

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : ( Sexplib0__.Sexp.t -> 'a ) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ( 'a -> Sexplib0__.Sexp.t ) -> 'a t -> Sexplib0__.Sexp.t

Container operations traverse deque elements front-to-back, like Front_to_back below. If you need faster traversal and don't care about the order, use Arbitrary_order below.

is_empty and length have worst-case complexity O(1).

include Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:( 'a -> 'a -> bool ) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:( 'a -> unit ) -> unit

val fold : 'a t -> init:'accum -> f:( 'accum -> 'a -> 'accum ) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'accum -> f:( 'accum -> 'a -> ( 'accum, 'e ) Base.Result.t ) -> ( 'accum, 'e ) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'accum -> f:( 'accum -> 'a -> ( 'accum, 'final ) Base.Container.Continue_or_stop.t ) -> finish:( 'accum -> 'final ) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:( 'a -> bool ) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:( 'a -> bool ) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:( 'a -> bool ) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base.Container.Summable with type t = 'sum) -> 'a t -> f:( 'a -> 'sum ) -> 'sum

Returns the sum of f i for all i in the container.

val find : 'a t -> f:( 'a -> bool ) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:( 'a -> 'b option ) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:( 'a -> 'a -> int ) -> 'a option

Returns a minimum (resp maximum) element from the collection using the provided compare function, or None if the collection is empty. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold.

val max_elt : 'a t -> compare:( 'a -> 'a -> int ) -> 'a option

include Base.Invariant.S1 with type 'a t := 'a t

val invariant : ( 'a -> unit ) -> 'a t -> unit

include Base.Monad.S with type 'a t := 'a t

val (>>=) : 'a t -> ( 'a -> 'b t ) -> 'b t

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a t -> ( 'a -> 'b ) -> 'b t

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a t -> f:( 'a -> 'b t ) -> 'b t

bind t ~f = t >>= f

val return : 'a -> 'a t

return v returns the (trivial) computation that returns v.

val map : 'a t -> f:( 'a -> 'b ) -> 'b t

map t ~f is t >>| f.

val join : 'a t t -> 'a t

join t is t >>= (fun t' -> t').

val ignore_m : 'a t -> unit t

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end

These are convenient to have in scope when programming with a monad:

module Arbitrary_order : sig ... end

Traverse deque elements in arbitrary order.

module Front_to_back : sig ... end

Traverse deque elements front-to-back. Incurs up to O(n) additional time and space cost over Arbitrary_order.

module Back_to_front : sig ... end

Traverse deque elements back-to-front. Incurs up to O(n) additional time and space cost over Arbitrary_order.

val empty : _ t

The empty deque.

val singleton : 'a -> 'a t

A one-element deque.

val of_list : 'a Base.List.t -> 'a t

of_list returns a deque with elements in the same front-to-back order as the list.

val rev : 'a t -> 'a t

rev t returns t, reversed.

Complexity: worst-case O(1).

val enqueue : 'a t -> [ `back | `front ] -> 'a -> 'a t

enqueue t side x produces t updated with x added to its side.

Complexity: worst-case O(1).

val enqueue_front : 'a t -> 'a -> 'a t

val enqueue_back : 'a t -> 'a -> 'a t

val peek : 'a t -> [ `back | `front ] -> 'a Base.Option.t

peek t side produces Some of the element at the side of t, or None if t is empty.

Complexity: worst-case O(1).

val peek_exn : 'a t -> [ `back | `front ] -> 'a

val peek_front : 'a t -> 'a Base.Option.t

val peek_front_exn : 'a t -> 'a

val peek_back : 'a t -> 'a Base.Option.t

val peek_back_exn : 'a t -> 'a

val drop : 'a t -> [ `back | `front ] -> 'a t Base.Option.t

drop t side produces Some of t with the element at its side removed, or None if t is empty.

Complexity: amortized O(1), worst-case O(length t).

val drop_exn : 'a t -> [ `back | `front ] -> 'a t

val drop_front : 'a t -> 'a t Base.Option.t

val drop_front_exn : 'a t -> 'a t

val drop_back : 'a t -> 'a t Base.Option.t

val drop_back_exn : 'a t -> 'a t

val dequeue : 'a t -> [ `back | `front ] -> ('a * 'a t) Base.Option.t

dequeue t side produces Option.both (peek t side) (drop t side).

Complexity: amortized O(1), worst-case O(length t).

val dequeue_exn : 'a t -> [ `back | `front ] -> 'a * 'a t

val dequeue_front : 'a t -> ('a * 'a t) Base.Option.t

val dequeue_front_exn : 'a t -> 'a * 'a t

val dequeue_back : 'a t -> ('a * 'a t) Base.Option.t

val dequeue_back_exn : 'a t -> 'a * 'a t

module Stable : sig ... end

module Private : sig ... end