Module type Immediate_option_intf.S_without_immediate

type t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

include S_without_immediate_plain with type t := t

type value

The immediate value carried by the immediate option.

Given the presence of unchecked_value, the value type should not have operations that depend on the value's validity for memory safety. In particular, unchecked_value is not called unsafe_value as it would be if it could return a value that later resulted in a segmentation fault. For pointer-like values, use Ext.Nullable, for example.

Constructors analogous to None and Some. If not (some_is_representable x) then some x may raise or return none.

val none : t

val some : value -> t

val some_is_representable : value -> Base.Bool.t

For some representations of immediate options, the encodings of none and some overlap. For these representations, some_is_representable value = false if value cannot be represented as an option. For example, Int.Option uses min_value to represent none. For other representations, some_is_representable always returns true.

val is_none : t -> Base.Bool.t

val is_some : t -> Base.Bool.t

val value : t -> default:value -> value

value (some x) ~default = x and value none ~default = default.

val value_exn : t -> value

value_exn (some x) = x. value_exn none raises. Unlike Option.value_exn, there is no ?message argument, so that calls to value_exn that do not raise also do not have to allocate.

val unchecked_value : t -> value

unchecked_value (some x) = x. unchecked_value none returns an unspecified value. unchecked_value t is intended as an optimization of value_exn t when is_some t is known to be true.

val to_option : t -> value Base.Option.t

val of_option : value Base.Option.t -> t

module Optional_syntax : Optional_syntax.S with type t := t with type value := value