Module Core.Indexed_container

include module type of struct include Base.Indexed_container end

module type Derived = Base.Indexed_container.Derived

module type Generic = Base.Indexed_container.Generic

module type Generic_with_creators = Base.Indexed_container.Generic_with_creators

module type S0 = Base.Indexed_container.S0

module type S0_with_creators = Base.Indexed_container.S0_with_creators

module type S1 = Base.Indexed_container.S1

module type S1_with_creators = Base.Indexed_container.S1_with_creators

include Derived

val foldi : fold:('t -> init:'acc -> f:('acc -> 'a -> 'acc) -> 'acc) -> 't -> init:'acc -> f:(int -> 'acc -> 'a -> 'acc) -> 'acc

val iteri : fold:('t -> init:int -> f:(int -> 'a -> int) -> int) -> 't -> f:(int -> 'a -> unit) -> unit

Generic definitions of indexed container operations in terms of foldi.

val counti : foldi:('t -> init:int -> f:(int -> int -> 'a -> int) -> int) -> 't -> f:(int -> 'a -> bool) -> int

Generic definitions of indexed container operations in terms of iteri.

val existsi : iteri:('t -> f:(int -> 'a -> unit) -> unit) -> 't -> f:(int -> 'a -> bool) -> bool

val for_alli : iteri:('t -> f:(int -> 'a -> unit) -> unit) -> 't -> f:(int -> 'a -> bool) -> bool

val findi : iteri:('t -> f:(int -> 'a -> unit) -> unit) -> 't -> f:(int -> 'a -> bool) -> (int * 'a) option

val find_mapi : iteri:('t -> f:(int -> 'a -> unit) -> unit) -> 't -> f:(int -> 'a -> 'b option) -> 'b option

module Make = Base.Indexed_container.Make

module Make0 = Base.Indexed_container.Make0

module Make_gen = Base.Indexed_container.Make_gen

module Make_with_creators = Base.Indexed_container.Make_with_creators

module Make0_with_creators = Base.Indexed_container.Make0_with_creators

module Make_gen_with_creators = Base.Indexed_container.Make_gen_with_creators

module type S1_permissions = sig ... end

module type S1_with_creators_permissions = sig ... end