Module Core.Int64

This module extends Base.Int64.

Interface from Base

type t = int64

include Sexplib0.Sexpable.S with type t := t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

include Base.Floatable.S with type t := t

val of_float : float -> t

val to_float : t -> float

include Base.Intable.S with type t := t

val of_int_exn : int -> t

val to_int_exn : t -> int

include Base.Identifiable.S with type t := t

include Sexplib0.Sexpable.S with type t := t

include Base.Stringable.S with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

include Base.Comparator.S with type t := t

type comparator_witness

include Base.Pretty_printer.S with type t := t

include Base.Comparable.With_zero with type t := t

val is_positive : t -> bool

val is_non_negative : t -> bool

val is_negative : t -> bool

val is_non_positive : t -> bool

val sign : t -> Base.Sign.t

Returns Neg, Zero, or Pos in a way consistent with the above functions.

include Base.Invariant.S with type t := t

val invariant : t -> unit

val to_string_hum : ?delimiter:char -> t -> string

delimiter is an underscore by default.

Infix operators and constants

val one : t

val minus_one : t

val (+) : t -> t -> t

val (-) : t -> t -> t

val (*) : t -> t -> t

val (**) : t -> t -> t

Integer exponentiation

Negation

val (~-) : t -> t

There are two pairs of integer division and remainder functions, /% and %, and / and rem. They both satisfy the same equation relating the quotient and the remainder:

x = (x /% y) * y + (x % y);
x = (x /  y) * y + (rem x y);

The functions return the same values if x and y are positive. They all raise if y = 0.

The functions differ if x < 0 or y < 0.

If y < 0, then % and /% raise, whereas / and rem do not.

x % y always returns a value between 0 and y - 1, even when x < 0. On the other hand, rem x y returns a negative value if and only if x < 0; that value satisfies abs (rem x y) <= abs y - 1.

val (/%) : t -> t -> t

val (%) : t -> t -> t

val (/) : t -> t -> t

val rem : t -> t -> t

val (//) : t -> t -> float

Float division of integers.

val (land) : t -> t -> t

Same as bit_and.

val (lor) : t -> t -> t

Same as bit_or.

val (lxor) : t -> t -> t

Same as bit_xor.

val (lsl) : t -> int -> t

Same as shift_left.

val (asr) : t -> int -> t

Same as shift_right.

Other common functions

round rounds an int to a multiple of a given to_multiple_of argument, according to a direction dir, with default dir being `Nearest. round will raise if to_multiple_of <= 0. If the result overflows (too far positive or too far negative), round returns an incorrect result.

       | `Down    | rounds toward Int.neg_infinity                          |
       | `Up      | rounds toward Int.infinity                              |
       | `Nearest | rounds to the nearest multiple, or `Up in case of a tie |
       | `Zero    | rounds toward zero                                      |

Here are some examples for round ~to_multiple_of:10 for each direction:

       | `Down    | {10 .. 19} --> 10 | { 0 ... 9} --> 0 | {-10 ... -1} --> -10 |
       | `Up      | { 1 .. 10} --> 10 | {-9 ... 0} --> 0 | {-19 .. -10} --> -10 |
       | `Zero    | {10 .. 19} --> 10 | {-9 ... 9} --> 0 | {-19 .. -10} --> -10 |
       | `Nearest | { 5 .. 14} --> 10 | {-5 ... 4} --> 0 | {-15 ... -6} --> -10 |

For convenience and performance, there are variants of round with dir hard-coded. If you are writing performance-critical code you should use these.

val round : ?dir:[ `Zero | `Nearest | `Up | `Down ] -> t -> to_multiple_of:t -> t

val round_towards_zero : t -> to_multiple_of:t -> t

val round_down : t -> to_multiple_of:t -> t

val round_up : t -> to_multiple_of:t -> t

val round_nearest : t -> to_multiple_of:t -> t

Successor and predecessor functions

val succ : t -> t

val pred : t -> t

Exponentiation

val pow : t -> t -> t

pow base exponent returns base raised to the power of exponent. It is OK if base <= 0. pow raises if exponent < 0, or an integer overflow would occur.

Bit-wise logical operations

val bit_and : t -> t -> t

These are identical to land, lor, etc. except they're not infix and have different names.

val bit_or : t -> t -> t

val bit_xor : t -> t -> t

val bit_not : t -> t

val popcount : t -> int

Returns the number of 1 bits in the binary representation of the input.

Bit-shifting operations

The results are unspecified for negative shifts and shifts >= num_bits.

val shift_left : t -> int -> t

Shifts left, filling in with zeroes.

val shift_right : t -> int -> t

Shifts right, preserving the sign of the input.

Increment and decrement functions for integer references

val decr : t Caml.ref -> unit

val incr : t Caml.ref -> unit

Conversion functions to related integer types

val of_int32_exn : int32 -> t

val to_int32_exn : t -> int32

val of_int64_exn : int64 -> t

val to_int64 : t -> int64

val of_nativeint_exn : nativeint -> t

val to_nativeint_exn : t -> nativeint

val of_float_unchecked : float -> t

of_float_unchecked truncates the given floating point number to an integer, rounding towards zero. The result is unspecified if the argument is nan or falls outside the range of representable integers.

val num_bits : int

The number of bits available in this integer type. Note that the integer representations are signed.

val max_value : t

The largest representable integer.

val min_value : t

The smallest representable integer.

val (lsr) : t -> int -> t

Same as shift_right_logical.

val shift_right_logical : t -> int -> t

Shifts right, filling in with zeroes, which will not preserve the sign of the input.

val ceil_pow2 : t -> t

ceil_pow2 x returns the smallest power of 2 that is greater than or equal to x. The implementation may only be called for x > 0. Example: ceil_pow2 17 = 32

val floor_pow2 : t -> t

floor_pow2 x returns the largest power of 2 that is less than or equal to x. The implementation may only be called for x > 0. Example: floor_pow2 17 = 16

val ceil_log2 : t -> int

ceil_log2 x returns the ceiling of log-base-2 of x, and raises if x <= 0.

val floor_log2 : t -> int

floor_log2 x returns the floor of log-base-2 of x, and raises if x <= 0.

val is_pow2 : t -> bool

is_pow2 x returns true iff x is a power of 2. is_pow2 raises if x <= 0.

val clz : t -> int

Returns the number of leading zeros in the binary representation of the input, as an integer between 0 and one less than num_bits.

The results are unspecified for t = 0.

val ctz : t -> int

Returns the number of trailing zeros in the binary representation of the input, as an integer between 0 and one less than num_bits.

The results are unspecified for t = 0.

module O : sig ... end

include module type of O

val (+) : t -> t -> t

val (-) : t -> t -> t

val (*) : t -> t -> t

val (/) : t -> t -> t

val (~-) : t -> t

val (**) : t -> t -> t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (>) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (land) : t -> t -> t

val (lor) : t -> t -> t

val (lxor) : t -> t -> t

val lnot : t -> t

val abs : t -> t

val neg : t -> t

val zero : t

val (%) : t -> t -> t

val (/%) : t -> t -> t

val (//) : t -> t -> float

val (lsl) : t -> int -> t

val (asr) : t -> int -> t

val (lsr) : t -> int -> t

Conversion functions

val of_int : int -> t

val of_int32 : int32 -> t

val of_int64 : t -> t

val to_int : t -> int option

val to_int32 : t -> int32 option

val of_nativeint : nativeint -> t

val to_nativeint : t -> nativeint option

Truncating conversions

These functions return the least-significant bits of the input. In cases where optional conversions return Some x, truncating conversions return x.

val to_int_trunc : t -> int

val to_int32_trunc : int64 -> int32

val to_nativeint_trunc : int64 -> nativeint

Low-level float conversions

val bits_of_float : float -> t

val float_of_bits : t -> float

Byte swap operations

See Int's byte swap section for a description of Base's approach to exposing byte swap primitives.

As of writing, these operations do not sign extend unnecessarily on 64 bit machines, unlike their int32 counterparts, and hence, are more performant. See the Int32 module for more details of the overhead entailed by the int32 byteswap functions.

val bswap16 : t -> t

val bswap32 : t -> t

val bswap48 : t -> t

val bswap64 : t -> t

Extensions

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

include Int_intf.Hexable with type t := t

module Hex : sig ... end

include Base.Int.Hexable with type t := t and module Hex := Hex

include Identifiable.S with type t := t with type comparator_witness := comparator_witness

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Ppx_compare_lib.Comparable.S with type t := t

include Ppx_hash_lib.Hashable.S with type t := t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Base.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Base.Pretty_printer.S with type t := t

val pp : Base.Formatter.t -> t -> unit

include Comparable.S_binable with type t := t with type comparator_witness := comparator_witness

include Base.Comparable.S with type t := t with type comparator_witness := comparator_witness

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t with type comparator_witness := comparator_witness

module Replace_polymorphic_compare : Base.Comparable.Polymorphic_compare with type t := t

include Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : ( t, comparator_witness ) Comparator.comparator

module Map : Map.S_binable with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Set.S_binable with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Hashable.S_binable with type t := t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val hashable : t Hashtbl.Hashable.t

module Table : Hashtbl.S_binable with type key = t

module Hash_set : Hash_set.S_binable with type elt = t

module Hash_queue : Hash_queue.S with type key = t

include Comparable.Validate_with_zero with type t := t

val validate_lbound : min:t Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Maybe_bound.t -> t Validate.check

val validate_bound : min:t Maybe_bound.t -> max:t Maybe_bound.t -> t Validate.check

val validate_positive : t Validate.check

val validate_non_negative : t Validate.check

val validate_negative : t Validate.check

val validate_non_positive : t Validate.check

include Quickcheckable.S_int with type t := t

include Quickcheck_intf.S_range with type t := t

include Quickcheck_intf.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val gen_incl : t -> t -> t Base_quickcheck.Generator.t

gen_incl lower_bound upper_bound produces values between lower_bound and upper_bound, inclusive. It uses an ad hoc distribution that stresses boundary conditions more often than a uniform distribution, while still able to produce any value in the range. Raises if lower_bound > upper_bound.

val gen_uniform_incl : t -> t -> t Base_quickcheck.Generator.t

gen_uniform_incl lower_bound upper_bound produces a generator for values uniformly distributed between lower_bound and upper_bound, inclusive. Raises if lower_bound > upper_bound.

val gen_log_uniform_incl : t -> t -> t Base_quickcheck.Generator.t

gen_log_uniform_incl lower_bound upper_bound produces a generator for values between lower_bound and upper_bound, inclusive, where the number of bits used to represent the value is uniformly distributed. Raises if (lower_bound < 0) || (lower_bound > upper_bound).

val gen_log_incl : t -> t -> t Base_quickcheck.Generator.t

gen_log_incl lower_bound upper_bound is like gen_log_uniform_incl, but weighted slightly more in favor of generating lower_bound and upper_bound specifically.