Module Interval.Stable

Stable is used to build stable protocols. It ensures backwards compatibility by checking the sexp and bin-io representations of a given module. Here it's also applied to the Float, Int, Time, Time_ns, and Ofday intervals.

module V1 : sig ... end