Module Make.Fill

Fill.bin_prot X.bin_write_t t x writes x to t in bin-prot form, advancing past the bytes written.

include Iobuf_intf.Accessors_write with type ('a, 'd, 'w) t = ([ Core_kernel__Perms.Read.t | Core_kernel__Perms.Write.t ], seek) t -> 'a -> unit with type 'a bin_prot := 'a Bin_prot.Type_class.writer

include Iobuf_intf.Accessors_common with type ('a, 'd, 'w) t = ([ Core_kernel__Perms.Read.t | Core_kernel__Perms.Write.t ], seek) t -> 'a -> unit with type 'a bin_prot := 'a Bin_prot.Type_class.writer

type ('a, 'd, 'w) t = ([ Core_kernel__Perms.Read.t | Core_kernel__Perms.Write.t ], seek) t -> 'a -> unit

('d, 'w) Iobuf.t accessor function manipulating 'a, either writing it to the iobuf or reading it from the iobuf.

val char : (char, 'd, 'w) t

val int64_t_be : (Core_kernel.Int64.t, 'd, 'w) t

val int64_t_le : (Core_kernel.Int64.t, 'd, 'w) t

val head_padded_fixed_string : padding:char -> len:int -> (string, 'd, 'w) t

val tail_padded_fixed_string : padding:char -> len:int -> (string, 'd, 'w) t

val string : str_pos:int -> len:int -> (string, 'd, 'w) t

val bytes : str_pos:int -> len:int -> (Core_kernel.Bytes.t, 'd, 'w) t

val bigstring : str_pos:int -> len:int -> (Core_kernel.Bigstring.t, 'd, 'w) t

val stringo : ?⁠str_pos:int -> ?⁠len:int -> (string, 'd, 'w) t

val byteso : ?⁠str_pos:int -> ?⁠len:int -> (Core_kernel.Bytes.t, 'd, 'w) t

val bigstringo : ?⁠str_pos:int -> ?⁠len:int -> (Core_kernel.Bigstring.t, 'd, 'w) t

val bin_prot : 'a Bin_prot.Type_class.writer -> ('a, 'd, 'w) t

val int8_trunc : (int, 'd, 'w) t

val int16_be_trunc : (int, 'd, 'w) t

val int16_le_trunc : (int, 'd, 'w) t

val int32_be_trunc : (int, 'd, 'w) t

val int32_le_trunc : (int, 'd, 'w) t

val int64_be : (int, 'd, 'w) t

val int64_le : (int, 'd, 'w) t

val uint8_trunc : (int, 'd, 'w) t

val uint16_be_trunc : (int, 'd, 'w) t

val uint16_le_trunc : (int, 'd, 'w) t

val uint32_be_trunc : (int, 'd, 'w) t

val uint32_le_trunc : (int, 'd, 'w) t

val uint64_be_trunc : (int, 'd, 'w) t

val uint64_le_trunc : (int, 'd, 'w) t

val decimal : (int, _, _) t

decimal t int is equivalent to Iobuf.Fill.string t (Int.to_string int), but with improved efficiency and no intermediate allocation.

In other words: It fills the decimal representation of int to t. t is advanced by the number of characters written and no terminator is added. If sufficient space is not available, decimal will raise.