Module Core.Iobuf_intf

See Iobuf for documentation.

type no_seek

Like read.

val sexp_of_no_seek : no_seek -> Ppx_sexp_conv_lib.Sexp.t

type seek = private no_seek

Like read_write.

val sexp_of_seek : seek -> Ppx_sexp_conv_lib.Sexp.t

module type Accessors_common = sig ... end

Collections of access functions. These abstract over Iobuf.Consume, Iobuf.Fill, Iobuf.Peek, and Iobuf.Poke.

module type Accessors_read = sig ... end

module type Accessors_write = sig ... end

module type Bound = sig ... end

An iobuf window bound, either upper or lower. You can't see its int value, but you can save and restore it.

type ('src, 'dst) consuming_blit = src:'src -> dst:'dst -> dst_pos:int -> len:int -> unit

type ('src, 'dst) consuming_blito = src:'src -> ?⁠src_len:int -> dst:'dst -> ?⁠dst_pos:int -> unit -> unit

module type Consuming_blit = sig ... end

module type Compound_hexdump = sig ... end

module Unix : sig ... end

This file is a modified version of unixLabels.mli from the OCaml distribution. Many of these functions raise exceptions but do not have a _exn suffixed name.