Module Unix.Process_channels

Similar to UnixLabels.open_process, but the second argument specifies the environment passed to the command. The result is a triple of channels connected to the standard output, standard input, and standard error of the command.

type t = {

stdin : Stdio.Out_channel.t;

stdout : Stdio.In_channel.t;

stderr : Stdio.In_channel.t;

}