Module Unix.Service

type t = {

name : string;

aliases : string array;

port : int;

proto : string;

}

Structure of entries in the services database.

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val getbyname : string -> protocol:string -> t option

Find an entry in services with the given name.

val getbyname_exn : string -> protocol:string -> t

val getbyport : int -> protocol:string -> t option

Find an entry in services with the given service number.

val getbyport_exn : int -> protocol:string -> t