Module type Syscall_result.S

type ok_value

type t = 'a t

val compare : t -> t -> int

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

include Base.Equal.S with type t := t

val equal : t Base.Equal.equal

val create_ok : ok_value -> t

val create_error : Unix.error -> t

val is_ok : t -> bool

val is_error : t -> bool

val to_result : t -> (ok_value, Unix.error) Core_kernel.Result.t

This returns a preallocated object for all errors and at least a few ok_values, so can be used in many contexts where avoiding allocation is important.

val ok_exn : t -> ok_value

val error_exn : t -> Unix.error

val reinterpret_error_exn : t -> 'a t

This is more efficient than calling error_exn and then the create_error of the destination type.

val ok_or_unix_error_exn : t -> syscall_name:string -> ok_value

val ok_or_unix_error_with_args_exn : t -> syscall_name:string -> 'a -> ('a -> Core_kernel.Sexp.t) -> ok_value

module Optional_syntax : Core_kernel.Optional_syntax.S with type t := t and type value := ok_value