Module Linux_ext.Sysinfo

type t = {

uptime : Base.Float.t;
Time since boot

load1 : int;
Load average over the last minute

load5 : int;
Load average over the last 5 minutes

load15 : int;
Load average over the last 15 minutes

total_ram : int;
Total usable main memory

free_ram : int;
Available memory size

shared_ram : int;
Amount of shared memory

buffer_ram : int;
Memory used by buffers

total_swap : int;
Total swap page size

free_swap : int;
Available swap space

procs : int;
Number of current processes

totalhigh : int;
Total high memory size

freehigh : int;
Available high memory size

mem_unit : int;
Memory unit size in bytes

}

Result of sysinfo syscall (man 2 sysinfo).

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val sysinfo : (unit -> t) Core_kernel.Or_error.t