Module Core.List

This module extends Base.List with bin_io and quickcheck.

The interface from Base

module Constructors = Base.List.Constructors

type ('a : value_or_null) t = 'a list

val compare : 'a. (('a : value_or_null) -> ('a : value_or_null) -> int) -> ('a : value_or_null) t -> ('a : value_or_null) t -> int

val compare__local : 'a. (('a : value_or_null) @ local -> (('a : value_or_null) @ local -> int) @ local) -> ('a : value_or_null) t @ local -> (('a : value_or_null) t @ local -> int) @ local

val equal : 'a. (('a : value_or_null) -> ('a : value_or_null) -> bool) -> ('a : value_or_null) t -> ('a : value_or_null) t -> bool

val equal__local : 'a. (('a : value_or_null) @ local -> (('a : value_or_null) @ local -> bool) @ local) -> ('a : value_or_null) t @ local -> (('a : value_or_null) t @ local -> bool) @ local

val globalize : 'a. ('a @ local -> 'a) -> 'a t @ local -> 'a t

val hash_fold_t : 'a. (Ppx_hash_lib.Std.Hash.state -> ('a : value_or_null) -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('a : value_or_null) t -> Ppx_hash_lib.Std.Hash.state

val sexp_of_t : 'a. (('a : value_or_null) -> Sexplib0.Sexp.t) -> ('a : value_or_null) t -> Sexplib0.Sexp.t

val sexp_of_t__stack : 'a. (('a : value_or_null) @ local -> Sexplib0.Sexp.t @ local) -> ('a : value_or_null) t @ local -> Sexplib0.Sexp.t @ local

val t_of_sexp : 'a. (Sexplib0.Sexp.t -> ('a : value_or_null)) -> Sexplib0.Sexp.t -> ('a : value_or_null) t

val t_sexp_grammar : 'a. ('a : value_or_null) Sexplib0.Sexp_grammar.t -> ('a : value_or_null) t Sexplib0.Sexp_grammar.t @@ portable

include Base.Indexed_container.S1_with_creators__''value_or_null''__stack with type ('a : value_or_null) t := 'a t

include Base.Container.S1__''value_or_null''__stack with type ('a : value_or_null) t := 'a t

include sig ... end

include Base.Container.Generic_types__''value_or_null'' with type (('a : value_or_null), _, _) t := 'a t with type ('a : value_or_null) elt := 'a

These are all like their equivalents in Container except that an index starting at 0 is added as the first argument to f.

val existsi : 'a 'p1 'p2. 'a t -> f:(int -> ('a -> bool) @ local) @ local -> bool

val for_alli : 'a 'p1 'p2. 'a t -> f:(int -> ('a -> bool) @ local) @ local -> bool

val counti : 'a 'p1 'p2. 'a t -> f:(int -> ('a -> bool) @ local) @ local -> int

val iteri_until : 'a 'p1 'p2 'final. 'a t -> f: (int -> ('a -> (unit, 'final) Base.Container.Continue_or_stop.t) @ local) @ local -> (finish:(int -> 'final) @ local -> 'final) @ local

val find_mapi : 'a 'p1 'p2 'b. 'a t -> f:(int -> ('a -> 'b option) @ local) @ local -> 'b option

val foldi_until : 'a 'p1 'p2 'acc 'final. 'a t -> init:'acc -> f: (int -> ('acc -> ('a -> ('acc, 'final) Base.Container.Continue_or_stop.t) @ local) @ local) @ local -> (finish:(int -> ('acc -> 'final) @ local) @ local -> 'final) @ local

val findi : 'a 'p1 'p2. 'a t -> f:(int -> ('a -> bool) @ local) @ local -> (int * 'a) option

include Base.Container.S1_with_creators__''value_or_null''__stack with type ('a : value_or_null) t := 'a t

include sig ... end

val is_empty : 'a 'p1 'p2. 'a t @ local -> bool

val exists : 'a 'p1 'p2. 'a t -> f:('a -> bool) @ local -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a 'p1 'p2. 'a t -> f:('a -> bool) @ local -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a 'p1 'p2. 'a t -> f:('a -> bool) @ local -> int

Returns the number of elements for which the provided function evaluates to true.

val find : 'a 'p1 'p2. 'a t -> f:('a -> bool) @ local -> 'a Option.t

Returns as an option the first element for which f evaluates to true.

val to_list : 'a 'p1 'p2. 'a t -> 'a list

val min_elt : 'a 'p1 'p2. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> 'a Option.t

Returns a min (resp. max) element from the collection using the provided compare function. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold. Returns None iff the collection is empty.

val max_elt : 'a 'p1 'p2. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> 'a Option.t

val sum : 'a 'sum 'p1 'p2. (module Base.Container.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) @ local -> 'sum

Returns the sum of f i for all i in the container. The order in which the elements will be summed is unspecified.

val iter_until : 'a 'p1 'p2 'final. 'a t -> f:('a -> (unit, 'final) Base.Container.Continue_or_stop.t) @ local -> (finish:(unit -> 'final) @ local -> 'final) @ local

iter_until t ~f ~finish is a short-circuiting version of iter. If f returns Stop x the computation ceases and returns x. If f always returns Continue () the final result is computed by finish.

val fold_result : 'a 'p1 'p2 'acc 'e. 'a t -> init:'acc -> f:('acc -> ('a -> ('acc, 'e) Result.t) @ local) @ local -> ('acc, 'e) Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val find_map : 'a 'p1 'p2 'b. 'a t -> f:('a -> 'b Option.t) @ local -> 'b Option.t

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val fold_until : 'a 'p1 'p2 'acc 'final. 'a t -> init:'acc -> f: ('acc -> ('a -> ('acc, 'final) Base.Container.Continue_or_stop.t) @ local) @ local -> (finish:('acc -> 'final) @ local -> 'final) @ local

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

  type maybe_negative =
    | Found_negative of int
    | All_nonnegative of { sum : int }

  (** [first_neg_or_sum list] returns the first negative number in [list], if any,
      otherwise returns the sum of the list. *)
  let first_neg_or_sum =
    List.fold_until ~init:0
      ~f:(fun sum x ->
        if x < 0
        then Stop (Found_negative x)
        else Continue (sum + x))
      ~finish:(fun sum -> All_nonnegative { sum })
  ;;

  let x = first_neg_or_sum [1; 2; 3; 4; 5]
  val x : maybe_negative = All_nonnegative {sum = 15}

  let y = first_neg_or_sum [1; 2; -3; 4; 5]
  val y : maybe_negative = Found_negative -3

include sig ... end

val of_list : 'a 'p1 'p2. 'a list -> 'a t

val concat : 'a 'p1 'p2. 'a t t -> 'a t

Concatenates a nested container. The elements of the inner containers are concatenated together in order to give the result.

val partition_tf : 'a 'p1 'p2. 'a t -> f:('a -> bool) @ local -> 'a t * 'a t

partition_tf t ~f returns a pair t1, t2, where t1 is all elements of t that satisfy f, and t2 is all elements of t that do not satisfy f. The "tf" suffix is mnemonic to remind readers that the result is (trues, falses).

val partition_map : 'a 'p1 'p2 'b 'c. 'a t -> f:('a -> ('b, 'c) Either.t) @ local -> 'b t * 'c t

partition_map t ~f partitions t according to f.

val to_array : 'a. 'a t -> 'a array

val of_array : 'a. 'a array -> 'a t

include Base.Container.Generic_types__''value_or_null'' with type (('a : value_or_null), _, _) t := 'a t with type ('a : value_or_null) elt := 'a

val partitioni_tf : 'a 'p1 'p2. 'a t -> f:(int -> ('a -> bool) @ local) @ local -> 'a t * 'a t

partitioni_tf t ~f is like partition_tf. Additionally, it passes the index as an argument.

val partition_mapi : 'a 'p1 'p2 'b 'c. 'a t -> f:(int -> ('a -> ('b, 'c) Base__.Either0.t) @ local) @ local -> 'b t * 'c t

partition_mapi t ~f is like partition_map. Additionally, it passes the index as an argument.

val invariant : ('a -> unit) -> 'a t -> unit

module Cartesian_product = Base.List.Cartesian_product

Implements cartesian-product behavior for map and bind.

The monad portion of Cartesian_product is re-exported at top level.

include Base.Monad.S__'value_or_null_mod_maybe_null'__local with type ('a : value_or_null) t := 'a t

val return : 'a 'i 'p 'q. 'a -> 'a t

Convert a value to a t.

val bind : 'a 'b 'i 'j 'k 'p 'q. 'a t -> f:('a -> 'b t) @ local -> 'b t

Sequences computations. bind t ~f computes f v for value(s) v in t. Well-behaved monads satisfy these "laws" (where ( >>= ) is the infix bind operator):

map t ~f is equivalent to bind t ~f:(fun x -> return (f x))

return x >>= f is equivalent to f x

t >>= return is equivalent to t

(t >>= f) >>= g is equivalent to t >>= fun x -> f x >>= g

val join : 'a 'i 'j 'k 'p 'q. 'a t t -> 'a t

Combines nested t into just one layer. Equivalent to bind t ~f:Fn.id.

val ignore_m : 'a 'i 'j 'p 'q. 'a t -> unit t

Ignores contained values of t. Equivalent to map t ~f:ignore.

val all : 'a 'i 'p 'q. 'a t list -> 'a list t

Combines a list of t.

val all_unit : 'i 'p 'q. unit t list -> unit t

Combines a list of t whose contents are unimportant.

val (>>=) : 'a 'b 'i 'j 'k 'p 'q. 'a t -> ('a -> 'b t) @ local -> 'b t

Infix bind.

val (>>|) : 'a 'b 'i 'j 'p 'q. 'a t -> ('a -> 'b) @ local -> 'b t

Infix map.

module Monad_infix = Base.List.Monad_infix

module Let_syntax = Base.List.Let_syntax

module Or_unequal_lengths = Base.List.Or_unequal_lengths

Or_unequal_lengths is used for functions that take multiple lists and that only make sense if all the lists have the same length, e.g., iter2, map3. Such functions check the list lengths prior to doing anything else, and return Unequal_lengths if not all the lists have the same length.

val create : 'a. len:int -> 'a -> 'a list @@ portable

create ~len x returns a list of length len populated x.

val singleton : 'a. 'a -> 'a t @@ portable

singleton x returns a list with a single element x.

val iter : 'a. 'a t -> f:('a -> unit) @ local -> unit

val iteri : 'a. 'a t -> f:(int -> ('a -> unit) @ local) @ local -> unit

val length : 'a. 'a t @ local immutable -> int

val nth_exn : 'a. 'a t -> int -> 'a

Return the n-th element of the given list. The first element (head of the list) is at position 0. Raise if the list is too short or n is negative.

val nth : 'a. 'a t -> int -> 'a option

val mem : 'a t -> 'a -> equal:('a -> ('a -> bool) @ local) @ local -> bool

val init : 'a. int -> f:(int -> 'a) @ local -> 'a t

init n ~f creates a list of length n where the element at index i is the value of f i. Raises if n < 0. f is called on indices from the highest to lowest, so the order of side effects is reversed: List.init 3 ~f:print_int prints 210.

val append : 'a. 'a t -> 'a t -> 'a t

val filteri : 'a. 'a t -> f:(int -> ('a -> bool) @ local) @ local -> 'a t

val filter : 'a. 'a t -> f:('a -> bool) @ local -> 'a t

val rev_append : 'a. 'a t -> 'a t -> 'a t

rev_append l1 l2 reverses l1 and concatenates it to l2. This is equivalent to (List.rev l1) @ l2, but rev_append is more efficient.

val rev : 'a. 'a t -> 'a t

List reversal.

val find_or_null : 'a t -> f:('a -> bool) @ local -> 'a Basement.Or_null_shim.t

val nth_or_null : 'a t -> int -> 'a Basement.Or_null_shim.t

val map : 'a 'b. 'a t -> f:('a -> 'b) @ local -> 'b t

val mapi : 'a 'b. 'a t -> f:(int -> ('a -> 'b) @ local) @ local -> 'b t

val rev_map : 'a 'b. 'a t -> f:('a -> 'b) @ local -> 'b t

rev_map l ~f gives the same result as List.rev (List.map l ~f), but is more efficient.

val rev_mapi : 'a 'b. 'a t -> f:(int -> ('a -> 'b) @ local) @ local -> 'b t

val filter_map : 'a 'b. 'a t -> f:('a -> 'b option) @ local -> 'b t

val filter_mapi : 'a 'b. 'a t -> f:(int -> ('a -> 'b option) @ local) @ local -> 'b t

val concat_map : 'a 'b. 'a t -> f:('a -> 'b t) @ local -> 'b t

val concat_mapi : 'a 'b. 'a t -> f:(int -> ('a -> 'b t) @ local) @ local -> 'b t

val fold : 'a 'b. 'a t -> init:'b -> f:('b -> ('a -> 'b) @ local) @ local -> 'b

val foldi : 'a 'b. 'a t -> init:'b -> f:(int -> ('b -> ('a -> 'b) @ local) @ local) @ local -> 'b

val unordered_append : 'a. 'a t -> 'a t -> 'a t @@ portable

unordered_append l1 l2 has the same elements as l1 @ l2, but in some unspecified order. Generally takes time proportional to length of first list, but is O(1) if either list is empty.

val iter2_exn : 'a 'b. 'a t -> 'b t -> f:('a -> ('b -> unit) @ local) @ local -> unit @@ portable

iter2 [a1; ...; an] [b1; ...; bn] ~f calls in turn f a1 b1; ...; f an bn. The exn version will raise if the two lists have different lengths.

val iter2 : 'a 'b. 'a t -> 'b t -> f:('a -> ('b -> unit) @ local) @ local -> unit Or_unequal_lengths.t @@ portable

val rev_map2_exn : 'a 'b 'c. 'a t -> 'b t -> f:('a -> ('b -> 'c) @ local) @ local -> 'c t @@ portable

rev_map2_exn l1 l2 ~f gives the same result as List.rev (List.map2_exn l1 l2 ~f), but is more efficient.

val rev_map2 : 'a 'b 'c. 'a t -> 'b t -> f:('a -> ('b -> 'c) @ local) @ local -> 'c t Or_unequal_lengths.t @@ portable

val fold2_exn : 'a 'b 'acc. 'a t -> 'b t -> init:'acc -> f:('acc -> ('a -> ('b -> 'acc) @ local) @ local) @ local -> 'acc @@ portable

fold2 ~f ~init:a [b1; ...; bn] [c1; ...; cn] is f (... (f (f a b1 c1) b2 c2) ...) bn cn. The exn version will raise if the two lists have different lengths.

val fold2 : 'a 'b 'acc. 'a t -> 'b t -> init:'acc -> f:('acc -> ('a -> ('b -> 'acc) @ local) @ local) @ local -> 'acc Or_unequal_lengths.t @@ portable

val fold_right2_exn : 'a 'b 'acc. 'a t -> 'b t -> f:('a -> ('b -> ('acc -> 'acc) @ local) @ local) @ local -> (init:'acc -> 'acc) @ local @@ portable

fold_right2 ~f [a1; ...; an] [b1; ...; bn] ~init:c is f a1 b1 (f a2 b2 (... (f an bn c) ...)). The exn version will raise if the two lists have different lengths.

val fold_right2 : 'a 'b 'acc. 'a t -> 'b t -> f:('a -> ('b -> ('acc -> 'acc) @ local) @ local) @ local -> (init:'acc -> 'acc Or_unequal_lengths.t) @ local @@ portable

val for_all2_exn : 'a 'b. 'a t -> 'b t -> f:('a -> ('b -> bool) @ local) @ local -> bool @@ portable

Like List.for_all, but for a two-argument predicate. The exn version will raise if the two lists have different lengths.

val for_all2 : 'a 'b. 'a t -> 'b t -> f:('a -> ('b -> bool) @ local) @ local -> bool Or_unequal_lengths.t @@ portable

val exists2_exn : 'a 'b. 'a t -> 'b t -> f:('a -> ('b -> bool) @ local) @ local -> bool @@ portable

Like List.exists, but for a two-argument predicate. The exn version will raise if the two lists have different lengths.

val exists2 : 'a 'b. 'a t -> 'b t -> f:('a -> ('b -> bool) @ local) @ local -> bool Or_unequal_lengths.t @@ portable

val rev_filter : 'a. 'a t -> f:('a -> bool) @ local -> 'a t @@ portable

Like filter, but reverses the order of the input list.

val partition3_map : 'a 'b 'c 'd. 'a t -> f:('a -> [ `Fst of 'b | `Snd of 'c | `Trd of 'd ]) @ local -> 'b t * 'c t * 'd t @@ portable

val partition_result : 'ok 'error. ('ok, 'error) Base.Result.t t -> 'ok t * 'error t @@ portable

partition_result l returns a pair of lists (l1, l2), where l1 is the list of all Ok elements in l and l2 is the list of all Error elements. The order of elements in the input list is preserved.

val split_n : 'a. 'a t -> int -> 'a t * 'a t @@ portable

split_n [e1; ...; em] n is ([e1; ...; en], [en+1; ...; em]).

If n >= m, ([e1; ...; em], []) is returned.

If n <= 0, ([], [e1; ...; em]) is returned.

In either of these cases, the input list is returned as one side of the pair, rather than being copied.

val sort : 'a. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> 'a t

Sort a list in increasing order according to a comparison function. The comparison function must return 0 if its arguments compare as equal, a positive integer if the first is greater, and a negative integer if the first is smaller (see Array.sort for a complete specification). For example, Poly.compare is a suitable comparison function.

The current implementation uses Merge Sort. It runs in linear heap space and logarithmic stack space.

Presently, the sort is stable, meaning that two equal elements in the input will be in the same order in the output.

val stable_sort : 'a. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> 'a t

Like sort, but guaranteed to be stable.

val merge : 'a. 'a t -> 'a t -> compare:('a -> ('a -> int) @ local) @ local -> 'a t @@ portable

Merges two lists: assuming that l1 and l2 are sorted according to the comparison function compare, merge compare l1 l2 will return a sorted list containing all the elements of l1 and l2. If several elements compare equal, the elements of l1 will be before the elements of l2.

val hd : 'a. 'a t -> 'a option @@ portable

val tl : 'a. 'a t -> 'a t option @@ portable

val hd_exn : 'a. 'a t -> 'a @@ portable

Returns the first element of the given list. Raises if the list is empty.

val tl_exn : 'a. 'a t -> 'a t @@ portable

Returns the given list without its first element. Raises if the list is empty.

val findi_exn : 'a. 'a t -> f:(int -> ('a -> bool) @ local) @ local -> int * 'a @@ portable

Like find_exn, but passes the index as an argument.

val find_exn : 'a. 'a t -> f:('a -> bool) @ local -> 'a @@ portable

find_exn t ~f returns the first element of t that satisfies f. It raises Stdlib.Not_found or Not_found_s if there is no such element.

val find_map_exn : 'a 'b. 'a t -> f:('a -> 'b option) @ local -> 'b @@ portable

Returns the first evaluation of f that returns Some. Raises Stdlib.Not_found or Not_found_s if f always returns None.

val find_mapi_exn : 'a 'b. 'a t -> f:(int -> ('a -> 'b option) @ local) @ local -> 'b @@ portable

Like find_map_exn, but passes the index as an argument.

folding_map is a version of map that threads an accumulator through calls to f.

val folding_map : 'a 'b 'acc. 'a t -> init:'acc -> f:('acc -> ('a -> 'acc * 'b) @ local) @ local -> 'b t @@ portable

val folding_mapi : 'a 'b 'acc. 'a t -> init:'acc -> f:(int -> ('acc -> ('a -> 'acc * 'b) @ local) @ local) @ local -> 'b t @@ portable

fold_map is a combination of fold and map that threads an accumulator through calls to f.

val fold_map : 'a 'b 'acc. 'a t -> init:'acc -> f:('acc -> ('a -> 'acc * 'b) @ local) @ local -> 'acc * 'b t @@ portable

val fold_mapi : 'a 'b 'acc. 'a t -> init:'acc -> f:(int -> ('acc -> ('a -> 'acc * 'b) @ local) @ local) @ local -> 'acc * 'b t @@ portable

map2 [a1; ...; an] [b1; ...; bn] ~f is [f a1 b1; ...; f an bn]. The exn version will raise if the two lists have different lengths.

val map2_exn : 'a 'b 'c. 'a t -> 'b t -> f:('a -> ('b -> 'c) @ local) @ local -> 'c t @@ portable

val map2 : 'a 'b 'c. 'a t -> 'b t -> f:('a -> ('b -> 'c) @ local) @ local -> 'c t Or_unequal_lengths.t @@ portable

Analogous to rev_map2.

val rev_map3_exn : 'a 'b 'c 'd. 'a t -> 'b t -> 'c t -> f:('a -> ('b -> ('c -> 'd) @ local) @ local) @ local -> 'd t @@ portable

val rev_map3 : 'a 'b 'c 'd. 'a t -> 'b t -> 'c t -> f:('a -> ('b -> ('c -> 'd) @ local) @ local) @ local -> 'd t Or_unequal_lengths.t @@ portable

Analogous to map2.

val map3_exn : 'a 'b 'c 'd. 'a t -> 'b t -> 'c t -> f:('a -> ('b -> ('c -> 'd) @ local) @ local) @ local -> 'd t @@ portable

val map3 : 'a 'b 'c 'd. 'a t -> 'b t -> 'c t -> f:('a -> ('b -> ('c -> 'd) @ local) @ local) @ local -> 'd t Or_unequal_lengths.t @@ portable

val rev_map_append : 'a 'b. 'a t -> 'b t -> f:('a -> 'b) @ local -> 'b t @@ portable

rev_map_append l1 l2 ~f reverses l1 mapping f over each element, and appends the result to the front of l2.

val fold_right : 'a 'acc. 'a t -> f:('a -> ('acc -> 'acc) @ local) @ local -> (init:'acc -> 'acc) @ local @@ portable

fold_right [a1; ...; an] ~f ~init:b is f a1 (f a2 (... (f an b) ...)).

val fold_left : 'a 'acc. 'a t -> init:'acc -> f:('acc -> ('a -> 'acc) @ local) @ local -> 'acc @@ portable

fold_left is the same as Container.S1.fold, and one should always use fold rather than fold_left, except in functors that are parameterized over a more general signature where this equivalence does not hold.

Transform a list of pairs into a pair of lists: unzip [(a1,b1); ...; (an,bn)] is ([a1; ...; an], [b1; ...; bn]).

val unzip : 'a 'b. ('a * 'b) t -> 'a t * 'b t @@ portable

val unzip3 : 'a 'b 'c. ('a * 'b * 'c) t -> 'a t * 'b t * 'c t @@ portable

Transform a pair of lists into an (optional) list of pairs: zip [a1; ...; an] [b1; ...; bn] is [(a1,b1); ...; (an,bn)]. Returns Unequal_lengths if the two lists have different lengths.

val zip : 'a 'b. 'a t -> 'b t -> ('a * 'b) t Or_unequal_lengths.t @@ portable

val zip_exn : 'a 'b. 'a t -> 'b t -> ('a * 'b) t @@ portable

Analogous to zip, but with 3 lists rather than 2.

val zip3 : 'a t -> 'b t -> 'c t -> ('a * 'b * 'c) t Or_unequal_lengths.t @@ portable

val zip3_exn : 'a t -> 'b t -> 'c t -> ('a * 'b * 'c) t @@ portable

val reduce_exn : 'a. 'a t -> f:('a -> ('a -> 'a) @ local) @ local -> 'a @@ portable

reduce_exn [a1; ...; an] ~f is f (... (f (f a1 a2) a3) ...) an. It fails on the empty list. Tail recursive.

val reduce : 'a. 'a t -> f:('a -> ('a -> 'a) @ local) @ local -> 'a option @@ portable

val reduce_balanced : 'a. 'a t -> f:('a -> ('a -> 'a) @ local) @ local -> 'a option @@ portable

reduce_balanced returns the same value as reduce when f is associative, but differs in that the tree of nested applications of f has logarithmic depth.

This is useful when your 'a grows in size as you reduce it and f becomes more expensive with bigger inputs. For example, reduce_balanced ~f:(^) takes n*log(n) time, while reduce ~f:(^) takes quadratic time.

val reduce_balanced_exn : 'a. 'a t -> f:('a -> ('a -> 'a) @ local) @ local -> 'a @@ portable

val group : 'a. 'a t -> break:('a -> ('a -> bool) @ local) @ local -> 'a t t @@ portable

group l ~break returns a list of lists (i.e., groups) whose concatenation is equal to the original list. Each group is broken where break returns true on a pair of successive elements.

Example:

  group ~break:(<>) ['M';'i';'s';'s';'i';'s';'s';'i';'p';'p';'i'] ->

  [['M'];['i'];['s';'s'];['i'];['s';'s'];['i'];['p';'p'];['i']]

val groupi : 'a. 'a t -> break:(int -> ('a -> ('a -> bool) @ local) @ local) @ local -> 'a t t @@ portable

This is just like group, except that you get the index in the original list of the current element along with the two elements.

Example, group the chars of "Mississippi" into triples:

  groupi ~break:(fun i _ _ -> i mod 3 = 0)
    ['M';'i';'s';'s';'i';'s';'s';'i';'p';'p';'i'] ->

  [['M'; 'i'; 's']; ['s'; 'i'; 's']; ['s'; 'i'; 'p']; ['p'; 'i']]

val sort_and_group : 'a. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> 'a t t @@ portable

Group equal elements into the same buckets. Sorting is stable.

val chunks_of : 'a. 'a t -> length:int -> 'a t t @@ portable

chunks_of l ~length returns a list of lists whose concatenation is equal to the original list. Every list has length elements, except for possibly the last list, which may have fewer. chunks_of raises if length <= 0.

val chunk_evenly : 'a t -> into:int -> 'a t t @@ portable

chunk_evenly l ~into returns a list of exactly into lists whose concatenation is equal to the original list. The number of items in each list are split as evenly as possible (i.e. lengths differ by at most 1).

Specifically (with n = length l), the output consists of n % into lists of length (n / into) + 1 followed by into - n % into of length n / into. Raises if into <= 0.

val last : 'a. 'a t -> 'a option @@ portable

The final element of a list. The _exn version raises on the empty list.

val last_exn : 'a. 'a t -> 'a @@ portable

val is_prefix : 'a. 'a t -> prefix:'a t -> equal:('a -> ('a -> bool) @ local) @ local -> bool @@ portable

is_prefix xs ~prefix returns true if xs starts with prefix.

val is_suffix : 'a. 'a t -> suffix:'a t -> equal:('a -> ('a -> bool) @ local) @ local -> bool @@ portable

is_suffix xs ~suffix returns true if xs ends with suffix.

val find_consecutive_duplicate : 'a. 'a t -> equal:('a -> ('a -> bool) @ local) @ local -> ('a * 'a) option @@ portable

find_consecutive_duplicate t ~equal returns the first pair of consecutive elements (a1, a2) in t such that equal a1 a2. They are returned in the same order as they appear in t. equal need not be an equivalence relation; it is simply used as a predicate on consecutive elements.

val remove_consecutive_duplicates : 'a. ?which_to_keep:[ `First | `Last ] -> 'a t -> equal:('a -> ('a -> bool) @ local) @ local -> 'a t @@ portable

Returns the given list with consecutive duplicates removed. The relative order of the other elements is unaffected. The element kept from a run of duplicates is determined by which_to_keep.

val dedup_and_sort : 'a. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> 'a t @@ portable

Returns the given list with duplicates removed and in sorted order. Of duplicates in the original list, the element occurring last in the original list is kept.

val stable_dedup : 'a. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> 'a t @@ portable

Returns the original list, dropping all occurrences of duplicates after the first.

val find_a_dup : 'a. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> 'a option @@ portable

find_a_dup returns a duplicate from the list (with no guarantees about which duplicate you get), or None if there are no dups.

val contains_dup : 'a. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> bool @@ portable

Returns true if there are any two elements in the list which are the same. O(n log n) time complexity.

val find_all_dups : 'a. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> 'a list @@ portable

find_all_dups returns a list of all elements that occur more than once, with no guarantees about order. O(n log n) time complexity.

val all_equal : 'a. 'a t -> equal:('a -> ('a -> bool) @ local) @ local -> 'a option @@ portable

all_equal returns a single element of the list that is equal to all other elements, or None if no such element exists.

val range : ?stride:int -> ?start:[ `inclusive | `exclusive ] -> ?stop:[ `inclusive | `exclusive ] -> int -> int -> int t @@ portable

range ?stride ?start ?stop start_i stop_i is the list of integers from start_i to stop_i, stepping by stride. If stride < 0 then we need start_i > stop_i for the result to be nonempty (or start_i = stop_i in the case where both bounds are inclusive).

val range' : 'a. compare:('a -> ('a -> int) @ local) @ local -> (stride:('a -> 'a) @ local -> (?start:[ `inclusive | `exclusive ] -> (?stop:[ `inclusive | `exclusive ] -> ('a -> ('a -> 'a t) @ local) @ local) @ local) @ local) @ local @@ portable

range' is analogous to range for general start/stop/stride types. range' raises if stride x returns x or if the direction that stride x moves x changes from one call to the next.

val filter_opt : 'a. 'a option t -> 'a t

filter_opt l is the sublist of l containing only elements which are Some e. In other words, filter_opt l = filter_map ~f:Fn.id l.

val rev_filter_map : 'a 'b. 'a t -> f:('a -> 'b option) @ local -> 'b t

rev_filter_map l ~f is the reversed sublist of l containing only elements for which f returns Some e.

val rev_filter_mapi : 'a 'b. 'a t -> f:(int -> 'a -> 'b option) -> 'b t

rev_filter_mapi is just like rev_filter_map, but it also passes in the index of each element as the first argument to the mapped function. Tail-recursive.

val of_iter : 'a. iter:(f:('a -> unit) @ local -> unit) @ local -> 'a t @@ portable

Create a list of every value iter passes to f, in chronological order.

val sub : 'a. 'a t -> pos:int -> len:int -> 'a t @@ portable

sub pos len l is the len-element sublist of l, starting at pos.

val take : 'a. 'a t -> int -> 'a t @@ portable

take l n returns the first n elements of l, or all of l if n > length l. take l n = fst (split_n l n). If n >= length l, returns l rather than a copy.

val drop : 'a. 'a t -> int -> 'a t @@ portable

drop l n returns l without the first n elements, or the empty list if n > length l. drop l n is equivalent to snd (split_n l n). If n <= 0, returns l rather than a copy.

val take_while : 'a. 'a t -> f:('a -> bool) @ local -> 'a t @@ portable

take_while l ~f returns the longest prefix of l for which f is true.

val drop_while : 'a. 'a t -> f:('a -> bool) @ local -> 'a t @@ portable

drop_while l ~f drops the longest prefix of l for which f is true.

val split_while : 'a. 'a t -> f:('a -> bool) @ local -> 'a t * 'a t @@ portable

split_while xs ~f = (take_while xs ~f, drop_while xs ~f).

val drop_last : 'a. 'a t -> 'a t option @@ portable

drop_last l drops the last element of l, returning None if l is empty.

val drop_last_exn : 'a. 'a t -> 'a t @@ portable

val concat_no_order : 'a. 'a t t -> 'a t @@ portable

Like concat, but faster and without preserving any ordering (i.e., for lists that are essentially viewed as multi-sets).

val cons : 'a. 'a -> 'a t -> 'a t @@ portable

val cartesian_product : 'a 'b. 'a t -> 'b t -> ('a * 'b) t @@ portable

Returns a list with all possible pairs -- if the input lists have length len1 and len2, the resulting list will have length len1 * len2.

val permute : 'a. ?random_state:Base.Random.State.t -> 'a t -> 'a t @@ portable

permute ?random_state t returns a permutation of t.

permute side-effects random_state by repeated calls to Random.State.int. If random_state is not supplied, permute uses Random.State.default.

val random_element : 'a. ?random_state:Base.Random.State.t -> 'a t -> 'a option @@ portable

random_element ?random_state t is None if t is empty, else it is Some x for some x chosen uniformly at random from t.

random_element side-effects random_state by calling Random.State.int. If random_state is not supplied, random_element uses Random.State.default.

val random_element_exn : 'a. ?random_state:Base.Random.State.t -> 'a t -> 'a @@ portable

val is_sorted : 'a. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> bool @@ portable

is_sorted t ~compare returns true iff for all adjacent a1; a2 in t, compare a1 a2 <= 0.

is_sorted_strictly is similar, except it uses < instead of <=.

val is_sorted_strictly : 'a. 'a t -> compare:('a -> ('a -> int) @ local) @ local -> bool @@ portable

module Infix = Base.List.Infix

val transpose : 'a. 'a t t -> 'a t t option @@ portable

transpose m transposes the rows and columns of the matrix m, considered as either a row of column lists or (dually) a column of row lists.

Example:

  transpose [ [ 1; 2; 3 ]; [ 4; 5; 6 ] ] = [ [ 1; 4 ]; [ 2; 5 ]; [ 3; 6 ] ]

On non-empty rectangular matrices, transpose is an involution (i.e., transpose (transpose m) = m). Transpose returns None when called on lists of lists with non-uniform lengths.

val transpose_exn : 'a. 'a t t -> 'a t t @@ portable

transpose_exn transposes the rows and columns of its argument, throwing an exception if the list is not rectangular.

val intersperse : 'a. 'a t -> sep:'a -> 'a t @@ portable

intersperse xs ~sep places sep between adjacent elements of xs. For example, intersperse [1;2;3] ~sep:0 = [1;0;2;0;3].

module Private = Base.List.Private

include sig ... end

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : 'a. ('a : value_or_null) Bin_prot.Size.sizer -> ('a : value_or_null) Base.List.t Bin_prot.Size.sizer

val bin_size_t__local : 'a. ('a : value_or_null) Bin_prot.Size.sizer__local -> ('a : value_or_null) Base.List.t Bin_prot.Size.sizer__local

val bin_write_t : 'a. ('a : value_or_null) Bin_prot.Write.writer -> ('a : value_or_null) Base.List.t Bin_prot.Write.writer

val bin_write_t__local : 'a. ('a : value_or_null) Bin_prot.Write.writer__local -> ('a : value_or_null) Base.List.t Bin_prot.Write.writer__local

val bin_writer_t : 'a. ('a : value_or_null) Bin_prot.Type_class.writer -> ('a : value_or_null) Base.List.t Bin_prot.Type_class.writer

val bin_read_t : 'a. ('a : value_or_null) Bin_prot.Read.reader -> ('a : value_or_null) Base.List.t Bin_prot.Read.reader

val __bin_read_t__ : 'a. ('a : value_or_null) Bin_prot.Read.reader -> ('a : value_or_null) Base.List.t Bin_prot.Read.vtag_reader

val bin_reader_t : 'a. ('a : value_or_null) Bin_prot.Type_class.reader -> ('a : value_or_null) Base.List.t Bin_prot.Type_class.reader

val bin_t : 'a. ('a : value_or_null) Bin_prot.Type_class.t -> ('a : value_or_null) Base.List.t Bin_prot.Type_class.t

val typerep_of_t : 'a Typerep_lib.Std.Typerep.t -> 'a Base.List.t Typerep_lib.Std.Typerep.t @@ portable

val typename_of_t : 'a Typerep_lib.Std.Typename.t -> 'a Base.List.t Typerep_lib.Std.Typename.t @@ portable

module Assoc : sig ... end

Extensions

exception Duplicate_found of Base.Unit.t -> Base.Sexp.t * Base.String.t

Only raised in exn_if_dup below.

deprecated [since 2018-03] stop matching on Duplicate_found. [exn_if_dup] will eventually raise a different and unspecified exception

val exn_if_dup : 'a. compare:('a -> 'a -> Base.Int.t) -> ?context:Base.String.t -> 'a t -> to_sexp:('a -> Base.Sexp.t) -> Base.Unit.t

exn_if_dup ~compare ?context t ~to_sexp raises if t contains a duplicate. It will specifically raise a Duplicate_found exception and use context as its second argument. O(n log n) time complexity.

val slice : 'a t -> Base.Int.t -> Base.Int.t -> 'a t @@ portable

slice t start stop returns a new list including elements t.(start) through t.(stop-1), normalized Python-style with the exception that stop = 0 is treated as stop = length t. This can still raise if indices are too large in the negative or positive direction.

Consider using List.sub instead, which has a more conventional interface that does not require reasoning about negative indices.

include Comparator.Derived__portable with type 'a t := 'a t

type !'cmp comparator_witness

val comparator : ('a, 'cmp) Base.Comparator.T.comparator -> ('a t, 'cmp comparator_witness) Base.Comparator.T.comparator

include Quickcheckable.S1__portable with type 'a t := 'a t

include Quickcheck_intf.Definitions.S1 with type 'a t := 'a t

val quickcheck_generator : 'a Base_quickcheck.Generator.t -> 'a t Base_quickcheck.Generator.t

val quickcheck_observer : 'a Base_quickcheck.Observer.t -> 'a t Base_quickcheck.Observer.t

val quickcheck_shrinker : 'a Base_quickcheck.Shrinker.t -> 'a t Base_quickcheck.Shrinker.t

val to_string : 'a. f:('a -> Base.String.t) -> 'a t -> Base.String.t @@ portable

val gen_non_empty : 'a Quickcheck.Generator.t -> 'a t Quickcheck.Generator.t @@ portable

Like gen, but never generates the empty list.

val gen_with_length : Base.Int.t -> 'a Quickcheck.Generator.t -> 'a t Quickcheck.Generator.t @@ portable

Like gen, but generates lists with the given length.

val gen_filtered : 'a t -> 'a t Quickcheck.Generator.t @@ portable

Randomly drops elements from the input list. Length is chosen uniformly between 0 and the length of the input, inclusive.

val gen_permutations : 'a t -> 'a t Quickcheck.Generator.t @@ portable

gen_permutations t generates all permutations of list. If t contains duplicate values, then gen_permutations t will produce duplicate lists.

val zip_with_remainder : 'a 'b. 'a Base.List.t -> 'b Base.List.t -> ('a * 'b) Base.List.t * ('a Base.List.t, 'b Base.List.t) Either.t Base.Option.t @@ portable

zip_with_remainder xs ys zips as many elements as possible of xs and ys together and also returns the un-zipped remainder of the longer input, if the inputs have different lengths.

If xs and ys have the same length, zip_with_remainder xs ys returns the same thing as (zip_exn xs ys, None)

module Stable : sig ... end