Module List.Assoc

type ('a, 'b) t = ( 'a, 'b ) Base.List.Assoc.t

include Bin_prot.Binable.S2 with type ('a, 'b) t := ( 'a, 'b ) t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ( 'a, 'b, ( 'a, 'b ) t ) Bin_prot.Size.sizer2

val bin_write_t : ( 'a, 'b, ( 'a, 'b ) t ) Bin_prot.Write.writer2

val bin_read_t : ( 'a, 'b, ( 'a, 'b ) t ) Bin_prot.Read.reader2

val __bin_read_t__ : ( 'a, 'b, int -> ( 'a, 'b ) t ) Bin_prot.Read.reader2

val bin_writer_t : ( 'a, 'b, ( 'a, 'b ) t ) Bin_prot.Type_class.S2.writer

val bin_reader_t : ( 'a, 'b, ( 'a, 'b ) t ) Bin_prot.Type_class.S2.reader

val bin_t : ( 'a, 'b, ( 'a, 'b ) t ) Bin_prot.Type_class.S2.t

val compare : ( 'a -> 'a -> Base.Int.t ) -> ( 'b -> 'b -> Base.Int.t ) -> ( 'a, 'b ) t -> ( 'a, 'b ) t -> Base.Int.t

include module type of struct include Base.List.Assoc end with type ('a, 'b) t := ( 'a, 'b ) t

Interpret a list of (key, value) pairs as a map in which only the first occurrence of a key affects the semantics, i.e.:

List.Assoc.xxx alist ...args... 

is always the same as (or at least sort of isomorphic to):

Map.xxx (alist |> Map.of_alist_multi |> Map.map ~f:List.hd) ...args... 

include Sexplib0.Sexpable.S2 with type ('a, 'b) t := ( 'a, 'b ) t

val t_of_sexp : ( Sexplib0__.Sexp.t -> 'a ) -> ( Sexplib0__.Sexp.t -> 'b ) -> Sexplib0__.Sexp.t -> ( 'a, 'b ) t

val sexp_of_t : ( 'a -> Sexplib0__.Sexp.t ) -> ( 'b -> Sexplib0__.Sexp.t ) -> ( 'a, 'b ) t -> Sexplib0__.Sexp.t

val t_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'b Sexplib0.Sexp_grammar.t -> ( 'a, 'b ) t Sexplib0.Sexp_grammar.t

val add : ( 'a, 'b ) t -> equal:( 'a -> 'a -> bool ) -> 'a -> 'b -> ( 'a, 'b ) t

val find : ( 'a, 'b ) t -> equal:( 'a -> 'a -> bool ) -> 'a -> 'b option

val find_exn : ( 'a, 'b ) t -> equal:( 'a -> 'a -> bool ) -> 'a -> 'b

val mem : ( 'a, 'b ) t -> equal:( 'a -> 'a -> bool ) -> 'a -> bool

val remove : ( 'a, 'b ) t -> equal:( 'a -> 'a -> bool ) -> 'a -> ( 'a, 'b ) t

val map : ( 'a, 'b ) t -> f:( 'b -> 'c ) -> ( 'a, 'c ) t

val inverse : ( 'a, 'b ) t -> ( 'b, 'a ) t

Bijectivity is not guaranteed because we allow a key to appear more than once.

val group : ('a * 'b) list -> equal:( 'a -> 'a -> bool ) -> ( 'a, 'b list ) t

Converts an association list with potential consecutive duplicate keys into an association list of (non-empty) lists with no (consecutive) duplicate keys. Any non-consecutive duplicate keys in the input will remain in the output.

val sort_and_group : ('a * 'b) list -> compare:( 'a -> 'a -> int ) -> ( 'a, 'b list ) t

Converts an association list with potential duplicate keys into an association list of (non-empty) lists with no duplicate keys.