Module List.Cartesian_product

Implements cartesian-product behavior for map and bind. *

include Base.Applicative.S with type 'a t := 'a list

val both : 'a list -> 'b list -> ('a * 'b) list

val (<*>) : ( 'a -> 'b ) list -> 'a list -> 'b list

same as apply

val (<*) : 'a list -> unit list -> 'a list

val (*>) : unit list -> 'a list -> 'a list

val (>>|) : 'a list -> ( 'a -> 'b ) -> 'b list

val apply : ( 'a -> 'b ) list -> 'a list -> 'b list

val map2 : 'a list -> 'b list -> f:( 'a -> 'b -> 'c ) -> 'c list

val map3 : 'a list -> 'b list -> 'c list -> f:( 'a -> 'b -> 'c -> 'd ) -> 'd list

module Applicative_infix : sig ... end

include Base.Monad.S with type 'a t := 'a list

val (>>=) : 'a list -> ( 'a -> 'b list ) -> 'b list

t >>= f returns a computation that sequences the computations represented by two monad elements. The resulting computation first does t to yield a value v, and then runs the computation returned by f v.

val (>>|) : 'a list -> ( 'a -> 'b ) -> 'b list

t >>| f is t >>= (fun a -> return (f a)).

module Monad_infix : sig ... end

val bind : 'a list -> f:( 'a -> 'b list ) -> 'b list

bind t ~f = t >>= f

val return : 'a -> 'a list

return v returns the (trivial) computation that returns v.

val map : 'a list -> f:( 'a -> 'b ) -> 'b list

map t ~f is t >>| f.

val join : 'a list list -> 'a list

join t is t >>= (fun t' -> t').

val ignore_m : 'a list -> unit list

ignore_m t is map t ~f:(fun _ -> ()). ignore_m used to be called ignore, but we decided that was a bad name, because it shadowed the widely used Caml.ignore. Some monads still do let ignore = ignore_m for historical reasons.

val all : 'a list list -> 'a list list

val all_unit : unit list list -> unit list

Like all, but ensures that every monadic value in the list produces a unit value, all of which are discarded rather than being collected into a list.

module Let_syntax : sig ... end

These are convenient to have in scope when programming with a monad: