Module Of_binable.V1

Parameters

module Binable : Minimal.S

module M : Conv_without_uuid with type binable := Binable.t

Signature

include Bin_prot.Binable.S_only_functions with type t := M.t

val bin_size_t : M.t Bin_prot.Size.sizer

val bin_write_t : M.t Bin_prot.Write.writer

val bin_read_t : M.t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> M.t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : M.t Bin_prot.Type_class.writer

val bin_reader_t : M.t Bin_prot.Type_class.reader

val bin_t : M.t Bin_prot.Type_class.t