Module Nothing.Hash_set

type elt = t

type t = elt Hash_set.t

include Ppx_compare_lib.Equal.S with type t := t

val equal : t -> t -> bool

val sexp_of_t : t -> Sexplib0.Sexp.t

include Base.Hash_set.Creators_generic with type 'a t := t with type 'a elt := elt with type ('a, 'z) create_options := ( 'a, 'z ) Base.Hash_set.create_options_without_first_class_module

val create : ( 'a, unit -> t ) Base.Hash_set.create_options_without_first_class_module

val of_list : ( 'a, elt list -> t ) Base.Hash_set.create_options_without_first_class_module

module Provide_of_sexp (X : sig ... end) : sig ... end

module Provide_bin_io (X : sig ... end) : sig ... end

val t_of_sexp : Sexplib0.Sexp.t -> t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t