Module Core.Option_array

This module extends Base.Option_array with bin_io.

type 'a t = 'a Base.Option_array.t

include Bin_prot.Binable.S1 with type 'a t := 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ( 'a, 'a t ) Bin_prot.Size.sizer1

val bin_write_t : ( 'a, 'a t ) Bin_prot.Write.writer1

val bin_read_t : ( 'a, 'a t ) Bin_prot.Read.reader1

val __bin_read_t__ : ( 'a, int -> 'a t ) Bin_prot.Read.reader1

val bin_writer_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.writer

val bin_reader_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.reader

val bin_t : ( 'a, 'a t ) Bin_prot.Type_class.S1.t

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : ( Sexplib0__.Sexp.t -> 'a ) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ( 'a -> Sexplib0__.Sexp.t ) -> 'a t -> Sexplib0__.Sexp.t

include module type of struct include Base.Option_array end with type 'a t := 'a t

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : ( Sexplib0__.Sexp.t -> 'a ) -> Sexplib0__.Sexp.t -> 'a t

val sexp_of_t : ( 'a -> Sexplib0__.Sexp.t ) -> 'a t -> Sexplib0__.Sexp.t

val t_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'a t Sexplib0.Sexp_grammar.t

val empty : _ t

val create : len:int -> _ t

Initially filled with all None

include Base.Indexed_container.Generic with type 'a t := 'a t and type 'a elt := 'a option

include Base.Container.Generic with type 'a t := 'a t with type 'a elt := 'a option

val length : _ t -> int

val is_empty : _ t -> bool

val iter : 'a t -> f:( 'a option -> unit ) -> unit

val fold : 'a t -> init:'accum -> f:( 'accum -> 'a option -> 'accum ) -> 'accum

val fold_result : 'a t -> init:'accum -> f:( 'accum -> 'a option -> ( 'accum, 'e ) Base.Result.t ) -> ( 'accum, 'e ) Base.Result.t

val fold_until : 'a t -> init:'accum -> f: ( 'accum -> 'a option -> ( 'accum, 'final ) Base.Container.Continue_or_stop.t ) -> finish:( 'accum -> 'final ) -> 'final

val exists : 'a t -> f:( 'a option -> bool ) -> bool

val for_all : 'a t -> f:( 'a option -> bool ) -> bool

val count : 'a t -> f:( 'a option -> bool ) -> int

val sum : (module Base.Container.Summable with type t = 'sum) -> 'a t -> f:( 'a option -> 'sum ) -> 'sum

val find : 'a t -> f:( 'a option -> bool ) -> 'a option option

val find_map : 'a t -> f:( 'a option -> 'b option ) -> 'b option

val to_list : 'a t -> 'a option list

val min_elt : 'a t -> compare:( 'a option -> 'a option -> int ) -> 'a option option

val max_elt : 'a t -> compare:( 'a option -> 'a option -> int ) -> 'a option option

These are all like their equivalents in Container except that an index starting at 0 is added as the first argument to f.

val foldi : 'a t -> init:_ -> f:( int -> _ -> 'a option -> _ ) -> _

val iteri : 'a t -> f:( int -> 'a option -> unit ) -> unit

val existsi : 'a t -> f:( int -> 'a option -> bool ) -> bool

val for_alli : 'a t -> f:( int -> 'a option -> bool ) -> bool

val counti : 'a t -> f:( int -> 'a option -> bool ) -> int

val findi : 'a t -> f:( int -> 'a option -> bool ) -> (int * 'a option) option

val find_mapi : 'a t -> f:( int -> 'a option -> 'b option ) -> 'b option

val init_some : int -> f:( int -> 'a ) -> 'a t

val init : int -> f:( int -> 'a option ) -> 'a t

val of_array : 'a option array -> 'a t

val of_array_some : 'a array -> 'a t

val to_array : 'a t -> 'a option Base.Array.t

val get : 'a t -> int -> 'a option

get t i returns the element number i of array t, raising if i is outside the range 0 to length t - 1.

val get_some_exn : 'a t -> int -> 'a

Raises if the element number i is None.

val is_none : _ t -> int -> bool

is_none t i = Option.is_none (get t i)

val is_some : _ t -> int -> bool

is_some t i = Option.is_some (get t i)

These can cause arbitrary behavior when used for an out-of-bounds array access.

val unsafe_get : 'a t -> int -> 'a option

val unsafe_get_some_exn : 'a t -> int -> 'a

unsafe_get_some_exn t i is unsafe because it does not bounds check i. It does, however check whether the value at index i is none or some, and raises if it is none.

val unsafe_get_some_assuming_some : 'a t -> int -> 'a

unsafe_get_some_assuming_some t i is unsafe both because it does not bounds check i and because it does not check whether the value at index i is none or some, assuming that it is some.

val unsafe_is_some : _ t -> int -> bool

val set : 'a t -> int -> 'a option -> unit

set t i x modifies array t in place, replacing element number i with x, raising if i is outside the range 0 to length t - 1.

val set_some : 'a t -> int -> 'a -> unit

val set_none : _ t -> int -> unit

val swap : _ t -> int -> int -> unit

val clear : _ t -> unit

Replaces all the elements of the array with None.

val map : 'a t -> f:( 'a option -> 'b option ) -> 'b t

map f [|a1; ...; an|] applies function f to a1, a2, ..., an, in order, and builds the option_array [|f a1; ...; f an|] with the results returned by f.

val map_some : 'a t -> f:( 'a -> 'b ) -> 'b t

map_some t ~f is like map, but None elements always map to None and Some always map to Some.

Unsafe versions of set*. Can cause arbitrary behaviour when used for an out-of-bounds array access.

val unsafe_set : 'a t -> int -> 'a option -> unit

val unsafe_set_some : 'a t -> int -> 'a -> unit

val unsafe_set_none : _ t -> int -> unit

include Base.Blit.S1 with type 'a t := 'a t

val blit : src:'a t -> src_pos:int -> dst:'a t -> dst_pos:int -> len:int -> unit

val blito : src:'a t -> ?src_pos:int -> ?src_len:int -> dst:'a t -> ?dst_pos:int -> unit -> unit

val unsafe_blit : src:'a t -> src_pos:int -> dst:'a t -> dst_pos:int -> len:int -> unit

val sub : 'a t -> pos:int -> len:int -> 'a t

val subo : ?pos:int -> ?len:int -> 'a t -> 'a t

val copy : 'a t -> 'a t

Makes a (shallow) copy of the array.