Module Core.Ordering

Extends Base.Ordering, intended to make code that matches on the result of a comparison more concise and easier to read.

type t = Base.Ordering.t =

| Less

| Equal

| Greater

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val compare : t -> t -> int

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

val all : t list

include Base.Equal.S with type t := t

val equal : t Base.Equal.equal

val of_int : int -> t

of_int n is:

      Less     if n < 0
      Equal    if n = 0
      Greater  if n > 0

val to_int : t -> int

to_int t is:

      Less     -> -1
      Equal    -> 0
      Greater  -> 1

It can be useful when writing a comparison function to allow one to return Ordering.t values and transform them to ints later.

module Export : sig ... end