Module Percent.Option

The value nan cannot be represented as an Option.t

type value := t

type t = private Base.Float.t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

include Immediate_option.S_without_immediate with type value := value and type t := t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

include Immediate_option_intf.S_without_immediate_plain with type t := t with type value := value

Constructors analogous to None and Some. If not (some_is_representable x) then some x may raise or return none.

val none : t

val some : value -> t

val some_is_representable : value -> Base.Bool.t

For some representations of immediate options, the encodings of none and some overlap. For these representations, some_is_representable value = false if value cannot be represented as an option. For example, Int.Option uses min_value to represent none. For other representations, some_is_representable always returns true.

val is_none : t -> Base.Bool.t

val is_some : t -> Base.Bool.t

val value : t -> default:value -> value

value (some x) ~default = x and value none ~default = default.

val value_exn : t -> value

value_exn (some x) = x. value_exn none raises. Unlike Option.value_exn, there is no ?message argument, so that calls to value_exn that do not raise also do not have to allocate.

val unchecked_value : t -> value

unchecked_value (some x) = x. unchecked_value none returns an unspecified value. unchecked_value t is intended as an optimization of value_exn t when is_some t is known to be true.

val to_option : t -> value Base.Option.t

val of_option : value Base.Option.t -> t

module Optional_syntax : Optional_syntax.S with type t := t with type value := value

val apply_with_none_as_nan : t -> Base.Float.t -> Base.Float.t

apply_with_none_as_nan (some x) y = apply x y, and apply_with_none_as_nan none y = apply (of_mult Float.nan) y

val of_mult_with_nan_as_none : Base.Float.t -> t

of_mult_with_nan_as_none Float.nan = none, and of_mult_with_nan_as_none x = some (of_mult x) otherwise

val to_mult_with_none_as_nan : t -> Base.Float.t

to_mult_with_none_as_nan none = Float.nan, and to_mult_with_none_as_nan (some x) = to_mult x