Module Stable.V1

module Bin_shape_same_as_float : sig ... end

Tl;dr: For new code use Stable.V3 if you care about exact round-trippability via sexp, or Almost_round_trippable if you want to round your output to 14 significant digits to hide common floating-point rounding errors and make it less of an eyesore but also less accurate.