Module V1.Set

include Set.Stable.V1.S with type elt := t with type elt_comparator_witness := comparator_witness

type nonrec t = (t, comparator_witness) Set.Stable.V1.t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Diffable.S with type t := t and type Diff.t = t Diffable.Set_diff.Stable.V1.t

module Diff : sig ... end

val stable_witness : t Stable_witness.t