Module type Quickcheck_intf.S_int

include S_range

include S

type t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val gen_incl : t -> t -> t Base_quickcheck.Generator.t

gen_incl lower_bound upper_bound produces values between lower_bound and upper_bound, inclusive. It uses an ad hoc distribution that stresses boundary conditions more often than a uniform distribution, while still able to produce any value in the range. Raises if lower_bound > upper_bound.

val gen_uniform_incl : t -> t -> t Base_quickcheck.Generator.t

gen_uniform_incl lower_bound upper_bound produces a generator for values uniformly distributed between lower_bound and upper_bound, inclusive. Raises if lower_bound > upper_bound.

val gen_log_uniform_incl : t -> t -> t Base_quickcheck.Generator.t

gen_log_uniform_incl lower_bound upper_bound produces a generator for values between lower_bound and upper_bound, inclusive, where the number of bits used to represent the value is uniformly distributed. Raises if (lower_bound < 0) || (lower_bound > upper_bound).

val gen_log_incl : t -> t -> t Base_quickcheck.Generator.t

gen_log_incl lower_bound upper_bound is like gen_log_uniform_incl, but weighted slightly more in favor of generating lower_bound and upper_bound specifically.