Module type Quickcheck_intf.Syntax

Includes Let_syntax from Monad.Syntax. Sets Open_on_rhs to be all of Generator, except that it does not shadow Let_syntax itself. Both Generator and Open_on_rhs are meant to be destructively assigned.

module Generator : Generator

An 'a t a generates values of type 'a with a specific probability distribution.

module Open_on_rhs : Generator with type 'a t := 'a Generator.t and module Let_syntax := Generator.Let_syntax

An 'a t a generates values of type 'a with a specific probability distribution.

include Base.Monad.Syntax with type 'a t := 'a Generator.t and module Let_syntax.Let_syntax.Open_on_rhs = Open_on_rhs

module Let_syntax : sig ... end

These are convenient to have in scope when programming with a monad: