Module Stable.V1_stable_unit_test

We export the unit test arg rather than instantiate the functor inside result.ml in order to avoid circular dependencies. The functor is instantiated in stable.ml.

type t

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : ( int -> t ) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val equal : t -> t -> Base.Bool.t

val tests : (t * Base.String.t * Base.String.t) Base.List.t

tests is a list of (value, sexp-representation, bin-io-representation) triples. The unit tests check that the type properly serializes and de-serializes according to the given representations.