Module Stable_unit_test.Make_sexp_deserialization_test

We provide separate access to sexp serialization and deserialization tests because some stable types will have varying sexp serializations. Notably, Time.sexp_of_t depends on the local timezone. For such types it is still important to check that all the sexps can be deserialized correctly.

Parameters

module T : sig ... end

Signature