Module String.Stable

Note that string is already stable by itself, since as a primitive type it is an integral part of the sexp / bin_io protocol. String.Stable exists only to introduce String.Stable.Set, String.Stable.Map, String.Stable.Table, and provide interface uniformity with other stable types.

module type Identifiable_without_binio := sig ... end

module V1 : sig ... end

module Utf8 : sig ... end

module Utf16le : sig ... end

module Utf16be : sig ... end

module Utf32le : sig ... end

module Utf32be : sig ... end