Module Stable.V1

type nonrec t = t

include Ppx_compare_lib.Equal.S with type t := t

val equal : t -> t -> bool

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

include Base.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Stable_comparable.With_stable_witness.V1 with type t := t with type comparator_witness = comparator_witness

val stable_witness : t Ppx_stable_witness_runtime.Stable_witness.t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Comparator.Stable.V1.S with type t := t with type comparator_witness = comparator_witness

type comparator_witness = comparator_witness

val comparator : (t, comparator_witness) Comparator.Stable.V1.comparator

include Comparable.Stable.V1.With_stable_witness.S with type comparable := t with type comparator_witness := comparator_witness

module Map : Map.Stable.V1.With_stable_witness.S with type key := t with type comparator_witness := comparator_witness

module Set : Set.Stable.V1.With_stable_witness.S with type elt := t with type elt_comparator_witness := comparator_witness

include Hashable.Stable.V1.With_stable_witness.S with type key := t

module Table : sig ... end

module Hash_set : sig ... end

val hashable : t Base.Hashable.t

include Diffable.S_atomic with type t := t

module Diff : sig ... end