Module type Substring_intf.S

type base

The type of strings that type t is a substring of.

type t

include Ppx_quickcheck_runtime.Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

include Base.Indexed_container.S0 with type t := t with type elt := Base.Char.t

include Base.Container.S0 with type t := t with type elt := Base.Char.t

val mem : t -> Base.Char.t -> bool

Checks whether the provided element is there, using equality on elts.

val length : t -> int

val is_empty : t -> bool

val iter : t -> f:( Base.Char.t -> unit ) -> unit

iter must allow exceptions raised in f to escape, terminating the iteration cleanly. The same holds for all functions below taking an f.

val fold : t -> init:'accum -> f:( 'accum -> Base.Char.t -> 'accum ) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t.

val fold_result : t -> init:'accum -> f:( 'accum -> Base.Char.t -> ( 'accum, 'e ) Base.Result.t ) -> ( 'accum, 'e ) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : t -> init:'accum -> f: ( 'accum -> Base.Char.t -> ( 'accum, 'final ) Base.Container.Continue_or_stop.t ) -> finish:( 'accum -> 'final ) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : t -> f:( Base.Char.t -> bool ) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : t -> f:( Base.Char.t -> bool ) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : t -> f:( Base.Char.t -> bool ) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base.Container.Summable with type t = 'sum) -> t -> f:( Base.Char.t -> 'sum ) -> 'sum

Returns the sum of f i for all i in the container.

val find : t -> f:( Base.Char.t -> bool ) -> Base.Char.t option

Returns as an option the first element for which f evaluates to true.

val find_map : t -> f:( Base.Char.t -> 'a option ) -> 'a option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : t -> Base.Char.t list

val to_array : t -> Base.Char.t array

val min_elt : t -> compare:( Base.Char.t -> Base.Char.t -> int ) -> Base.Char.t option

Returns a min (resp. max) element from the collection using the provided compare function. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold. Returns None iff the collection is empty.

val max_elt : t -> compare:( Base.Char.t -> Base.Char.t -> int ) -> Base.Char.t option

These are all like their equivalents in Container except that an index starting at 0 is added as the first argument to f.

val foldi : t -> init:_ -> f:( int -> _ -> Base.Char.t -> _ ) -> _

val iteri : t -> f:( int -> Base.Char.t -> unit ) -> unit

val existsi : t -> f:( int -> Base.Char.t -> bool ) -> bool

val for_alli : t -> f:( int -> Base.Char.t -> bool ) -> bool

val counti : t -> f:( int -> Base.Char.t -> bool ) -> int

val findi : t -> f:( int -> Base.Char.t -> bool ) -> (int * Base.Char.t) option

val find_mapi : t -> f:( int -> Base.Char.t -> 'a option ) -> 'a option

val base : t -> base

val pos : t -> Base.Int.t

pos refers to the position in the base string, not any other substring that this substring was generated from.

val get : t -> Base.Int.t -> Base.Char.t

Per String.get and Bigstring.get, this raises an exception if the index is out of bounds.

val create : ?pos:Base.Int.t -> ?len:Base.Int.t -> base -> t

create ?pos ?len base creates a substring of the base sequence of length len starting at position pos, i.e.,

base.[pos], base.[pos + 1], ... base.[pos + len - 1] 

An exception is raised if any of those indices into base is invalid.

It does not copy the characters, so mutating base mutates t and vice versa.

val sub : ?pos:Base.Int.t -> ?len:Base.Int.t -> t -> t

Blit functions

For copying characters from a substring to and from both strings and substrings.

val blit_to_string : t -> dst:Base.Bytes.t -> dst_pos:Base.Int.t -> Base.Unit.t

val blit_to_bytes : t -> dst:Base.Bytes.t -> dst_pos:Base.Int.t -> Base.Unit.t

val blit_to_bigstring : t -> dst:Bigstring.t -> dst_pos:Base.Int.t -> Base.Unit.t

val blit_from_string : t -> src:Base.String.t -> src_pos:Base.Int.t -> len:Base.Int.t -> Base.Unit.t

val blit_from_bigstring : t -> src:Bigstring.t -> src_pos:Base.Int.t -> len:Base.Int.t -> Base.Unit.t

String concatenation

These functions always copy.

val concat : t Base.List.t -> t

val concat_string : t Base.List.t -> Base.String.t

val concat_bigstring : t Base.List.t -> Bigstring.t

Conversion to/from substrings

These functions always copy.

val to_string : t -> Base.String.t

val to_bigstring : t -> Bigstring.t

These functions always copy. Use create if you want sharing.

val of_string : Base.String.t -> t

val of_bigstring : Bigstring.t -> t

Prefixes and suffixes

The result of these functions share data with their input, but don't mutate the underlying string.

val drop_prefix : t -> Base.Int.t -> t

val drop_suffix : t -> Base.Int.t -> t

val prefix : t -> Base.Int.t -> t

val suffix : t -> Base.Int.t -> t