Module Core.Time

module Span : sig ... end

module Zone : sig ... end

module Ofday : sig ... end

type t = Base.Float.t

A fully qualified point in time, independent of timezone.

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val compare : t -> t -> int

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

module type S_kernel_without_zone = sig ... end

module type S_kernel = sig ... end

type underlying = Base.Float.t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val compare : t -> t -> Base.Int.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

module Replace_polymorphic_compare : Core_kernel.Comparable.Polymorphic_compare with type t := t

include Core_kernel.Comparable.Polymorphic_compare with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val (>=.) : t -> t -> bool

val (<=.) : t -> t -> bool

val (=.) : t -> t -> bool

val (>.) : t -> t -> bool

val (<.) : t -> t -> bool

val (<>.) : t -> t -> bool

val robustly_compare : t -> t -> int

val add : t -> Base.Float.t -> t

val sub : t -> Base.Float.t -> t

val diff : t -> t -> Base.Float.t

val next : t -> t

next t returns the next t (forwards in time)

val prev : t -> t

prev t returns the previous t (backwards in time)

val to_span_since_epoch : t -> Base.Float.t

val of_span_since_epoch : Base.Float.t -> t

include Core_kernel.Comparable.S_common with type t := t and module Replace_polymorphic_compare := Replace_polymorphic_compare

include Base.Comparable.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

module Date_and_ofday : sig ... end with type Date_and_ofday.absolute := t

Equivalent to a Date.t and an Ofday.t with no time zone. A Date_and_ofday.t does not correspond to a single, unambiguous point in time.

val next_multiple : ?⁠can_equal_after:Base.Bool.t -> base:t -> after:t -> interval:Base.Float.t -> Base.Unit.t -> t

next_multiple ~base ~after ~interval returns the smallest time of the form:

time = base + k * interval

where k >= 0 and time > after. It is an error if interval <= 0.

Supplying ~can_equal_after:true allows the result to satisfy time >= after.

val prev_multiple : ?⁠can_equal_before:Base.Bool.t -> base:t -> before:t -> interval:Base.Float.t -> Base.Unit.t -> t

prev_multiple ~base ~before ~interval returns the largest time of the form:

time = base + k * interval

where k >= 0 and time < before. It is an error if interval <= 0.

Supplying ~can_equal_before:true allows the result to satisfy time <= before.

val now : Base.Unit.t -> t

module Stable : sig ... end

val now : Base.Unit.t -> t

now () returns a t representing the current time

Basic operations on times

val add : t -> Base.Float.t -> t

add t s adds the span s to time t and returns the resulting time.

NOTE: adding spans as a means of adding days is not accurate, and may run into trouble due to shifts in daylight savings time, float arithmetic issues, and leap seconds. See the comment at the top of Zone.mli for a more complete discussion of some of the issues of time-keeping. For spans that cross date boundaries, use date functions instead.

val sub : t -> Base.Float.t -> t

sub t s subtracts the span s from time t and returns the resulting time. See important note for add.

val diff : t -> t -> Base.Float.t

diff t1 t2 returns time t1 minus time t2.

val abs_diff : t -> t -> Base.Float.t

abs_diff t1 t2 returns the absolute span of time t1 minus time t2.

include Core_kernel.Quickcheck.S_range with type t := t

include Core_kernel.Quickcheck_intf.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val gen_incl : t -> t -> t Base_quickcheck.Generator.t

gen_incl lower_bound upper_bound produces values between lower_bound and upper_bound, inclusive. It uses an ad hoc distribution that stresses boundary conditions more often than a uniform distribution, while still able to produce any value in the range. Raises if lower_bound > upper_bound.

val gen_uniform_incl : t -> t -> t Base_quickcheck.Generator.t

gen_uniform_incl lower_bound upper_bound produces a generator for values uniformly distributed between lower_bound and upper_bound, inclusive. Raises if lower_bound > upper_bound.

Comparisons

val is_earlier : t -> than:t -> Base.Bool.t

val is_later : t -> than:t -> Base.Bool.t

Conversions

val of_date_ofday : zone:Core_kernel__Zone.t -> Core_kernel__Date0.t -> Base.Float.t -> t

val of_date_ofday_precise : Core_kernel__Date0.t -> Base.Float.t -> zone:Core_kernel__Zone.t -> [ `Once of t | `Twice of t * t | `Never of t ]

Because timezone offsets change throughout the year (clocks go forward or back) some local times can occur twice or not at all. In the case that they occur twice, this function gives `Twice with both occurrences in order; if they do not occur at all, this function gives `Never with the time at which the local clock skips over the desired time of day.

Note that this is really only intended to work with DST transitions and not unusual or dramatic changes, like the calendar change in 1752 (run "cal 9 1752" in a shell to see). In particular it makes the assumption that midnight of each day is unambiguous.

Most callers should use of_date_ofday rather than this function. In the `Twice and `Never cases, of_date_ofday will return reasonable times for most uses.

val to_date_ofday : t -> zone:Core_kernel__Zone.t -> Core_kernel__Date0.t * Base.Float.t

val to_date_ofday_precise : t -> zone:Core_kernel__Zone.t -> Core_kernel__Date0.t * Base.Float.t * [ `Only | `Also_at of t | `Also_skipped of Core_kernel__Date0.t * Base.Float.t ]

Always returns the Date.t * Ofday.t that to_date_ofday would have returned, and in addition returns a variant indicating whether the time is associated with a time zone transition.

      - `Only         -> there is a one-to-one mapping between [t]'s and
                         [Date.t * Ofday.t] pairs
      - `Also_at      -> there is another [t] that maps to the same [Date.t * Ofday.t]
                         (this date/time pair happened twice because the clock fell back)
      - `Also_skipped -> there is another [Date.t * Ofday.t] pair that never happened (due
                         to a jump forward) that [of_date_ofday] would map to the same
                         [t].

val to_date : t -> zone:Core_kernel__Zone.t -> Core_kernel__Date0.t

val to_ofday : t -> zone:Core_kernel__Zone.t -> Base.Float.t

val reset_date_cache : Base.Unit.t -> Base.Unit.t

For performance testing only; reset_date_cache () resets an internal cache used to speed up to_date and related functions when called repeatedly on times that fall within the same day.

val epoch : t

midnight, Jan 1, 1970 in UTC

val convert : from_tz:Core_kernel__Zone.t -> to_tz:Core_kernel__Zone.t -> Core_kernel__Date0.t -> Base.Float.t -> Core_kernel__Date0.t * Base.Float.t

It's unspecified what happens if the given date/ofday/zone correspond to more than one date/ofday pair in the other zone.

val utc_offset : t -> zone:Core_kernel__Zone.t -> Base.Float.t

Other string conversions

The {to,of}_string functions in Time convert to UTC time, because a local time zone is not necessarily available. They are generous in what they will read in.

val of_string : string -> t

val to_string : t -> string

val to_filename_string : t -> zone:Core_kernel__Zone.t -> Base.String.t

to_filename_string t ~zone converts t to string with format YYYY-MM-DD_HH-MM-SS.mmm which is suitable for using in filenames.

val of_filename_string : Base.String.t -> zone:Core_kernel__Zone.t -> t

of_filename_string s ~zone converts s that has format YYYY-MM-DD_HH-MM-SS.mmm into time.

val to_string_abs : t -> zone:Core_kernel__Zone.t -> Base.String.t

to_string_abs ~zone t is the same as to_string t except that it uses the given time zone.

val to_string_abs_trimmed : t -> zone:Core_kernel__Zone.t -> Base.String.t

to_string_abs_trimmed is the same as to_string_abs, but drops trailing seconds and milliseconds if they are 0.

val to_string_abs_parts : t -> zone:Core_kernel__Zone.t -> Base.String.t Base.List.t

val to_string_trimmed : t -> zone:Core_kernel__Zone.t -> Base.String.t

Same as to_string_abs_trimmed, except it leaves off the timezone, so won't reliably round trip.

val to_sec_string : t -> zone:Core_kernel__Zone.t -> Base.String.t

Same as to_string_abs, but without milliseconds

val of_localized_string : zone:Core_kernel__Zone.t -> Base.String.t -> t

of_localized_string ~zone str read in the given string assuming that it represents a time in zone and return the appropriate Time.t

val of_string_gen : default_zone:(Base.Unit.t -> Core_kernel__Zone.t) -> find_zone:(Base.String.t -> Core_kernel__Zone.t) -> Base.String.t -> t

of_string_gen ~default_zone ~find_zone s attempts to parse s as a t, calling out to default_zone and find_zone as needed.

val to_string_iso8601_basic : t -> zone:Core_kernel__Zone.t -> Base.String.t

to_string_iso8601_basic return a string representation of the following form: %Y-%m-%dT%H:%M:%S.%s%Z e.g. to_string_iso8601_basic ~zone:Time.Zone.utc epoch = "1970-01-01T00:00:00.000000Z"

val occurrence : [ `First_after_or_at | `Last_before_or_at ] -> t -> ofday:Base.Float.t -> zone:Core_kernel__Zone.t -> t

occurrence side time ~ofday ~zone returns a Time.t that is the occurrence of ofday (in the given zone) that is the latest occurrence (<=) time or the earliest occurrence (>=) time, according to side.

NOTE: If the given time converted to wall clock time in the given zone is equal to ofday then the t returned will be equal to the t given.

module Stable : sig ... end

val arg_type : t Core_kernel.Command.Arg_type.t

String conversions use the local timezone by default. Sexp conversions use get_sexp_zone () by default, which can be overridden by calling set_sexp_zone. These default time zones are used when writing a time, and when reading a time with no explicit zone or UTC offset.

include Core_kernel.Identifiable.S with type t := t and type comparator_witness := comparator_witness and module Replace_polymorphic_compare := Replace_polymorphic_compare

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Core_kernel.Identifiable.S_common with type t := t

val compare : t -> t -> Base.Int.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

include Base.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Base.Pretty_printer.S with type t := t

val pp : Base.Formatter.t -> t -> unit

include Core_kernel.Comparable.S_binable with type t := t with type comparator_witness := comparator_witness with module Replace_polymorphic_compare := Replace_polymorphic_compare

include Base.Comparable.S with type t := t with type comparator_witness := comparator_witness

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

include Core_kernel.Comparator.S with type t := t with type comparator_witness := comparator_witness

val comparator : (t, comparator_witness) Core_kernel.Comparator.comparator

module Map : Core_kernel.Map.S_binable with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Core_kernel.Set.S_binable with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Core_kernel.Hashable.S_binable with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val hashable : t Core_kernel.Hashtbl.Hashable.t

module Table : Core_kernel.Hashtbl.S_binable with type key = t

module Hash_set : Core_kernel.Hash_set.S_binable with type elt = t

module Hash_queue : Core_kernel.Hash_queue.S with type key = t

val get_sexp_zone : unit -> Zone.t

val set_sexp_zone : Zone.t -> unit

val (>=.) : t -> t -> bool

val (<=.) : t -> t -> bool

val (=.) : t -> t -> bool

val (>.) : t -> t -> bool

val (<.) : t -> t -> bool

val (<>.) : t -> t -> bool

val robustly_compare : t -> t -> int

val of_tm : Unix.tm -> zone:Zone.t -> t

of_tm converts a Unix.tm (mirroring a struct tm from the C stdlib) into a Time.t. Note that the tm_wday, tm_yday, and tm_isdst fields are ignored.

val of_date_ofday_zoned : Core_kernel.Date.t -> Ofday.Zoned.t -> t

Conversion functions that involved Ofday.Zoned.t, exactly analogous to the conversion functions that involve Ofday.t

val to_date_ofday_zoned : t -> zone:Core_kernel__Zone.t -> Core_kernel.Date.t * Ofday.Zoned.t

val to_ofday_zoned : t -> zone:Core_kernel__Zone.t -> Ofday.Zoned.t

val to_string_fix_proto : [ `Utc | `Local ] -> t -> string

val of_string_fix_proto : [ `Utc | `Local ] -> string -> t

val of_string_abs : string -> t

This is like of_string except that if the string doesn't specify the zone then it raises rather than assume the local timezone.

val of_string_gen : if_no_timezone:[ `Fail | `Local | `Use_this_one of Zone.t ] -> string -> t

of_string_gen ~if_no_timezone s attempts to parse s to a t. If s doesn't supply a time zone if_no_timezone is consulted.

val t_of_sexp_abs : Core_kernel.Sexp.t -> t

t_of_sexp_abs sexp as t_of_sexp, but demands that sexp indicate the timezone the time is expressed in.

val sexp_of_t_abs : t -> zone:Zone.t -> Core_kernel.Sexp.t

Miscellaneous

val pause : Span.t -> unit

pause span sleeps for span time.

val interruptible_pause : Span.t -> [ `Ok | `Remaining of Span.t ]

interruptible_pause span sleeps for span time unless interrupted (e.g. by delivery of a signal), in which case the remaining unslept portion of time is returned.

val pause_forever : unit -> Core_kernel.Nothing.t

pause_forever sleeps indefinitely.

val format : t -> string -> zone:Zone.t -> string

format t fmt formats the given time according to fmt, which follows the formatting rules given in 'man strftime'. The time is output in the given timezone. Here are some commonly used control codes:

      %Y - year (4 digits)
      %y - year (2 digits)
      %m - month
      %d - day
      %H - hour
      %M - minute
      %S - second

a common choice would be: %Y-%m-%d %H:%M:%S

Although %Z and %z are interpreted as format strings, neither are correct in the current implementation. %Z always refers to the local machine timezone, and does not correctly detect whether DST is active. The effective local timezone can be controlled by setting the "TZ" environment variable before calling format. %z behaves unreliably and should be avoided.

Not all strftime control codes are standard; the supported subset will depend on the C libraries linked into a given executable.

val parse : string -> fmt:string -> zone:Zone.t -> t

parse string ~fmt ~zone parses string, according to fmt, which follows the formatting rules given in 'man strptime'. The time is assumed to be in the given timezone.

      %Y - year (4 digits)
      %y - year (2 digits)
      %m - month
      %d - day
      %H - hour
      %M - minute
      %S - second

module Exposed_for_tests : sig ... end

module Stable : sig ... end