Module Stable.With_utc_sexp

Provides a sexp representation that is independent of the time zone of the machine writing it. V1.t_of_sexp will be generous in what sexps it accepts, while V2.t_of_sexp should not be expected to parse sexps not generated by V2.sexp_of_t

module V1 : Core_kernel.Stable_comparable.V1 with type t = t with type comparator_witness = comparator_witness

module V2 : sig ... end