Module type Time_common.S

type t

include Core_kernel.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Core_kernel.Comparable.S with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_ubound : max:t Base.Maybe_bound.t -> t Base.Validate.check

val validate_bound : min:t Base.Maybe_bound.t -> max:t Base.Maybe_bound.t -> t Base.Validate.check

module Replace_polymorphic_compare : sig ... end

module Map : Core_kernel.Map.S with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Core_kernel.Set.S with type Elt.t = t with type Elt.comparator_witness = comparator_witness

include Core_kernel.Hashable.S with type t := t

include Core_kernel.Hashable.Common with type t := t

val compare : t -> t -> Base.Int.t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val hashable : t Core_kernel.Hashtbl.Hashable.t

module Table : Core_kernel.Hashtbl.S with type key = t

module Hash_set : Core_kernel.Hash_set.S with type elt = t

module Hash_queue : Core_kernel.Hash_queue.S with type key = t

include Core_kernel.Sexpable.S with type t := t

val t_of_sexp : Base.Sexp.t -> t

val sexp_of_t : t -> Base.Sexp.t

include Base.Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

module Span : sig ... end

module Ofday : sig ... end

val epoch : t

val now : unit -> t

val add : t -> Span.t -> t

val sub : t -> Span.t -> t

val diff : t -> t -> Span.t

val abs_diff : t -> t -> Span.t

val to_string_fix_proto : [ `Local | `Utc ] -> t -> string

val of_string_fix_proto : [ `Local | `Utc ] -> string -> t

val to_string_abs : t -> zone:Core_kernel__Zone.t -> string

val of_string_abs : string -> t

val next_multiple : ?⁠can_equal_after:bool -> base:t -> after:t -> interval:Span.t -> unit -> t

val of_date_ofday : zone:Core_kernel__Zone.t -> Core_kernel.Date.t -> Ofday.t -> t

val to_ofday : t -> zone:Core_kernel__Zone.t -> Ofday.t

val to_date : t -> zone:Core_kernel__Zone.t -> Core_kernel.Date.t

val occurrence : [ `First_after_or_at | `Last_before_or_at ] -> t -> ofday:Ofday.t -> zone:Core_kernel__Zone.t -> t

val pause : Span.t -> unit

val interruptible_pause : Span.t -> [ `Ok | `Remaining of Span.t ]

val pause_forever : unit -> Core_kernel.Nothing.t