Module Time_ns.Option

Option.t is like t option, except that the value is immediate. This module should mainly be used to avoid allocations.

type value := t

type t = private Span.Option.t

include Ppx_compare_lib.Comparable.S with type t := t

include Bin_prot.Binable.S with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t

include Immediate_option.S_int63 with type value := value with type t := t

include Immediate_option_intf.S_without_immediate with type t := t with type value := value

include Ppx_compare_lib.Comparable.S with type t := t

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

include Typerep_lib.Typerepable.S with type t := t

val typerep_of_t : t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : t Typerep_lib.Typename.t

include Immediate_option_intf.S_without_immediate_plain with type t := t with type value := value

Constructors analogous to None and Some. If not (some_is_representable x) then some x may raise or return none.

val none : t

val some : value -> t

val some_is_representable : value -> Base.Bool.t

For some representations of immediate options, the encodings of none and some overlap. For these representations, some_is_representable value = false if value cannot be represented as an option. For example, Int.Option uses min_value to represent none. For other representations, some_is_representable always returns true.

val is_none : t -> Base.Bool.t

val is_some : t -> Base.Bool.t

val value : t -> default:value -> value

value (some x) ~default = x and value none ~default = default.

val value_exn : t -> value

value_exn (some x) = x. value_exn none raises. Unlike Option.value_exn, there is no ?message argument, so that calls to value_exn that do not raise also do not have to allocate.

val unchecked_value : t -> value

unchecked_value (some x) = x. unchecked_value none returns an unspecified value. unchecked_value t is intended as an optimization of value_exn t when is_some t is known to be true.

val to_option : t -> value Base.Option.t

val of_option : value Base.Option.t -> t

module Optional_syntax : Optional_syntax.S with type t := t with type value := value

include Quickcheck.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

module Stable : sig ... end

val sexp_of_t : [ `Use_Time_ns_unix ]

deprecated [since 2023-09] Use [Time_ns_unix.Option]

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

module Alternate_sexp : sig ... end

Just like Time_ns.Alternate_sexp, this is different from the sexp format exposed in Time_ns_unix. See the comment above.