Module Type_equal.Id

include module type of struct include Base.Type_equal.Id end

type 'a t = 'a Base.Type_equal.Id.t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

module type Arg0 = Base.Type_equal.Id.Arg0

module type Arg1 = Base.Type_equal.Id.Arg1

module type Arg2 = Base.Type_equal.Id.Arg2

module type Arg3 = Base.Type_equal.Id.Arg3

module type S0 = Base.Type_equal.Id.S0

module type S1 = Base.Type_equal.Id.S1

module type S2 = Base.Type_equal.Id.S2

module type S3 = Base.Type_equal.Id.S3

val uid : _ t -> Base.Type_equal.Id.Uid.t

val create : name:string -> ('a -> Base.Sexp.t) -> 'a t

create ~name defines a new type identity. Two calls to create will result in two distinct identifiers, even for the same arguments with the same type. If the type 'a doesn't support sexp conversion, then a good practice is to have the converter be [%sexp_of: _], (or sexp_of_opaque, if not using ppx_sexp_conv).

Accessors

val hash : _ t -> int

val name : _ t -> string

val to_sexp : 'a t -> 'a -> Base.Sexp.t

val hash_fold_t : Base.Hash.state -> _ t -> Base.Hash.state

same_witness t1 t2 and same_witness_exn t1 t2 return a type equality proof iff the two identifiers are the same (i.e., physically equal, resulting from the same call to create). This is a useful way to achieve a sort of dynamic typing. same_witness does not allocate a Some every time it is called.

same t1 t2 = is_some (same_witness t1 t2).

val same : _ t -> _ t -> bool

val same_witness : 'a t -> 'b t -> ('a, 'b) Base.Type_equal.equal option

val same_witness_exn : 'a t -> 'b t -> ('a, 'b) Base.Type_equal.equal

module Create0 = Base.Type_equal.Id.Create0

module Create1 = Base.Type_equal.Id.Create1

module Create2 = Base.Type_equal.Id.Create2

module Create3 = Base.Type_equal.Id.Create3

module Uid : sig ... end