Module Int63.For_testing

val reset_counter : Base.Unit.t -> Base.Unit.t

Resets the counter to its default starting value. The nth call to create after reset_counter has been called has the same ID value as the nth call to create after the start of the program (before reset_counter has been called).

This should only be used in testing to set up a deterministic environment when potentially running multiple tests in a row, because calling this will break the guarantee that create always returns a unique value.