Module Unix.Clock

type t =

| Realtime

| Monotonic

| Process_cpu

| Process_thread

Supported clocks. See clock_gettime(3) man pages for semantics.

val getres : (t -> Core_kernel.Int63.t) Core_kernel.Or_error.t

Return the resolution of the given clock in nanoseconds.

val gettime : (t -> Core_kernel.Int63.t) Core_kernel.Or_error.t

Return the current time of the given clock since 00:00:00 UTC, Jan. 1, 1970, in nanoseconds.