Module Unix.Host

type t = {

name : string;

aliases : string array;

family : Protocol_family.t;

addresses : Inet_addr.t array;

}

Structure of entries in the hosts database.

val sexp_of_t : t -> Ppx_sexp_conv_lib.Sexp.t

val getbyname : string -> t option

Find an entry in hosts with the given name.

NOTE: This function is not thread safe with certain versions of winbind using "wins" name resolution.

val getbyname_exn : string -> t

val getbyaddr : Inet_addr.t -> t option

Find an entry in hosts with the given address.

val getbyaddr_exn : Inet_addr.t -> t

val have_address_in_common : t -> t -> bool