Module Core__Iobuf

A non-moving (in the GC sense) contiguous range of bytes, useful for I/O operations.

An iobuf consists of:

bigstring

limits -- a subrange of the bigstring

window -- a subrange of the limits

All iobuf operations are restricted to operate within the limits. Initially, the window of an iobuf is identical to its limits. A phantom type, the "seek" permission, controls whether or not code is allowed to change the limits and window. With seek permission, the limits can be narrowed, but can never be widened, and the window can be set to an arbitrary subrange of the limits.

A phantom type controls whether code can read and write bytes in the bigstring (within the limits) or can only read them.

To present a restricted view of an iobuf to a client, one can create a sub-iobuf or add a type constraint.

Functions operate on the window unless the documentation or naming indicates otherwise.