Module Syslog.Open_option

type t =

| PID
Include PID with each message

| CONS
Write directly to system console if there is an error while sending to system logger

| ODELAY
Delay opening of the connection until syslog is called

| NDELAY
No delay opening connection to syslog daemon

| NOWAIT
Do not wait for child processes while logging message

| PERROR
Print to stderr as well

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t