Module Core.Core_private

To be used in implementing Core, but not by end users.

module Digit_string_helpers : sig ... end

Utility functions for parsing and outputing strings containing known numbers of digits. Used primarily for building functions for reading in and writing out Time related values.

module Time_zone : sig ... end

module Ofday_helpers : sig ... end

This module is only used in the implementation of ofday.ml, and isn't exposed as a submodule of Core.

module Span_float : sig ... end

module Bigbuffer_internal : sig ... end

module Stable_internal : sig ... end

module Std_internal : sig ... end

module Time_ns_alternate_sexp : sig ... end

A Time_ns that uses its alternate sexp representation. *

module Timezone_js_loader : sig ... end

module Timezone_types : sig ... end