Module Core.Deque

A double-ended queue that can shrink and expand on both ends.

An index is assigned to an element when it enters the queue, and the index of an element is static (i.e., an index refers to a distinct element until that element is removed from the queue, no matter how many intervening push/pop operations occur).

One consequence of this is that the minimum index may be less than zero.

The "front" is the smallest valid index, while the "back" is the largest.

All operations are amortized O(1) with a small constant.

type 'a t

include Bin_prot.Binable.S1 with type 'a t := 'a t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'a t) Bin_prot.Size.sizer1

val bin_write_t : ('a, 'a t) Bin_prot.Write.writer1

val bin_read_t : ('a, 'a t) Bin_prot.Read.reader1

val __bin_read_t__ : ('a, int -> 'a t) Bin_prot.Read.reader1

val bin_writer_t : ('a, 'a t) Bin_prot.Type_class.S1.writer

val bin_reader_t : ('a, 'a t) Bin_prot.Type_class.S1.reader

val bin_t : ('a, 'a t) Bin_prot.Type_class.S1.t

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

val t_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'a t Sexplib0.Sexp_grammar.t

include Binary_searchable.S1 with type 'a t := 'a t

val binary_search : ?pos:int -> ?len:int -> 'a t -> compare:('a -> 'key -> int) -> [ `Last_strictly_less_than | `Last_less_than_or_equal_to | `Last_equal_to | `First_equal_to | `First_greater_than_or_equal_to | `First_strictly_greater_than ] -> 'key -> int option

val binary_search_segmented : ?pos:int -> ?len:int -> 'a t -> segment_of:('a -> [ `Left | `Right ]) -> [ `Last_on_left | `First_on_right ] -> int option

include Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'acc -> f:('acc -> 'a -> 'acc) -> 'acc

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'acc -> f:('acc -> 'a -> ('acc, 'e) Base.Result.t) -> ('acc, 'e) Base.Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'acc -> f:('acc -> 'a -> ('acc, 'final) Base.Container.Continue_or_stop.t) -> finish:('acc -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

  type maybe_negative =
    | Found_negative of int
    | All_nonnegative of { sum : int }

  (** [first_neg_or_sum list] returns the first negative number in [list], if any,
      otherwise returns the sum of the list. *)
  let first_neg_or_sum =
    List.fold_until ~init:0
      ~f:(fun sum x ->
        if x < 0
        then Stop (Found_negative x)
        else Continue (sum + x))
      ~finish:(fun sum -> All_nonnegative { sum })
  ;;

  let x = first_neg_or_sum [1; 2; 3; 4; 5]
  val x : maybe_negative = All_nonnegative {sum = 15}

  let y = first_neg_or_sum [1; 2; -3; 4; 5]
  val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

Returns the sum of f i for all i in the container.

val sum : (module Base.Container.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

val find : 'a t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

Returns a minimum (resp maximum) element from the collection using the provided compare function, or None if the collection is empty. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold.

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

val create : ?initial_length:Base.Int.t -> ?never_shrink:Base.Bool.t -> Base.Unit.t -> _ t

create ?initial_length ?never_shrink () creates a new t. initial_length is the initial length of the dequeue; it will be able to hold initial_length elements without resizing. It must be positive. If never_shrink is true, the physical array will never shrink, only expand. If initial_length is given without never_shrink, then never_shrink is presumed to be true, otherwise never_shrink defaults to false.

parameter initial_length
defaults to 7

val of_array : 'a Base.Array.t -> 'a t

of_array arr creates a dequeue containing the elements of arr. The first element of the array will be at the front of the dequeue.

val front_index : _ t -> Base.Int.t Base.Option.t

front_index t return the index of the front item in t.

val front_index_exn : _ t -> Base.Int.t

front_index_exn t throws an exception if t is empty, otherwise returns the index of the front item in t.

val back_index : _ t -> Base.Int.t Base.Option.t

back_index t return the index of the back item in t.

val back_index_exn : _ t -> Base.Int.t

back_index_exn t throws an exception if t is empty, otherwise returns the index of the back item in t.

val get_opt : 'a t -> Base.Int.t -> 'a Base.Option.t

get_opt t i returns the element at index i. Return None if i is invalid.

val get : 'a t -> Base.Int.t -> 'a

get t i returns the element at index i. Raise an exception if i is invalid.

val peek : 'a t -> [ `back | `front ] -> 'a Base.Option.t

peek t back_or_front returns the value at the back or front of the dequeue without removing it.

val peek_front : 'a t -> 'a Base.Option.t

val peek_front_exn : 'a t -> 'a

val peek_back : 'a t -> 'a Base.Option.t

val peek_back_exn : 'a t -> 'a

val set_exn : 'a t -> Base.Int.t -> 'a -> Base.Unit.t

set_exn t i v mutates the element at i.

val iter' : 'a t -> [ `front_to_back | `back_to_front ] -> f:('a -> Base.Unit.t) -> Base.Unit.t

iter' t ~f iterates over the elements of t.

val iteri : 'a t -> f:(Base.Int.t -> 'a -> Base.Unit.t) -> Base.Unit.t

iteri t ~f iterates over the elements of t `front_to_back passing in the index.

val iteri' : 'a t -> [ `front_to_back | `back_to_front ] -> f:(Base.Int.t -> 'a -> Base.Unit.t) -> Base.Unit.t

iteri' t ~f is the same as iter', but also passes in the index of the current element.

val fold' : 'a t -> [ `front_to_back | `back_to_front ] -> init:'b -> f:('b -> 'a -> 'b) -> 'b

fold' t ~init ~f folds over the elements of t.

val foldi : 'a t -> init:'b -> f:(Base.Int.t -> 'b -> 'a -> 'b) -> 'b

foldi t ~init ~f is the same as fold, but also passes in the index of the current element to f.

val foldi' : 'a t -> [ `front_to_back | `back_to_front ] -> init:'b -> f:(Base.Int.t -> 'b -> 'a -> 'b) -> 'b

foldi' t ~init ~f is the same as fold', but also passes in the index of the current element to f.

val enqueue : 'a t -> [ `back | `front ] -> 'a -> Base.Unit.t

enqueue t back_or_front v pushes v onto the back_or_front of t.

val enqueue_front : 'a t -> 'a -> Base.Unit.t

val enqueue_back : 'a t -> 'a -> Base.Unit.t

val clear : _ t -> Base.Unit.t

clear t removes all elements from t.

val drop : ?n:Base.Int.t -> _ t -> [ `back | `front ] -> Base.Unit.t

drop ?n t back_or_front drops n elements (default 1) from the back_or_front of t. If t has fewer than n elements then it is cleared.

val drop_front : ?n:Base.Int.t -> _ t -> Base.Unit.t

val drop_back : ?n:Base.Int.t -> _ t -> Base.Unit.t

val dequeue : 'a t -> [ `back | `front ] -> 'a Base.Option.t

dequeue t back_or_front removes and returns the back_or_front of t.

val dequeue_exn : 'a t -> [ `back | `front ] -> 'a

val dequeue_front : 'a t -> 'a Base.Option.t

val dequeue_front_exn : 'a t -> 'a

val dequeue_back : 'a t -> 'a Base.Option.t

val dequeue_back_exn : 'a t -> 'a