Module Gc.Allocation_policy

The policy used for allocating in the heap.

The Next_fit policy is quite fast but can result in fragmentation.

The First_fit policy can be slower in some cases but can be better for programs with fragmentation problems.

The Best_fit policy is as fast as Next_fit and has less fragmentation than First_fit.

The default is Best_fit.

type t =

| Next_fit

| First_fit

| Best_fit

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t -> t -> int

include Ppx_compare_lib.Equal.S with type t := t

val equal : t -> t -> bool

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : Base.Hash.state -> t -> Base.Hash.state

val hash : t -> Base.Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t