Module Gc.Stat

type t = {

minor_words : Base.Float.t;
(*
Number of words allocated in the minor heap since the program was started. This number is accurate in byte-code programs, but only an approximation in programs compiled to native code.
*)

promoted_words : Base.Float.t;
(*
Number of words allocated in the minor heap that survived a minor collection and were moved to the major heap since the program was started.
*)

major_words : Base.Float.t;
(*
Number of words allocated in the major heap, including the promoted words, since the program was started.
*)

minor_collections : Base.Int.t;
(*
Number of minor collections since the program was started.
*)

major_collections : Base.Int.t;
(*
Number of major collection cycles completed since the program was started.
*)

heap_words : Base.Int.t;
(*
Total size of the major heap, in words.
*)

heap_chunks : Base.Int.t;
(*
Number of contiguous pieces of memory that make up the major heap.
*)

live_words : Base.Int.t;
(*
Number of words of live data in the major heap, including the header words.
*)

live_blocks : Base.Int.t;
(*
Number of live blocks in the major heap.
*)

free_words : Base.Int.t;
(*
Number of words in the free list.
*)

free_blocks : Base.Int.t;
(*
Number of blocks in the free list.
*)

largest_free : Base.Int.t;
(*
Size (in words) of the largest block in the free list.
*)

fragments : Base.Int.t;
(*
Number of wasted words due to fragmentation. These are 1-words free blocks placed between two live blocks. They are not available for allocation.
*)

compactions : Base.Int.t;
(*
Number of heap compactions since the program was started.
*)

top_heap_words : Base.Int.t;
(*
Maximum size reached by the major heap, in words.
*)

stack_size : Base.Int.t;
(*
Current size of the stack, in words.
*)

forced_major_collections : Base.Int.t;
(*
Number of forced full major collection cycles completed since the program was started.
*)

}

val sexp_of_t : t -> Sexplib0.Sexp.t

val forced_major_collections : t -> Base.Int.t

val stack_size : t -> Base.Int.t

val top_heap_words : t -> Base.Int.t

val compactions : t -> Base.Int.t

val fragments : t -> Base.Int.t

val largest_free : t -> Base.Int.t

val free_blocks : t -> Base.Int.t

val free_words : t -> Base.Int.t

val live_blocks : t -> Base.Int.t

val live_words : t -> Base.Int.t

val heap_chunks : t -> Base.Int.t

val heap_words : t -> Base.Int.t

val major_collections : t -> Base.Int.t

val minor_collections : t -> Base.Int.t

val major_words : t -> Base.Float.t

val promoted_words : t -> Base.Float.t

val minor_words : t -> Base.Float.t

module Fields : sig ... end

include Comparable.S_plain with type t := t

include Base.Comparable.S with type t := t

include Base.Comparisons.S with type t := t

include Base.Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Base.Or_error.t

include Base.Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Base.Comparator.comparator

val validate_lbound : min:t Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Maybe_bound.t -> t Validate.check

val validate_bound : min:t Maybe_bound.t -> max:t Maybe_bound.t -> t Validate.check

module Replace_polymorphic_compare : Base.Comparable.Comparisons with type t := t

module Map : Map.S_plain with type Key.t = t with type Key.comparator_witness = comparator_witness

module Set : Set.S_plain with type Elt.t = t with type Elt.comparator_witness = comparator_witness

val add : t -> t -> t

add first second computes first+second pointwise across each field; this helps in aggregating statistics across processes.

val diff : t -> t -> t

diff after before computes after-before pointwise across each field; this helps show the effect (on all stats) of some code block.