Module Core.Int64

This module extends Base.Int64.

Interface from Base

val globalize : int64 -> int64

include Sexplib0.Sexpable.S with type t := int64

val t_sexp_grammar : int64 Sexplib0.Sexp_grammar.t

include Base.Floatable.S with type t := int64

val of_float : float -> int64

val to_float : int64 -> float

include Base.Intable.S with type t := int64

val of_int_exn : int -> int64

val to_int_exn : int64 -> int

include Base.Identifiable.S with type t := int64

include Sexplib0.Sexpable.S with type t := int64

include Base.Stringable.S with type t := int64

include Base.Comparable.S with type t := int64

include Base.Comparisons.S with type t := int64

include Base.Comparisons.Infix with type t := int64

include Base.Comparator.S with type t := int64

type comparator_witness = Base.Int64.comparator_witness

include Base.Pretty_printer.S with type t := int64

include Base.Comparable.With_zero with type t := int64

val is_positive : int64 -> bool

val is_non_negative : int64 -> bool

val is_negative : int64 -> bool

val is_non_positive : int64 -> bool

val sign : int64 -> Base.Sign.t

Returns Neg, Zero, or Pos in a way consistent with the above functions.

val compare__local : int64 -> int64 -> int

val equal__local : int64 -> int64 -> bool

include Base.Invariant.S with type t := int64

val invariant : int64 -> unit

val of_string_opt : string -> int64 option

val to_string_hum : ?delimiter:char -> int64 -> string

delimiter is an underscore by default.

Infix operators and constants

val one : int64

val minus_one : int64

Negation

There are two pairs of integer division and remainder functions, /% and %, and / and rem. They both satisfy the same equation relating the quotient and the remainder:

  x = (x /% y) * y + (x % y);
  x = (x /  y) * y + (rem x y);

The functions return the same values if x and y are positive. They all raise if y = 0.

The functions differ if x < 0 or y < 0.

If y < 0, then % and /% raise, whereas / and rem do not.

x % y always returns a value between 0 and y - 1, even when x < 0. On the other hand, rem x y returns a negative value if and only if x < 0; that value satisfies abs (rem x y) <= abs y - 1.

val rem : int64 -> int64 -> int64

Other common functions

round rounds an int to a multiple of a given to_multiple_of argument, according to a direction dir, with default dir being `Nearest. round will raise if to_multiple_of <= 0. If the result overflows (too far positive or too far negative), round returns an incorrect result.

 | `Down    | rounds toward Int.neg_infinity                          |
 | `Up      | rounds toward Int.infinity                              |
 | `Nearest | rounds to the nearest multiple, or `Up in case of a tie |
 | `Zero    | rounds toward zero                                      |

Here are some examples for round ~to_multiple_of:10 for each direction:

 | `Down    | {10 .. 19} --> 10 | { 0 ... 9} --> 0 | {-10 ... -1} --> -10 |
 | `Up      | { 1 .. 10} --> 10 | {-9 ... 0} --> 0 | {-19 .. -10} --> -10 |
 | `Zero    | {10 .. 19} --> 10 | {-9 ... 9} --> 0 | {-19 .. -10} --> -10 |
 | `Nearest | { 5 .. 14} --> 10 | {-5 ... 4} --> 0 | {-15 ... -6} --> -10 |

For convenience and performance, there are variants of round with dir hard-coded. If you are writing performance-critical code you should use these.

val round : ?dir:[ `Zero | `Nearest | `Up | `Down ] -> int64 -> to_multiple_of:int64 -> int64

val round_towards_zero : int64 -> to_multiple_of:int64 -> int64

val round_down : int64 -> to_multiple_of:int64 -> int64

val round_up : int64 -> to_multiple_of:int64 -> int64

val round_nearest : int64 -> to_multiple_of:int64 -> int64

Successor and predecessor functions

val succ : int64 -> int64

val pred : int64 -> int64

Exponentiation

val pow : int64 -> int64 -> int64

pow base exponent returns base raised to the power of exponent. It is OK if base <= 0. pow raises if exponent < 0, or an integer overflow would occur.

Bit-wise logical operations

val bit_and : int64 -> int64 -> int64

These are identical to land, lor, etc. except they're not infix and have different names.

val bit_or : int64 -> int64 -> int64

val bit_xor : int64 -> int64 -> int64

val bit_not : int64 -> int64

val popcount : int64 -> int

Returns the number of 1 bits in the binary representation of the input.

Bit-shifting operations

The results are unspecified for negative shifts and shifts >= num_bits.

val shift_left : int64 -> int -> int64

Shifts left, filling in with zeroes.

val shift_right : int64 -> int -> int64

Shifts right, preserving the sign of the input.

Increment and decrement functions for integer references

val decr : int64 ref -> unit

val incr : int64 ref -> unit

Conversion functions to related integer types

val of_int32_exn : int32 -> int64

val to_int32_exn : int64 -> int32

val of_int64_exn : int64 -> int64

val to_int64 : int64 -> int64

val of_nativeint_exn : nativeint -> int64

val to_nativeint_exn : int64 -> nativeint

val of_float_unchecked : float -> int64

of_float_unchecked truncates the given floating point number to an integer, rounding towards zero. The result is unspecified if the argument is nan or falls outside the range of representable integers.

val num_bits : int

The number of bits available in this integer type. Note that the integer representations are signed.

val max_value : int64

The largest representable integer.

val min_value : int64

The smallest representable integer.

val shift_right_logical : int64 -> int -> int64

Shifts right, filling in with zeroes, which will not preserve the sign of the input.

val ceil_pow2 : int64 -> int64

ceil_pow2 x returns the smallest power of 2 that is greater than or equal to x. The implementation may only be called for x > 0. Example: ceil_pow2 17 = 32

val floor_pow2 : int64 -> int64

floor_pow2 x returns the largest power of 2 that is less than or equal to x. The implementation may only be called for x > 0. Example: floor_pow2 17 = 16

val ceil_log2 : int64 -> int

ceil_log2 x returns the ceiling of log-base-2 of x, and raises if x <= 0.

val floor_log2 : int64 -> int

floor_log2 x returns the floor of log-base-2 of x, and raises if x <= 0.

val is_pow2 : int64 -> bool

is_pow2 x returns true iff x is a power of 2. is_pow2 raises if x <= 0.

val clz : int64 -> int

Returns the number of leading zeros in the binary representation of the input, as an integer between 0 and one less than num_bits.

The results are unspecified for t = 0.

val ctz : int64 -> int

Returns the number of trailing zeros in the binary representation of the input, as an integer between 0 and one less than num_bits.

The results are unspecified for t = 0.

module O = Base.Int64.O

include module type of O

val (+) : Base.Int64.t -> Base.Int64.t -> Base.Int64.t

val (-) : Base.Int64.t -> Base.Int64.t -> Base.Int64.t

val (*) : Base.Int64.t -> Base.Int64.t -> Base.Int64.t

val (/) : Base.Int64.t -> Base.Int64.t -> Base.Int64.t

val (~-) : Base.Int64.t -> Base.Int64.t

val (**) : Base.Int64.t -> Base.Int64.t -> Base.Int64.t

val (land) : Base.Int64.t -> Base.Int64.t -> Base.Int64.t

val (lor) : Base.Int64.t -> Base.Int64.t -> Base.Int64.t

val (lxor) : Base.Int64.t -> Base.Int64.t -> Base.Int64.t

val lnot : Base.Int64.t -> Base.Int64.t

val abs : Base.Int64.t -> Base.Int64.t

val neg : Base.Int64.t -> Base.Int64.t

val zero : Base.Int64.t

val (%) : Base.Int64.t -> Base.Int64.t -> Base.Int64.t

val (/%) : Base.Int64.t -> Base.Int64.t -> Base.Int64.t

val (//) : Base.Int64.t -> Base.Int64.t -> float

val (lsl) : Base.Int64.t -> int -> Base.Int64.t

val (asr) : Base.Int64.t -> int -> Base.Int64.t

val (lsr) : Base.Int64.t -> int -> Base.Int64.t

Conversion functions

val of_int : int -> int64

val of_int32 : int32 -> int64

val of_int64 : int64 -> int64

val to_int : int64 -> int option

val to_int32 : int64 -> int32 option

val of_nativeint : nativeint -> int64

val to_nativeint : int64 -> nativeint option

Truncating conversions

These functions return the least-significant bits of the input. In cases where optional conversions return Some x, truncating conversions return x.

val to_int_trunc : int64 -> int

val to_int32_trunc : int64 -> int32

val to_nativeint_trunc : int64 -> nativeint

Low-level float conversions

val bits_of_float : float -> int64

bits_of_float will always allocate its result on the heap unless the _unboxed C function call is chosen by the compiler.

val float_of_bits : int64 -> float

float_of_bits will always allocate its result on the heap unless the _unboxed C function call is chosen by the compiler.

Byte swap operations

See Int's byte swap section for a description of Base's approach to exposing byte swap primitives.

As of writing, these operations do not sign extend unnecessarily on 64 bit machines, unlike their int32 counterparts, and hence, are more performant. See the Int32 module for more details of the overhead entailed by the int32 byteswap functions.

val bswap16 : int64 -> int64

val bswap32 : int64 -> int64

val bswap48 : int64 -> int64

val bswap64 : int64 -> int64

Extensions

include Bin_prot.Binable.S with type t := int64

include Bin_prot.Binable.S_only_functions with type t := int64

include Typerep_lib.Typerepable.S with type t := int64

val typerep_of_t : int64 Typerep_lib.Std_internal.Typerep.t

val typename_of_t : int64 Typerep_lib.Typename.t

include Int_intf.Binaryable with type t := int64

module Binary : sig ... end

include Base.Int.Binaryable with type t := int64 and module Binary := Binary

include Int_intf.Hexable with type t := int64

module Hex : sig ... end

include Base.Int.Hexable with type t := int64 and module Hex := Hex

include Identifiable.S with type t := int64 with type comparator_witness := comparator_witness

include Bin_prot.Binable.S with type t := int64

include Bin_prot.Binable.S_only_functions with type t := int64

include Ppx_hash_lib.Hashable.S with type t := int64

include Sexplib0.Sexpable.S with type t := int64

val t_of_sexp : Sexplib0.Sexp.t -> int64

include Ppx_compare_lib.Comparable.S with type t := int64

include Ppx_hash_lib.Hashable.S with type t := int64

val sexp_of_t : int64 -> Sexplib0.Sexp.t

include Base.Stringable.S with type t := int64

val of_string : string -> int64

val to_string : int64 -> string

include Base.Pretty_printer.S with type t := int64

val pp : Base.Formatter.t -> int64 -> unit

include Comparable.S_binable with type t := int64 with type comparator_witness := comparator_witness

include Base.Comparable.S with type t := int64 with type comparator_witness := comparator_witness

include Base.Comparisons.S with type t := int64

include Base.Comparisons.Infix with type t := int64

val (>=) : int64 -> int64 -> bool

val (<=) : int64 -> int64 -> bool

val (=) : int64 -> int64 -> bool

val (>) : int64 -> int64 -> bool

val (<) : int64 -> int64 -> bool

val (<>) : int64 -> int64 -> bool

val equal : int64 -> int64 -> bool

val compare : int64 -> int64 -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : int64 -> int64 -> int64

val max : int64 -> int64 -> int64

val ascending : int64 -> int64 -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : int64 -> int64 -> int

val between : int64 -> low:int64 -> high:int64 -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : int64 -> min:int64 -> max:int64 -> int64

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : int64 -> min:int64 -> max:int64 -> int64 Base.Or_error.t

include Base.Comparator.S with type t := int64 with type comparator_witness := comparator_witness

module Replace_polymorphic_compare : Base.Comparable.Comparisons with type t := int64

include Comparator.S with type t := int64 with type comparator_witness := comparator_witness

val comparator : (int64, comparator_witness) Base.Comparator.comparator

module Map : Map.S_binable with type Key.t = int64 with type Key.comparator_witness = comparator_witness

module Set : Set.S_binable with type Elt.t = int64 with type Elt.comparator_witness = comparator_witness

include Hashable.S_binable with type t := int64

include Ppx_hash_lib.Hashable.S with type t := int64

val hash_fold_t : Base.Hash.state -> int64 -> Base.Hash.state

val hash : int64 -> Base.Hash.hash_value

val hashable : int64 Base.Hashable.t

module Table : Hashtbl.S_binable with type key = int64

module Hash_set : Hash_set.S_binable with type elt = int64

module Hash_queue : Hash_queue.S with type key = int64

include Comparable.Validate_with_zero with type t := int64

val validate_lbound : min:int64 Maybe_bound.t -> int64 Validate.check

val validate_ubound : max:int64 Maybe_bound.t -> int64 Validate.check

val validate_bound : min:int64 Maybe_bound.t -> max:int64 Maybe_bound.t -> int64 Validate.check

val validate_positive : int64 Validate.check

val validate_non_negative : int64 Validate.check

val validate_negative : int64 Validate.check

val validate_non_positive : int64 Validate.check

include Quickcheckable.S_int with type t := int64

include Quickcheck_intf.S_range with type t := int64

include Quickcheck_intf.S with type t := int64

val quickcheck_generator : int64 Base_quickcheck.Generator.t

val quickcheck_observer : int64 Base_quickcheck.Observer.t

val quickcheck_shrinker : int64 Base_quickcheck.Shrinker.t

val gen_incl : int64 -> int64 -> int64 Base_quickcheck.Generator.t

gen_incl lower_bound upper_bound produces values between lower_bound and upper_bound, inclusive. It uses an ad hoc distribution that stresses boundary conditions more often than a uniform distribution, while still able to produce any value in the range. Raises if lower_bound > upper_bound.

val gen_uniform_incl : int64 -> int64 -> int64 Base_quickcheck.Generator.t

gen_uniform_incl lower_bound upper_bound produces a generator for values uniformly distributed between lower_bound and upper_bound, inclusive. Raises if lower_bound > upper_bound.

val gen_log_uniform_incl : int64 -> int64 -> int64 Base_quickcheck.Generator.t

gen_log_uniform_incl lower_bound upper_bound produces a generator for values between lower_bound and upper_bound, inclusive, where the number of bits used to represent the value is uniformly distributed. Raises if (lower_bound < 0) || (lower_bound > upper_bound).

val gen_log_incl : int64 -> int64 -> int64 Base_quickcheck.Generator.t

gen_log_incl lower_bound upper_bound is like gen_log_uniform_incl, but weighted slightly more in favor of generating lower_bound and upper_bound specifically.

include sig ... end

type nonrec t = int64

include Bin_prot.Binable.S_local with type t := t

include Bin_prot.Binable.S_local_only_functions with type t := t

include Bin_prot.Binable.S_only_functions with type t := t

val bin_size_t : t Bin_prot.Size.sizer

val bin_write_t : t Bin_prot.Write.writer

val bin_read_t : t Bin_prot.Read.reader

val __bin_read_t__ : (int -> t) Bin_prot.Read.reader

This function only needs implementation if t exposed to be a polymorphic variant. Despite what the type reads, this does *not* produce a function after reading; instead it takes the constructor tag (int) before reading and reads the rest of the variant t afterwards.

val bin_size_t__local : t Bin_prot.Size.sizer_local

val bin_write_t__local : t Bin_prot.Write.writer_local

val bin_shape_t : Bin_prot.Shape.t

val bin_writer_t : t Bin_prot.Type_class.writer

val bin_reader_t : t Bin_prot.Type_class.reader

val bin_t : t Bin_prot.Type_class.t