Module Core.Result

This module extends Base.Result.

type ('a, 'b) t = ('a, 'b) Base.Result.t =

| Ok of 'a

| Error of 'b

include Bin_prot.Binable.S_local2 with type ('a, 'b) t := ('a, 'b) t

include Bin_prot.Binable.S2 with type ('a, 'b) t := ('a, 'b) t

val bin_shape_t : Bin_prot.Shape.t -> Bin_prot.Shape.t -> Bin_prot.Shape.t

val bin_size_t : ('a, 'b, ('a, 'b) t) Bin_prot.Size.sizer2

val bin_write_t : ('a, 'b, ('a, 'b) t) Bin_prot.Write.writer2

val bin_read_t : ('a, 'b, ('a, 'b) t) Bin_prot.Read.reader2

val __bin_read_t__ : ('a, 'b, int -> ('a, 'b) t) Bin_prot.Read.reader2

val bin_writer_t : ('a, 'b, ('a, 'b) t) Bin_prot.Type_class.S2.writer

val bin_reader_t : ('a, 'b, ('a, 'b) t) Bin_prot.Type_class.S2.reader

val bin_t : ('a, 'b, ('a, 'b) t) Bin_prot.Type_class.S2.t

val bin_size_t__local : ('a, 'b, ('a, 'b) t) Bin_prot.Size.sizer_local2

val bin_write_t__local : ('a, 'b, ('a, 'b) t) Bin_prot.Write.writer_local2

include Ppx_compare_lib.Comparable.S2 with type ('a, 'b) t := ('a, 'b) t

include Ppx_compare_lib.Equal.S2 with type ('a, 'b) t := ('a, 'b) t

include Ppx_hash_lib.Hashable.S2 with type ('a, 'b) t := ('a, 'b) t

include Typerep_lib.Typerepable.S2 with type ('a, 'b) t := ('a, 'b) t

val typerep_of_t : 'a Typerep_lib.Std_internal.Typerep.t -> 'b Typerep_lib.Std_internal.Typerep.t -> ('a, 'b) t Typerep_lib.Std_internal.Typerep.t

val typename_of_t : 'a Typerep_lib.Typename.t -> 'b Typerep_lib.Typename.t -> ('a, 'b) t Typerep_lib.Typename.t

include Sexplib0.Sexpable.S2 with type ('a, 'b) t := ('a, 'b) t

module Diff : sig ... end

include Sexplib0.Sexpable.S2 with type ('ok, 'err) t := ('ok, 'err) t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> (Sexplib0.Sexp.t -> 'b) -> Sexplib0.Sexp.t -> ('a, 'b) t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> ('b -> Sexplib0.Sexp.t) -> ('a, 'b) t -> Sexplib0.Sexp.t

val t_sexp_grammar : 'ok Sexplib0.Sexp_grammar.t -> 'err Sexplib0.Sexp_grammar.t -> ('ok, 'err) t Sexplib0.Sexp_grammar.t

val compare : ('a -> 'a -> int) -> ('b -> 'b -> int) -> ('a, 'b) t -> ('a, 'b) t -> int

val compare__local : ('a -> 'a -> int) -> ('b -> 'b -> int) -> ('a, 'b) t -> ('a, 'b) t -> int

val equal : ('a -> 'a -> bool) -> ('b -> 'b -> bool) -> ('a, 'b) t -> ('a, 'b) t -> bool

val equal__local : ('a -> 'a -> bool) -> ('b -> 'b -> bool) -> ('a, 'b) t -> ('a, 'b) t -> bool

val hash_fold_t : (Base.Hash.state -> 'a -> Base.Hash.state) -> (Base.Hash.state -> 'b -> Base.Hash.state) -> Base.Hash.state -> ('a, 'b) t -> Base.Hash.state

val globalize : ('ok -> 'ok) -> ('err -> 'err) -> ('ok, 'err) t -> ('ok, 'err) t

include Base.Monad.S2_local with type ('a, 'err) t := ('a, 'err) t

val (>>=) : ('a, 'e) t -> ('a -> ('b, 'e) t) -> ('b, 'e) t

val (>>|) : ('a, 'e) t -> ('a -> 'b) -> ('b, 'e) t

module Let_syntax = Base.Result.Let_syntax

module Monad_infix = Base.Result.Monad_infix

val bind : ('a, 'e) t -> f:('a -> ('b, 'e) t) -> ('b, 'e) t

val return : 'a -> ('a, _) t

val join : (('a, 'e) t, 'e) t -> ('a, 'e) t

val ignore_m : (_, 'e) t -> (unit, 'e) t

val all : ('a, 'e) t list -> ('a list, 'e) t

val all_unit : (unit, 'e) t list -> (unit, 'e) t

module Error = Base.Result.Error

val invariant : ('a -> unit) -> ('b -> unit) -> ('a, 'b) t -> unit

val fail : 'err -> (_, 'err) t

val failf : ('a, unit, string, (_, string) t) format4 -> 'a

e.g., failf "Couldn't find bloogle %s" (Bloogle.to_string b).

val is_ok : (_, _) t -> bool

val is_error : (_, _) t -> bool

val ok : ('ok, _) t -> 'ok option

val ok_exn : ('ok, exn) t -> 'ok

val ok_or_failwith : ('ok, string) t -> 'ok

val error : (_, 'err) t -> 'err option

val of_option : 'ok option -> error:'err -> ('ok, 'err) t

val iter : ('ok, _) t -> f:('ok -> unit) -> unit

val iter_error : (_, 'err) t -> f:('err -> unit) -> unit

val map : ('ok, 'err) t -> f:('ok -> 'c) -> ('c, 'err) t

val map_error : ('ok, 'err) t -> f:('err -> 'c) -> ('ok, 'c) t

val combine : ('ok1, 'err) t -> ('ok2, 'err) t -> ok:('ok1 -> 'ok2 -> 'ok3) -> err:('err -> 'err -> 'err) -> ('ok3, 'err) t

Returns Ok if both are Ok and Error otherwise.

val combine_errors : ('ok, 'err) t list -> ('ok list, 'err list) t

combine_errors ts returns Ok if every element in ts is Ok, else it returns Error with all the errors in ts.

This is similar to all from Monad.S2, with the difference that all only returns the first error.

val combine_errors_unit : (unit, 'err) t list -> (unit, 'err list) t

combine_errors_unit returns Ok if every element in ts is Ok (), else it returns Error with all the errors in ts, like combine_errors.

val to_either : ('ok, 'err) t -> ('ok, 'err) Base__.Either0.t

to_either is useful with List.partition_map. For example:

  let ints, exns =
    List.partition_map ["1"; "two"; "three"; "4"] ~f:(fun string ->
      Result.to_either (Result.try_with (fun () -> Int.of_string string)))

val of_either : ('ok, 'err) Base__.Either0.t -> ('ok, 'err) t

val ok_if_true : bool -> error:'err -> (unit, 'err) t

ok_if_true returns Ok () if bool is true, and Error error if it is false.

val try_with : (unit -> 'a) -> ('a, exn) t

module Export = Base.Result.Export

module Stable : sig ... end