Module Stable_unit_test.Unordered_container_test

Unordered container tests are for types with serializations that will contain a certain set of elements (each represented by a single sexp or bin-io string) which may appear in any order, such as hash tables and hash sets.

type t = {

sexps : Base.String.t Base.List.t;

bin_io_header : Base.String.t;

bin_io_elements : Base.String.t Base.List.t;

}