Module Stable.Span

module V1 : sig ... end

V1's sexps use single-unit format and support units from d to ms; it does not support us or ns. V1's sexp conversions do not round-trip precisely.

module V2 : sig ... end

V2's sexps use single-unit format and support units from d to ns. V2 can read V1 sexps but not vice versa. V2's sexp conversions do not round-trip precisely.

module V3 : sig ... end

V3 uses mixed-unit format and supports units from d to ns. V3 can read V2 and V1 sexps but not vice versa. V3's sexp conversions round-trip precisely.